feat(SFT-27): improving readability of code
