+++ packages/plugin/src/Resources/js/client/client.js
@@ -1,1 +1,1 @@
-!function(){"use strict";var e,t,n,r={35858:function(e,t,n){n.r(t);var r,a=n(67294);function o(){return o=Object.assign?Object.assign.bind():function(e){for(var t=1;t<arguments.length;t++){var n=arguments[t];for(var r in n)Object.prototype.hasOwnProperty.call(n,r)&&(e[r]=n[r])}return e},o.apply(this,arguments)}t.default=function(e){return a.createElement("svg",o({xmlns:"http://www.w3.org/2000/svg",viewBox:"0 0 512 512"},e),r||(r=a.createElement("path",{d:"M495.9 166.6c3.2 8.7.5 18.4-6.4 24.6l-43.3 39.4c1.1 8.3 1.7 16.8 1.7 25.4s-.6 17.1-1.7 25.4l43.3 39.4c6.9 6.2 9.6 15.9 6.4 24.6-4.4 11.9-9.7 23.3-15.8 34.3l-4.7 8.1c-6.6 11-14 21.4-22.1 31.2-5.9 7.2-15.7 9.6-24.5 6.8l-55.7-17.7c-13.4 10.3-28.2 18.9-44 25.4l-12.5 57.1c-2 9.1-9 16.3-18.2 17.8-13.8 2.3-28 3.5-42.5 3.5s-28.7-1.2-42.5-3.5c-9.2-1.5-16.2-8.7-18.2-17.8l-12.5-57.1c-15.8-6.5-30.6-15.1-44-25.4l-55.6 17.8c-8.8 2.8-18.6.3-24.5-6.8-8.1-9.8-15.5-20.2-22.1-31.2l-4.7-8.1c-6.1-11-11.4-22.4-15.8-34.3-3.2-8.7-.5-18.4 6.4-24.6l43.3-39.4c-1.1-8.4-1.7-16.9-1.7-25.5s.6-17.1 1.7-25.4l-43.3-39.4c-6.9-6.2-9.6-15.9-6.4-24.6 4.4-11.9 9.7-23.3 15.8-34.3l4.7-8.1c6.6-11 14-21.4 22.1-31.2 5.9-7.2 15.7-9.6 24.5-6.8l55.7 17.7c13.4-10.3 28.2-18.9 44-25.4l12.5-57.1c2-9.1 9-16.3 18.2-17.8C227.3 1.2 241.5 0 256 0s28.7 1.2 42.5 3.5c9.2 1.5 16.2 8.7 18.2 17.8l12.5 57.1c15.8 6.5 30.6 15.1 44 25.4l55.7-17.7c8.8-2.8 18.6-.3 24.5 6.8 8.1 9.8 15.5 20.2 22.1 31.2l4.7 8.1c6.1 11 11.4 22.4 15.8 34.3zM256 336c44.2 0 80-35.8 80-80s-35.8-80-80-80-80 35.8-80 80 35.8 80 80 80z"})))}},65721:function(e,t,n){n.r(t);var r,a=n(67294);function o(){return o=Object.assign?Object.assign.bind():function(e){for(var t=1;t<arguments.length;t++){var n=arguments[t];for(var r in n)Object.prototype.hasOwnProperty.call(n,r)&&(e[r]=n[r])}return e},o.apply(this,arguments)}t.default=function(e){return a.createElement("svg",o({xmlns:"http://www.w3.org/2000/svg",viewBox:"0 0 512 512"},e),r||(r=a.createElement("path",{fill:"currentColor",d:"M416 208c0 45.9-14.9 88.3-40 122.7l126.6 126.7c12.5 12.5 12.5 32.8 0 45.3s-32.8 12.5-45.3 0L330.7 376c-34.4 25.2-76.8 40-122.7 40C93.1 416 0 322.9 0 208S93.1 0 208 0s208 93.1 208 208zM208 352c79.5 0 144-64.5 144-144S287.5 64 208 64 64 128.5 64 208s64.5 144 144 144z"})))}},56137:function(e,t,n){n.r(t);var r,a=n(67294);function o(){return o=Object.assign?Object.assign.bind():function(e){for(var t=1;t<arguments.length;t++){var n=arguments[t];for(var r in n)Object.prototype.hasOwnProperty.call(n,r)&&(e[r]=n[r])}return e},o.apply(this,arguments)}t.default=function(e){return a.createElement("svg",o({xmlns:"http://www.w3.org/2000/svg",viewBox:"0 0 512 512"},e),r||(r=a.createElement("path",{fill:"currentColor",d:"M0 416c0-17.7 14.3-32 32-32h54.7c12.3-28.3 40.5-48 73.3-48s61 19.7 73.3 48H480c17.7 0 32 14.3 32 32s-14.3 32-32 32H233.3c-12.3 28.3-40.5 48-73.3 48s-61-19.7-73.3-48H32c-17.7 0-32-14.3-32-32zm192 0c0-17.7-14.3-32-32-32s-32 14.3-32 32 14.3 32 32 32 32-14.3 32-32zm192-160c0-17.7-14.3-32-32-32s-32 14.3-32 32 14.3 32 32 32 32-14.3 32-32zm-32-80c32.8 0 61 19.7 73.3 48H480c17.7 0 32 14.3 32 32s-14.3 32-32 32h-54.7c-12.3 28.3-40.5 48-73.3 48s-61-19.7-73.3-48H32c-17.7 0-32-14.3-32-32s14.3-32 32-32h246.7c12.3-28.3 40.5-48 73.3-48zM192 64c-17.7 0-32 14.3-32 32s14.3 32 32 32 32-14.3 32-32-14.3-32-32-32zm73.3 0H480c17.7 0 32 14.3 32 32s-14.3 32-32 32H265.3c-12.3 28.3-40.5 48-73.3 48s-61-19.7-73.3-48H32c-17.7 0-32-14.3-32-32s14.3-32 32-32h86.7C131 35.7 159.2 16 192 16s61 19.7 73.3 48z"})))}},27796:function(e,t,n){n.r(t);var r,a=n(67294);function o(){return o=Object.assign?Object.assign.bind():function(e){for(var t=1;t<arguments.length;t++){var n=arguments[t];for(var r in n)Object.prototype.hasOwnProperty.call(n,r)&&(e[r]=n[r])}return e},o.apply(this,arguments)}t.default=function(e){return a.createElement("svg",o({xmlns:"http://www.w3.org/2000/svg",viewBox:"0 0 384 512"},e),r||(r=a.createElement("path",{fill:"currentColor",d:"M41.4 233.4c-12.5 12.5-12.5 32.8 0 45.3l192 192c12.5 12.5 32.8 12.5 45.3 0s12.5-32.8 0-45.3L109.3 256 278.6 86.6c12.5-12.5 12.5-32.8 0-45.3s-32.8-12.5-45.3 0l-192 192z"})))}},33742:function(e,t,n){n.r(t);var r,a=n(67294);function o(){return o=Object.assign?Object.assign.bind():function(e){for(var t=1;t<arguments.length;t++){var n=arguments[t];for(var r in n)Object.prototype.hasOwnProperty.call(n,r)&&(e[r]=n[r])}return e},o.apply(this,arguments)}t.default=function(e){return a.createElement("svg",o({xmlns:"http://www.w3.org/2000/svg",viewBox:"0 0 512 512"},e),r||(r=a.createElement("path",{d:"M256 512c141.4 0 256-114.6 256-256S397.4 0 256 0 0 114.6 0 256s114.6 256 256 256zm-81-337c9.4-9.4 24.6-9.4 33.9 0l47 47 47-47c9.4-9.4 24.6-9.4 33.9 0s9.4 24.6 0 33.9l-47 47 47 47c9.4 9.4 9.4 24.6 0 33.9s-24.6 9.4-33.9 0l-47-47-47 47c-9.4 9.4-24.6 9.4-33.9 0s-9.4-24.6 0-33.9l47-47-47-47c-9.4-9.4-9.4-24.6 0-33.9z"})))}},80755:function(e,t,n){var r=this&&this.__importDefault||function(e){return e&&e.__esModule?e:{default:e}};Object.defineProperty(t,"__esModule",{value:!0});var a=n(88454),o=n(48101),i=r(n(9669));i.default.defaults.baseURL=(0,o.generateUrl)("/"),i.default.defaults.headers.get&&(i.default.defaults.headers.get.Accept="application/json"),i.default.defaults.headers.post&&(i.default.defaults.headers.post.Accept="application/json"),i.default.interceptors.request.use((function(e){return["post","put","patch","delete"].includes(e.method)&&e.data&&void 0!==n.g.Craft&&(e.data[Craft.csrfTokenName]=Craft.csrfTokenValue),e})),i.default.interceptors.response.use(null,(function(e){return e.response.data.error&&(e.message=e.response.data.error),e.response.data.errors?Promise.reject(new a.APIError(e.message,e.response.data.errors)):Promise.reject(e)}))},87662:function(e,t,n){Object.defineProperty(t,"__esModule",{value:!0}),n(80755)},63438:function(e,t,n){Object.defineProperty(t,"__esModule",{value:!0}),t.queryClient=void 0;var r=n(88767);t.queryClient=new r.QueryClient({defaultOptions:{queries:{cacheTime:6e5,retry:!1,refetchOnWindowFocus:!1}}})},72094:function(e,t,n){var r=this&&this.__makeTemplateObject||function(e,t){return Object.defineProperty?Object.defineProperty(e,"raw",{value:t}):e.raw=t,e},a=this&&this.__importDefault||function(e){return e&&e.__esModule?e:{default:e}};Object.defineProperty(t,"__esModule",{value:!0}),t.AppWrapper=void 0;var o,i=a(n(12788));t.AppWrapper=i.default.div(o||(o=r([""],[""])))},38957:function(e,t,n){var r=this&&this.__importDefault||function(e){return e&&e.__esModule?e:{default:e}};Object.defineProperty(t,"__esModule",{value:!0});var a=r(n(67294)),o=n(49818),i=n(72094);t.default=function(){return a.default.createElement(i.AppWrapper,{id:"freeform-client-app"},a.default.createElement(o.Outlet,null))}},16583:function(e,t,n){var r=this&&this.__createBinding||(Object.create?function(e,t,n,r){void 0===r&&(r=n);var a=Object.getOwnPropertyDescriptor(t,n);a&&!("get"in a?!t.__esModule:a.writable||a.configurable)||(a={enumerable:!0,get:function(){return t[n]}}),Object.defineProperty(e,r,a)}:function(e,t,n,r){void 0===r&&(r=n),e[r]=t[n]}),a=this&&this.__setModuleDefault||(Object.create?function(e,t){Object.defineProperty(e,"default",{enumerable:!0,value:t})}:function(e,t){e.default=t}),o=this&&this.__importStar||function(e){if(e&&e.__esModule)return e;var t={};if(null!=e)for(var n in e)"default"!==n&&Object.prototype.hasOwnProperty.call(e,n)&&r(t,e,n);return a(t,e),t},i=this&&this.__importDefault||function(e){return e&&e.__esModule?e:{default:e}};Object.defineProperty(t,"__esModule",{value:!0}),t.CpNavigation=void 0;var l=o(n(67294)),u=n(73935),d=n(49818),c=i(n(87351)),s=function(e){return e.isActive?"sel":void 0};t.CpNavigation=function(){return(0,l.useEffect)((function(){var e;null===(e=document.querySelector("#nav-freeform > a.sel"))||void 0===e||e.classList.remove("sel")}),[]),(0,u.createPortal)(l.default.createElement("ul",{className:"subnav react"},l.default.createElement("li",null,l.default.createElement(d.NavLink,{to:"",className:s},(0,c.default)("Dashboard"))),l.default.createElement("li",null,l.default.createElement(d.NavLink,{to:"forms",className:s},(0,c.default)("Forms"))),l.default.createElement("li",null,l.default.createElement(d.NavLink,{to:"settings",className:s},(0,c.default)("Settings")))),document.getElementById("nav-freeform"))}},47170:function(e,t,n){var r,a=this&&this.__extends||(r=function(e,t){return r=Object.setPrototypeOf||{__proto__:[]}instanceof Array&&function(e,t){e.__proto__=t}||function(e,t){for(var n in t)Object.prototype.hasOwnProperty.call(t,n)&&(e[n]=t[n])},r(e,t)},function(e,t){if("function"!=typeof t&&null!==t)throw new TypeError("Class extends value "+String(t)+" is not a constructor or null");function n(){this.constructor=e}r(e,t),e.prototype=null===t?Object.create(t):(n.prototype=t.prototype,new n)}),o=this&&this.__importDefault||function(e){return e&&e.__esModule?e:{default:e}};Object.defineProperty(t,"__esModule",{value:!0}),t.ErrorBoundary=void 0;var i=o(n(67294)),l=function(e){function t(t){var n=e.call(this,t)||this;return n.state={hasError:!1},n}return a(t,e),t.getDerivedStateFromError=function(){return{hasError:!0}},t.prototype.componentDidCatch=function(e,t){console.log(e,t)},t.prototype.render=function(){return this.state.hasError?i.default.createElement("div",null,this.props.message):this.props.children},t}(i.default.Component);t.ErrorBoundary=l},92617:function(e,t,n){var r=this&&this.__makeTemplateObject||function(e,t){return Object.defineProperty?Object.defineProperty(e,"raw",{value:t}):e.raw=t,e},a=this&&this.__importDefault||function(e){return e&&e.__esModule?e:{default:e}};Object.defineProperty(t,"__esModule",{value:!0}),t.FieldInstructions=t.FieldHeading=t.Field=void 0;var o,i,l,u=a(n(12788));t.Field=u.default.div(o||(o=r(["\n  margin: 0;\n  padding: 0;\n  width: 100%;\n  display: block;\n"],["\n  margin: 0;\n  padding: 0;\n  width: 100%;\n  display: block;\n"]))),t.FieldHeading=u.default.div(i||(i=r(["\n  margin: 0;\n  font-weight: bold;\n  padding: 0 0 10px 0;\n"],["\n  margin: 0;\n  font-weight: bold;\n  padding: 0 0 10px 0;\n"]))),t.FieldInstructions=u.default.div(l||(l=r(["\n  margin: 0;\n  font-size: 90%;\n  font-style: italic;\n  padding: 0 0 10px 0;\n"],["\n  margin: 0;\n  font-size: 90%;\n  font-style: italic;\n  padding: 0 0 10px 0;\n"])))},71967:function(e,t,n){var r=this&&this.__importDefault||function(e){return e&&e.__esModule?e:{default:e}};Object.defineProperty(t,"__esModule",{value:!0}),t.BaseControl=void 0;var a=r(n(67294)),o=n(92617);t.BaseControl=function(e){var t=e.property,n=e.children,r=t.label,i=t.handle,l=t.instructions;return a.default.createElement(o.Field,{className:"field"},!!r&&a.default.createElement(o.FieldHeading,null,a.default.createElement("label",{htmlFor:i},r)),!!l&&a.default.createElement(o.FieldInstructions,null,l),a.default.createElement("div",{className:"input"},n))}},42922:function(e,t,n){var r=this&&this.__importDefault||function(e){return e&&e.__esModule?e:{default:e}};Object.defineProperty(t,"__esModule",{value:!0});var a=r(n(67294)),o=n(32969),i=n(30440),l=n(71967);t.default=function(e){var t=e.value,n=e.property,r=e.onUpdateValue;return a.default.createElement(l.BaseControl,{property:n},a.default.createElement(i.InputWrapper,null,a.default.createElement(o.LightSwitchInput,{id:n.handle,enabled:t,onClick:function(){return r(!t)}})))}},81218:function(e,t,n){var r=this&&this.__makeTemplateObject||function(e,t){return Object.defineProperty?Object.defineProperty(e,"raw",{value:t}):e.raw=t,e},a=this&&this.__importDefault||function(e){return e&&e.__esModule?e:{default:e}};Object.defineProperty(t,"__esModule",{value:!0}),t.Overlay=t.Popover=t.SelectedColor=t.Swatch=void 0;var o,i,l,u,d=a(n(12788));t.Swatch=d.default.div(o||(o=r(["\n  margin: 0;\n  padding: 5px;\n  cursor: pointer;\n  border-radius: 1px;\n  display: inline-block;\n  background-color: #ffffff;\n  box-shadow: 0 0 0 1px rgba(0, 0, 0, 0.1);,\n"],["\n  margin: 0;\n  padding: 5px;\n  cursor: pointer;\n  border-radius: 1px;\n  display: inline-block;\n  background-color: #ffffff;\n  box-shadow: 0 0 0 1px rgba(0, 0, 0, 0.1);,\n"]))),t.SelectedColor=d.default.div(i||(i=r(["\n  margin: 0;\n  padding: 0;\n  width: 36px;\n  height: 14px;\n  border-radius: 2px;\n"],["\n  margin: 0;\n  padding: 0;\n  width: 36px;\n  height: 14px;\n  border-radius: 2px;\n"]))),t.Popover=d.default.div(l||(l=r(["\n  z-index: 2;\n  position: absolute;\n"],["\n  z-index: 2;\n  position: absolute;\n"]))),t.Overlay=d.default.div(u||(u=r(["\n  top: 0;\n  left: 0;\n  right: 0;\n  bottom: 0;\n  position: fixed;\n"],["\n  top: 0;\n  left: 0;\n  right: 0;\n  bottom: 0;\n  position: fixed;\n"])))},25857:function(e,t,n){var r=this&&this.__createBinding||(Object.create?function(e,t,n,r){void 0===r&&(r=n);var a=Object.getOwnPropertyDescriptor(t,n);a&&!("get"in a?!t.__esModule:a.writable||a.configurable)||(a={enumerable:!0,get:function(){return t[n]}}),Object.defineProperty(e,r,a)}:function(e,t,n,r){void 0===r&&(r=n),e[r]=t[n]}),a=this&&this.__setModuleDefault||(Object.create?function(e,t){Object.defineProperty(e,"default",{enumerable:!0,value:t})}:function(e,t){e.default=t}),o=this&&this.__importStar||function(e){if(e&&e.__esModule)return e;var t={};if(null!=e)for(var n in e)"default"!==n&&Object.prototype.hasOwnProperty.call(e,n)&&r(t,e,n);return a(t,e),t};Object.defineProperty(t,"__esModule",{value:!0});var i=o(n(67294)),l=n(93565),u=n(81218),d=n(71967);t.default=function(e){var t=e.value,n=e.property,r=e.onUpdateValue,a=(0,i.useState)(!1),o=a[0],c=a[1];return i.default.createElement(d.BaseControl,{property:n},i.default.createElement(u.Swatch,{onClick:function(){return c(!o)}},i.default.createElement(u.SelectedColor,{style:{backgroundColor:t}})),o&&i.default.createElement(u.Popover,null,i.default.createElement(u.Overlay,{onClick:function(){return c(!1)}}),i.default.createElement(l.SketchPicker,{color:t,onChangeComplete:function(e){var t=e.hex;return r(t)}})))}},42250:function(e,t,n){var r=this&&this.__importDefault||function(e){return e&&e.__esModule?e:{default:e}};Object.defineProperty(t,"__esModule",{value:!0});var a=r(n(67294)),o=r(n(9198)),i=r(n(89361)),l=r(n(49546)),u=r(n(14372)),d=n(71967);n(59339);var c="MM/dd/yyyy";t.default=function(e){var t=e.value,n=e.property,r=e.onUpdateValue,s=new Date,f=t||(0,l.default)(s,c),p=(0,u.default)(f,c,s);return a.default.createElement(d.BaseControl,{property:n},a.default.createElement(o.default,{id:n.handle,minDate:s,selected:p,dateFormat:c,className:(0,i.default)("text","fullwidth"),onChange:function(e){return r((0,l.default)(e,c))}}))}},458:function(e,t,n){var r=this&&this.__importDefault||function(e){return e&&e.__esModule?e:{default:e}};Object.defineProperty(t,"__esModule",{value:!0}),t.textarea=t.string=t.select=t.int=t.datePicker=t.colorPicker=t.bool=void 0;var a=n(42922);Object.defineProperty(t,"bool",{enumerable:!0,get:function(){return r(a).default}});var o=n(25857);Object.defineProperty(t,"colorPicker",{enumerable:!0,get:function(){return r(o).default}});var i=n(42250);Object.defineProperty(t,"datePicker",{enumerable:!0,get:function(){return r(i).default}});var l=n(37280);Object.defineProperty(t,"int",{enumerable:!0,get:function(){return r(l).default}});var u=n(15509);Object.defineProperty(t,"select",{enumerable:!0,get:function(){return r(u).default}});var d=n(21969);Object.defineProperty(t,"string",{enumerable:!0,get:function(){return r(d).default}});var c=n(62553);Object.defineProperty(t,"textarea",{enumerable:!0,get:function(){return r(c).default}})},37280:function(e,t,n){var r=this&&this.__importDefault||function(e){return e&&e.__esModule?e:{default:e}};Object.defineProperty(t,"__esModule",{value:!0});var a=r(n(67294)),o=r(n(89361)),i=n(71967);t.default=function(e){var t=e.value,n=e.property,r=e.onUpdateValue,l=n.handle,u=n.placeholder;return a.default.createElement(i.BaseControl,{property:n},a.default.createElement("input",{id:l,type:"number",placeholder:u,className:(0,o.default)("text","fullwidth"),value:t.toString(),onChange:function(e){return r(Number(e.target.value))}}))}},15509:function(e,t,n){var r=this&&this.__importDefault||function(e){return e&&e.__esModule?e:{default:e}};Object.defineProperty(t,"__esModule",{value:!0});var a=r(n(67294)),o=n(71967);t.default=function(e){var t=e.value,n=e.property,r=e.onUpdateValue,i=n.options;return a.default.createElement(o.BaseControl,{property:n},a.default.createElement("select",{id:n.handle,value:t||"",className:"text fullwidth",onChange:function(e){return r(e.target.value)}},i.map((function(e,t){var n=e.value,r=e.label;return a.default.createElement("option",{key:t,value:n,label:r})}))))}},21969:function(e,t,n){var r=this&&this.__importDefault||function(e){return e&&e.__esModule?e:{default:e}};Object.defineProperty(t,"__esModule",{value:!0});var a=r(n(67294)),o=r(n(89361)),i=n(71967);t.default=function(e){var t=e.value,n=e.property,r=e.onUpdateValue,l=n.handle,u=n.placeholder;return a.default.createElement(i.BaseControl,{property:n},a.default.createElement("input",{id:l,type:"text",placeholder:u,className:(0,o.default)("text","fullwidth",n.flags.includes("code")&&"code"),value:t||"",onChange:function(e){return r(e.target.value)}}))}},62553:function(e,t,n){var r=this&&this.__importDefault||function(e){return e&&e.__esModule?e:{default:e}};Object.defineProperty(t,"__esModule",{value:!0});var a=r(n(67294)),o=r(n(89361)),i=n(71967);t.default=function(e){var t=e.value,n=e.property,r=e.onUpdateValue,l=n.handle,u=n.placeholder,d=n.rows;return a.default.createElement(i.BaseControl,{property:n},a.default.createElement("textarea",{id:l,rows:d,placeholder:u,className:(0,o.default)("text","fullwidth",n.flags.includes("code")&&"code"),value:t||"",onChange:function(e){return r(e.target.value)}}))}},82384:function(e,t,n){var r=this&&this.__createBinding||(Object.create?function(e,t,n,r){void 0===r&&(r=n);var a=Object.getOwnPropertyDescriptor(t,n);a&&!("get"in a?!t.__esModule:a.writable||a.configurable)||(a={enumerable:!0,get:function(){return t[n]}}),Object.defineProperty(e,r,a)}:function(e,t,n,r){void 0===r&&(r=n),e[r]=t[n]}),a=this&&this.__setModuleDefault||(Object.create?function(e,t){Object.defineProperty(e,"default",{enumerable:!0,value:t})}:function(e,t){e.default=t}),o=this&&this.__importStar||function(e){if(e&&e.__esModule)return e;var t={};if(null!=e)for(var n in e)"default"!==n&&Object.prototype.hasOwnProperty.call(e,n)&&r(t,e,n);return a(t,e),t},i=this&&this.__importDefault||function(e){return e&&e.__esModule?e:{default:e}};Object.defineProperty(t,"__esModule",{value:!0}),t.FormControlGenerator=void 0;var l=o(n(67294)),u=n(47785),d=n(61552),c=i(n(76884)),s=n(47170),f=o(n(458));t.FormControlGenerator=function(e){var t=e.onValueUpdate,n=e.namespace,r=e.property,a=(0,u.useSelector)((0,d.selectFormSetting)(n,r.handle)),o=(0,c.default)(r.type),i=f[o];return void 0===i?l.default.createElement("div",null,"...".concat(r.handle," <").concat(r.type," [").concat(o,"]>")):(i.displayName="Setting <".concat(r.type,">"),l.default.createElement(s.ErrorBoundary,{message:"...".concat(r.handle," <").concat(r.type,">")},l.default.createElement(l.Suspense,null,l.default.createElement(i,{value:a,property:r,onUpdateValue:t}))))}},3399:function(e,t,n){var r=this&&this.__importDefault||function(e){return e&&e.__esModule?e:{default:e}};Object.defineProperty(t,"__esModule",{value:!0}),t.CheckboxInput=void 0;var a=r(n(67294));t.CheckboxInput=function(e){var t=e.id,n=e.label,r=e.onClick;return a.default.createElement("div",{className:"checkbox-wrapper"},a.default.createElement("input",{id:t,type:"checkbox",className:"checkbox",onClick:r}),a.default.createElement("label",{htmlFor:t},n))}},30440:function(e,t,n){var r=this&&this.__makeTemplateObject||function(e,t){return Object.defineProperty?Object.defineProperty(e,"raw",{value:t}):e.raw=t,e},a=this&&this.__createBinding||(Object.create?function(e,t,n,r){void 0===r&&(r=n);var a=Object.getOwnPropertyDescriptor(t,n);a&&!("get"in a?!t.__esModule:a.writable||a.configurable)||(a={enumerable:!0,get:function(){return t[n]}}),Object.defineProperty(e,r,a)}:function(e,t,n,r){void 0===r&&(r=n),e[r]=t[n]}),o=this&&this.__setModuleDefault||(Object.create?function(e,t){Object.defineProperty(e,"default",{enumerable:!0,value:t})}:function(e,t){e.default=t}),i=this&&this.__importStar||function(e){if(e&&e.__esModule)return e;var t={};if(null!=e)for(var n in e)"default"!==n&&Object.prototype.hasOwnProperty.call(e,n)&&a(t,e,n);return o(t,e),t};Object.defineProperty(t,"__esModule",{value:!0}),t.Wrapper=t.Handle=t.InputWrapper=void 0;var l=n(240),u=i(n(12788));t.InputWrapper=u.default.div(d||(d=r(["\n  padding: 6px 0;\n"],["\n  padding: 6px 0;\n"]))),t.Handle=u.default.div(c||(c=r(["\n  position: absolute;\n  top: 1px;\n  left: 1px;\n\n  width: 20px;\n  height: 20px;\n\n  border-radius: 10px;\n  background-color: ",";\n\n  transition: left 0.1s ease-out;\n"],["\n  position: absolute;\n  top: 1px;\n  left: 1px;\n\n  width: 20px;\n  height: 20px;\n\n  border-radius: 10px;\n  background-color: ",";\n\n  transition: left 0.1s ease-out;\n"])),l.colors.white);var d,c,s,f,p=(0,u.css)(s||(s=r(["\n  background-image: linear-gradient(\n    to right,\n    ",",\n    ","\n  );\n\n  > "," {\n    left: calc(100% - 21px);\n  }\n"],["\n  background-image: linear-gradient(\n    to right,\n    ",",\n    ","\n  );\n\n  > "," {\n    left: calc(100% - 21px);\n  }\n"])),l.colors.teal550,l.colors.teal550,t.Handle);t.Wrapper=u.default.button(f||(f=r(["\n  cursor: pointer;\n  position: relative;\n\n  display: block;\n  overflow: hidden;\n\n  width: 34px;\n  height: 22px;\n\n  border: none;\n  border-radius: 11px;\n\n  user-select: none;\n  transition: background-image 0.1s linear;\n  background-image: linear-gradient(\n    to right,\n    ",",\n    ","\n  );\n\n  ","\n"],["\n  cursor: pointer;\n  position: relative;\n\n  display: block;\n  overflow: hidden;\n\n  width: 34px;\n  height: 22px;\n\n  border: none;\n  border-radius: 11px;\n\n  user-select: none;\n  transition: background-image 0.1s linear;\n  background-image: linear-gradient(\n    to right,\n    ",",\n    ","\n  );\n\n  ","\n"])),l.colors.gray400,l.colors.gray400,(function(e){return e.enabled&&p}))},32969:function(e,t,n){var r=this&&this.__importDefault||function(e){return e&&e.__esModule?e:{default:e}};Object.defineProperty(t,"__esModule",{value:!0}),t.LightSwitchInput=void 0;var a=r(n(67294)),o=n(30440);t.LightSwitchInput=function(e){var t=e.enabled,n=e.onClick;return a.default.createElement(o.Wrapper,{enabled:t,onClick:function(){return n&&n(!t)}},a.default.createElement(o.Handle,null))}},1847:function(e,t,n){var r=this&&this.__makeTemplateObject||function(e,t){return Object.defineProperty?Object.defineProperty(e,"raw",{value:t}):e.raw=t,e},a=this&&this.__importDefault||function(e){return e&&e.__esModule?e:{default:e}};Object.defineProperty(t,"__esModule",{value:!0}),t.Space=void 0;var o,i=n(240),l=a(n(12788));t.Space=l.default.div(o||(o=r(["\n  width: 100%;\n  height: 0;\n  margin-bottom: ",";\n"],["\n  width: 100%;\n  height: 0;\n  margin-bottom: ",";\n"])),i.spacings.xl)},4524:function(e,t,n){var r=this&&this.__makeTemplateObject||function(e,t){return Object.defineProperty?Object.defineProperty(e,"raw",{value:t}):e.raw=t,e},a=this&&this.__createBinding||(Object.create?function(e,t,n,r){void 0===r&&(r=n);var a=Object.getOwnPropertyDescriptor(t,n);a&&!("get"in a?!t.__esModule:a.writable||a.configurable)||(a={enumerable:!0,get:function(){return t[n]}}),Object.defineProperty(e,r,a)}:function(e,t,n,r){void 0===r&&(r=n),e[r]=t[n]}),o=this&&this.__setModuleDefault||(Object.create?function(e,t){Object.defineProperty(e,"default",{enumerable:!0,value:t})}:function(e,t){e.default=t}),i=this&&this.__importStar||function(e){if(e&&e.__esModule)return e;var t={};if(null!=e)for(var n in e)"default"!==n&&Object.prototype.hasOwnProperty.call(e,n)&&a(t,e,n);return o(t,e),t},l=this&&this.__importDefault||function(e){return e&&e.__esModule?e:{default:e}};Object.defineProperty(t,"__esModule",{value:!0}),t.SidebarSlider=void 0;var u,d,c=i(n(67294)),s=n(39920),f=n(240),p=l(n(12788)),v=n(66264),h=p.default.div(u||(u=r(["\n  position: absolute !important;\n  left: 0;\n  right: 0;\n  top: 0;\n  bottom: 0;\n"],["\n  position: absolute !important;\n  left: 0;\n  right: 0;\n  top: 0;\n  bottom: 0;\n"]))),m=(0,p.default)(s.animated.div)(d||(d=r(["\n  position: absolute;\n  top: 0;\n  left: 0;\n  right: 0;\n  bottom: 0;\n\n  box-shadow: ",";\n"],["\n  position: absolute;\n  top: 0;\n  left: 0;\n  right: 0;\n  bottom: 0;\n\n  box-shadow: ",";\n"])),f.shadows.left);t.SidebarSlider=function(e){var t=e.swiped,n=e.children,r=(0,c.useRef)(null),a=(0,s.useSprings)(n.length,n.map((function(e,n){return{x:n?t?-300:0:t?0:300,config:{mass:1,tension:open?200:350,friction:26}}})));return c.default.createElement(v.Sidebar,{style:{overflow:"hidden"},ref:r},c.default.createElement(h,null,n.map((function(e,t){return c.default.createElement(m,{style:a[t],key:t},e)}))))}},66264:function(e,t,n){var r=this&&this.__makeTemplateObject||function(e,t){return Object.defineProperty?Object.defineProperty(e,"raw",{value:t}):e.raw=t,e},a=this&&this.__importDefault||function(e){return e&&e.__esModule?e:{default:e}};Object.defineProperty(t,"__esModule",{value:!0}),t.Sidebar=void 0;var o,i=n(240),l=a(n(12788));t.Sidebar=l.default.div(o||(o=r(["\n  position: relative;\n\n  flex-basis: 300px;\n  width: 300px;\n  padding: ",";\n  box-sizing: border-box;\n\n  border-bottom-left-radius: ",";\n  box-shadow: inset -1px 0 0 0 rgb(154 165 177 / 25%);\n  background: ",";\n"],["\n  position: relative;\n\n  flex-basis: 300px;\n  width: 300px;\n  padding: ",";\n  box-sizing: border-box;\n\n  border-bottom-left-radius: ",";\n  box-shadow: inset -1px 0 0 0 rgb(154 165 177 / 25%);\n  background: ",";\n"])),(function(e){return e.lean?i.spacings.sm:i.spacings.lg}),i.borderRadius.lg,i.colors.gray050)},59518:function(e,t,n){var r=this&&this.__importDefault||function(e){return e&&e.__esModule?e:{default:e}};Object.defineProperty(t,"__esModule",{value:!0});var a=r(n(76884));t.default=function(e,t){return(void 0===t?[!1]:t)[0]&&(e=(0,a.default)(e)),e.replace(/[^a-zA-Z0-9\-_]/g,"")}},99834:function(e,t,n){var r=this&&this.__importDefault||function(e){return e&&e.__esModule?e:{default:e}};Object.defineProperty(t,"__esModule",{value:!0}),t.handle=void 0;var a=n(59518);Object.defineProperty(t,"handle",{enumerable:!0,get:function(){return r(a).default}})},16430:function(e,t,n){var r=this&&this.__createBinding||(Object.create?function(e,t,n,r){void 0===r&&(r=n);var a=Object.getOwnPropertyDescriptor(t,n);a&&!("get"in a?!t.__esModule:a.writable||a.configurable)||(a={enumerable:!0,get:function(){return t[n]}}),Object.defineProperty(e,r,a)}:function(e,t,n,r){void 0===r&&(r=n),e[r]=t[n]}),a=this&&this.__setModuleDefault||(Object.create?function(e,t){Object.defineProperty(e,"default",{enumerable:!0,value:t})}:function(e,t){e.default=t}),o=this&&this.__importStar||function(e){if(e&&e.__esModule)return e;var t={};if(null!=e)for(var n in e)"default"!==n&&Object.prototype.hasOwnProperty.call(e,n)&&r(t,e,n);return a(t,e),t};Object.defineProperty(t,"__esModule",{value:!0}),t.applyMiddleware=void 0;var i=o(n(99834));t.applyMiddleware=function(e,t,n){var r=e;return t.forEach((function(t){var a=t[0],o=t[1];i[a]&&(r=i[a](e,o,n))})),r}},54981:function(e,t,n){var r=this&&this.__makeTemplateObject||function(e,t){return Object.defineProperty?Object.defineProperty(e,"raw",{value:t}):e.raw=t,e},a=this&&this.__importDefault||function(e){return e&&e.__esModule?e:{default:e}};Object.defineProperty(t,"__esModule",{value:!0}),t.Wrapper=void 0;var o,i=a(n(12788));t.Wrapper=i.default.div(o||(o=r(["\n  color: white;\n  background: red;\n  border: 1px solid darkred;\n"],["\n  color: white;\n  background: red;\n  border: 1px solid darkred;\n"])))},4696:function(e,t,n){var r=this&&this.__importDefault||function(e){return e&&e.__esModule?e:{default:e}};Object.defineProperty(t,"__esModule",{value:!0}),t.ErrorBlock=void 0;var a=r(n(67294)),o=n(54981);t.ErrorBlock=function(e){var t=e.children;return a.default.createElement(o.Wrapper,null,t)}},80306:function(e,t,n){var r=this&&this.__importDefault||function(e){return e&&e.__esModule?e:{default:e}};Object.defineProperty(t,"__esModule",{value:!0}),t.ThemedSkeleton=void 0;var a=r(n(67294)),o=n(84849);t.ThemedSkeleton=function(e){var t=e.children;return a.default.createElement(o.SkeletonTheme,{baseColor:"#e6eaee",highlightColor:"#ced1d4"},t)}},33658:function(e,t,n){var r=this&&this.__importDefault||function(e){return e&&e.__esModule?e:{default:e}};Object.defineProperty(t,"__esModule",{value:!0}),t.Dashboard=void 0;var a=r(n(67294));t.Dashboard=function(){return a.default.createElement("div",null,"Dashboard")}},73803:function(e,t,n){var r=this&&this.__makeTemplateObject||function(e,t){return Object.defineProperty?Object.defineProperty(e,"raw",{value:t}):e.raw=t,e},a=this&&this.__importDefault||function(e){return e&&e.__esModule?e:{default:e}};Object.defineProperty(t,"__esModule",{value:!0}),t.BuilderContent=t.BuilderWrapper=void 0;var o,i,l=n(240),u=a(n(12788));t.BuilderWrapper=u.default.div(o||(o=r(["\n  display: flex;\n  flex-direction: column;\n  height: 100%;\n\n  box-shadow: 0 0 0 1px ",", 0 2px 12px rgb(205 216 228 / 50%);\n  border-radius: ",";\n"],["\n  display: flex;\n  flex-direction: column;\n  height: 100%;\n\n  box-shadow: 0 0 0 1px ",", 0 2px 12px rgb(205 216 228 / 50%);\n  border-radius: ",";\n"])),l.colors.gray200,l.borderRadius.lg),t.BuilderContent=u.default.div(i||(i=r(["\n  flex-grow: 1;\n"],["\n  flex-grow: 1;\n"])))},93787:function(e,t,n){var r=this&&this.__importDefault||function(e){return e&&e.__esModule?e:{default:e}};Object.defineProperty(t,"__esModule",{value:!0}),t.Builder=void 0;var a=r(n(67294)),o=n(84520),i=n(27318),l=n(49818),u=n(39130),d=n(678),c=n(26678),s=n(70793),f=n(86419),p=n(73803),v=n(19179);t.Builder=function(){return a.default.createElement(o.DndProvider,{backend:i.HTML5Backend},a.default.createElement(p.BuilderWrapper,null,a.default.createElement(v.Tabs,null),a.default.createElement(p.BuilderContent,null,a.default.createElement(l.Routes,null,a.default.createElement(l.Route,{index:!0,element:a.default.createElement(f.LayoutEditor,null)}),a.default.createElement(l.Route,{path:"notifications",element:a.default.createElement("div",null,"Notifications")}),a.default.createElement(l.Route,{path:"integrations",element:a.default.createElement(s.Integrations,null)},a.default.createElement(l.Route,{index:!0,element:a.default.createElement(c.EmptyEditor,null)}),a.default.createElement(l.Route,{path:":id/:handle",element:a.default.createElement(d.Editor,null)})),a.default.createElement(l.Route,{path:"rules",element:a.default.createElement("div",null,"Rules")}),a.default.createElement(l.Route,{path:":namespace",element:a.default.createElement(u.FormSettings,null)})))))}},98104:function(e,t,n){var r=this&&this.__makeTemplateObject||function(e,t){return Object.defineProperty?Object.defineProperty(e,"raw",{value:t}):e.raw=t,e},a=this&&this.__importDefault||function(e){return e&&e.__esModule?e:{default:e}};Object.defineProperty(t,"__esModule",{value:!0}),t.GroupsCollection=t.FormSettingsContainer=t.FormSettingsWrapper=void 0;var o,i,l,u=n(240),d=a(n(12788));t.FormSettingsWrapper=d.default.div(o||(o=r(["\n  display: flex;\n  height: 100%;\n  background: ",";\n"],["\n  display: flex;\n  height: 100%;\n  background: ",";\n"])),u.colors.white),t.FormSettingsContainer=d.default.div(i||(i=r(["\n  width: calc(100% - 300px);\n\n  padding: ",";\n"],["\n  width: calc(100% - 300px);\n\n  padding: ",";\n"])),u.spacings.lg),t.GroupsCollection=d.default.div(l||(l=r(["\n  display: flex;\n  flex-direction: column;\n  gap: ",";\n"],["\n  display: flex;\n  flex-direction: column;\n  gap: ",";\n"])),u.spacings.xl)},39130:function(e,t,n){var r=this&&this.__createBinding||(Object.create?function(e,t,n,r){void 0===r&&(r=n);var a=Object.getOwnPropertyDescriptor(t,n);a&&!("get"in a?!t.__esModule:a.writable||a.configurable)||(a={enumerable:!0,get:function(){return t[n]}}),Object.defineProperty(e,r,a)}:function(e,t,n,r){void 0===r&&(r=n),e[r]=t[n]}),a=this&&this.__setModuleDefault||(Object.create?function(e,t){Object.defineProperty(e,"default",{enumerable:!0,value:t})}:function(e,t){e.default=t}),o=this&&this.__importStar||function(e){if(e&&e.__esModule)return e;var t={};if(null!=e)for(var n in e)"default"!==n&&Object.prototype.hasOwnProperty.call(e,n)&&r(t,e,n);return a(t,e),t},i=this&&this.__spreadArray||function(e,t,n){if(n||2===arguments.length)for(var r,a=0,o=t.length;a<o;a++)!r&&a in t||(r||(r=Array.prototype.slice.call(t,0,a)),r[a]=t[a]);return e.concat(r||Array.prototype.slice.call(t))};Object.defineProperty(t,"__esModule",{value:!0}),t.FormSettings=void 0;var l=o(n(67294)),u=n(49818),d=n(82384),c=n(66264),s=n(16430),f=n(85767),p=n(61552),v=n(23493),h=n(62762),m=n(98104);t.FormSettings=function(){var e=(0,f.useAppDispatch)(),t=(0,u.useParams)().namespace,n=(0,v.useQueryFormSettings)(),r=n.data,a=n.isFetching,o=(0,l.useCallback)((function(n){return function(r){e((function(e,a){e((0,p.modifySettings)({namespace:t,key:n.handle,value:(0,s.applyMiddleware)(r,n.middleware,a)}))}))}}),[t,e]);if(!r&&a)return l.default.createElement("div",null,"Loading...");var g=r.find((function(e){return e.handle===t}));if(!g)return null;var y=g.groups,_=g.properties;return l.default.createElement(m.FormSettingsWrapper,null,l.default.createElement(c.Sidebar,null),l.default.createElement(m.FormSettingsContainer,null,l.default.createElement(m.GroupsCollection,null,i(i([],y,!0),[null],!1).map((function(e,n){var r=_.filter((function(t){return!e&&!t.group||(null==e?void 0:e.handle)===t.group})).sort((function(e,t){return e.order-t.order}));return r.length?l.default.createElement(h.Group,{key:n,group:e},r.map((function(e){return l.default.createElement(d.FormControlGenerator,{key:e.handle,namespace:t,property:e,onValueUpdate:o(e)})}))):null})))))}},28741:function(e,t,n){var r=this&&this.__makeTemplateObject||function(e,t){return Object.defineProperty?Object.defineProperty(e,"raw",{value:t}):e.raw=t,e},a=this&&this.__importDefault||function(e){return e&&e.__esModule?e:{default:e}};Object.defineProperty(t,"__esModule",{value:!0}),t.GroupHeader=t.GroupContainer=t.GroupWrapper=void 0;var o,i,l,u=n(240),d=a(n(12788));t.GroupWrapper=d.default.div(o||(o=r([""],[""]))),t.GroupContainer=d.default.div(i||(i=r(["\n  display: grid;\n  gap: ",";\n  grid-template-columns: repeat(2, 1fr);\n"],["\n  display: grid;\n  gap: ",";\n  grid-template-columns: repeat(2, 1fr);\n"])),u.spacings.xl),t.GroupHeader=d.default.h2(l||(l=r(["\n  padding: 0 0 ",";\n  margin: 0;\n"],["\n  padding: 0 0 ",";\n  margin: 0;\n"])),u.spacings.lg)},62762:function(e,t,n){var r=this&&this.__importDefault||function(e){return e&&e.__esModule?e:{default:e}};Object.defineProperty(t,"__esModule",{value:!0}),t.Group=void 0;var a=r(n(67294)),o=r(n(87351)),i=n(28741);t.Group=function(e){var t=e.group,n=e.children;return a.default.createElement(i.GroupWrapper,null,!!(null==t?void 0:t.label)&&a.default.createElement(i.GroupHeader,null,(0,o.default)(t.label)),a.default.createElement(i.GroupContainer,null,n))}},80968:function(e,t,n){var r=this&&this.__makeTemplateObject||function(e,t){return Object.defineProperty?Object.defineProperty(e,"raw",{value:t}):e.raw=t,e},a=this&&this.__importDefault||function(e){return e&&e.__esModule?e:{default:e}};Object.defineProperty(t,"__esModule",{value:!0}),t.SaveButton=t.SaveButtonWrapper=t.TabsWrapper=t.FormName=t.Heading=t.TabWrapper=void 0;var o,i,l,u,d,c,s=n(240),f=a(n(12788));t.TabWrapper=f.default.nav(o||(o=r(["\n  position: relative;\n\n  display: grid;\n  grid-template-columns: 300px auto min-content;\n  align-items: center;\n\n  height: 50px;\n  flex: 0 0 50px;\n\n  box-sizing: border-box;\n  overflow-x: hidden;\n\n  background: ",";\n  box-shadow: inset 0 -1px 0 0 rgb(154 165 177 / 25%);\n"],["\n  position: relative;\n\n  display: grid;\n  grid-template-columns: 300px auto min-content;\n  align-items: center;\n\n  height: 50px;\n  flex: 0 0 50px;\n\n  box-sizing: border-box;\n  overflow-x: hidden;\n\n  background: ",";\n  box-shadow: inset 0 -1px 0 0 rgb(154 165 177 / 25%);\n"])),s.colors.gray050),t.Heading=f.default.h1(i||(i=r(["\n  position: relative;\n\n  display: flex;\n  align-items: center;\n  gap: ",";\n\n  margin: 0;\n  padding: 0 ",";\n\n  a {\n    svg {\n      position: relative;\n\n      width: 18px;\n      height: 18px;\n      color: ",";\n\n      transition: all 0.2s ease-out;\n    }\n\n    &:hover svg {\n      color: ",";\n      transform: translateX(-5px);\n    }\n  }\n"],["\n  position: relative;\n\n  display: flex;\n  align-items: center;\n  gap: ",";\n\n  margin: 0;\n  padding: 0 ",";\n\n  a {\n    svg {\n      position: relative;\n\n      width: 18px;\n      height: 18px;\n      color: ",";\n\n      transition: all 0.2s ease-out;\n    }\n\n    &:hover svg {\n      color: ",";\n      transform: translateX(-5px);\n    }\n  }\n"])),s.spacings.sm,s.spacings.lg,s.colors.gray200,s.colors.gray700),t.FormName=f.default.span(l||(l=r(["\n  font-size: 18px;\n  font-weight: 700;\n  line-height: 1.2;\n  color: ",";\n"],["\n  font-size: 18px;\n  font-weight: 700;\n  line-height: 1.2;\n  color: ",";\n"])),s.colors.gray700),t.TabsWrapper=f.default.div(u||(u=r(["\n  display: flex;\n  align-self: flex-end;\n\n  a {\n    display: flex;\n    align-items: center;\n\n    height: 42px;\n    padding: 0 12px;\n\n    white-space: nowrap;\n\n    color: var(--light-text-color);\n    border-radius: "," "," 0 0;\n\n    &:hover {\n      text-decoration: none;\n      background-color: rgba(154, 165, 177, 0.15);\n    }\n\n    &.active {\n      background: ",";\n\n      box-shadow: 0 0 0 1px ",", 0 2px 12px rgb(205 216 228 / 50%) !important;\n\n      color: ",";\n    }\n  }\n"],["\n  display: flex;\n  align-self: flex-end;\n\n  a {\n    display: flex;\n    align-items: center;\n\n    height: 42px;\n    padding: 0 12px;\n\n    white-space: nowrap;\n\n    color: var(--light-text-color);\n    border-radius: "," "," 0 0;\n\n    &:hover {\n      text-decoration: none;\n      background-color: rgba(154, 165, 177, 0.15);\n    }\n\n    &.active {\n      background: ",";\n\n      box-shadow: 0 0 0 1px ",", 0 2px 12px rgb(205 216 228 / 50%) !important;\n\n      color: ",";\n    }\n  }\n"])),s.borderRadius.md,s.borderRadius.md,s.colors.white,s.colors.gray200,s.colors.gray700),t.SaveButtonWrapper=f.default.div(d||(d=r(["\n  padding: 0 ",";\n"],["\n  padding: 0 ",";\n"])),s.spacings.lg),t.SaveButton=f.default.button.attrs({className:"btn submit"})(c||(c=r([""],[""])))},19179:function(e,t,n){var r=this&&this.__importDefault||function(e){return e&&e.__esModule?e:{default:e}};Object.defineProperty(t,"__esModule",{value:!0}),t.Tabs=void 0;var a=r(n(67294)),o=n(49818),i=n(85767),l=n(80926),u=r(n(27796)),d=n(12020),c=n(23493),s=r(n(87351)),f=n(80968);t.Tabs=function(){var e=(0,i.useAppDispatch)(),t=(0,c.useQueryFormSettings)().data,n=function(){e((0,l.save)())};return(0,d.useOnKeypress)({callback:function(e){if("s"===e.key){var t=window.navigator.platform.match(/Mac/);if(t&&!e.metaKey)return;if(!t&&!e.ctrlKey)return;return e.preventDefault(),n(),!1}},type:"keydown"}),a.default.createElement(f.TabWrapper,null,a.default.createElement(f.Heading,null,a.default.createElement(o.Link,{to:"..",title:(0,s.default)("Back to form list")},a.default.createElement(u.default,null)),a.default.createElement(f.FormName,null,"My form name")),a.default.createElement(f.TabsWrapper,null,a.default.createElement(o.NavLink,{to:"",end:!0},(0,s.default)("Layout")),a.default.createElement(o.NavLink,{to:"notifications"},(0,s.default)("Notifications")),a.default.createElement(o.NavLink,{to:"rules"},(0,s.default)("Rules")),a.default.createElement(o.NavLink,{to:"integrations"},(0,s.default)("Integrations")),(t||[]).map((function(e){return a.default.createElement(o.NavLink,{key:e.handle,to:e.handle},(0,s.default)(e.label))}))),a.default.createElement(f.SaveButtonWrapper,null,a.default.createElement(f.SaveButton,{onClick:n},(0,s.default)("Save"))))}},72467:function(e,t,n){var r=this&&this.__makeTemplateObject||function(e,t){return Object.defineProperty?Object.defineProperty(e,"raw",{value:t}):e.raw=t,e},a=this&&this.__importDefault||function(e){return e&&e.__esModule?e:{default:e}};Object.defineProperty(t,"__esModule",{value:!0}),t.SettingsWrapper=t.EditorWrapper=void 0;var o,i,l=n(240),u=a(n(12788));t.EditorWrapper=u.default.div(o||(o=r(["\n  flex: 1;\n\n  background: ",";\n  padding: ",";\n"],["\n  flex: 1;\n\n  background: ",";\n  padding: ",";\n"])),l.colors.white,l.spacings.md),t.SettingsWrapper=u.default.div(i||(i=r(["\n  display: flex;\n  flex-direction: column;\n  gap: ",";\n"],["\n  display: flex;\n  flex-direction: column;\n  gap: ",";\n"])),l.spacings.xl)},678:function(e,t,n){var r=this&&this.__importDefault||function(e){return e&&e.__esModule?e:{default:e}};Object.defineProperty(t,"__esModule",{value:!0}),t.Editor=void 0;var a=r(n(67294)),o=n(47785),i=n(49818),l=n(1847),u=n(68887),d=n(32156),c=n(72467),s=n(26678);t.Editor=function(){var e=(0,i.useParams)().id,t=((0,o.useDispatch)(),(0,o.useSelector)((0,u.selectIntegration)(Number(e))));if(!t)return a.default.createElement(s.EmptyEditor,null);var n=t.id,r=t.handle,f=t.name,p=t.description,v=(t.enabled,t.settings);return a.default.createElement(c.EditorWrapper,null,a.default.createElement("h1",{title:r},f),!!p&&a.default.createElement("p",null,p),a.default.createElement(l.Space,null),a.default.createElement(c.SettingsWrapper,null,v.map((function(e){return a.default.createElement(d.Setting,{key:e.handle,id:n,setting:e})}))))}},26678:function(e,t,n){var r=this&&this.__importDefault||function(e){return e&&e.__esModule?e:{default:e}};Object.defineProperty(t,"__esModule",{value:!0}),t.EmptyEditor=void 0;var a=r(n(67294)),o=r(n(84849)),i=n(80306),l=n(72467);t.EmptyEditor=function(){return a.default.createElement(l.EditorWrapper,null,a.default.createElement(i.ThemedSkeleton,null,a.default.createElement(o.default,{width:120,height:20}),a.default.createElement("br",null),a.default.createElement(o.default,{width:100,height:10}),a.default.createElement(o.default,{width:50,height:20}),a.default.createElement("br",null),a.default.createElement(o.default,{width:200,height:10}),a.default.createElement(o.default,{width:500,height:10}),a.default.createElement(o.default,{height:30}),a.default.createElement("br",null),a.default.createElement(o.default,{width:150,height:10}),a.default.createElement(o.default,{width:300,height:10}),a.default.createElement(o.default,{height:30})))}},32156:function(e,t){Object.defineProperty(t,"__esModule",{value:!0}),t.Setting=void 0;t.Setting=function(e){e.id,e.setting;return null}},36408:function(e,t,n){var r=this&&this.__makeTemplateObject||function(e,t){return Object.defineProperty?Object.defineProperty(e,"raw",{value:t}):e.raw=t,e},a=this&&this.__importDefault||function(e){return e&&e.__esModule?e:{default:e}};Object.defineProperty(t,"__esModule",{value:!0}),t.IntegrationsWrapper=void 0;var o,i=a(n(12788));t.IntegrationsWrapper=i.default.div(o||(o=r(["\n  display: flex;\n  height: 100%;\n"],["\n  display: flex;\n  height: 100%;\n"])))},70793:function(e,t,n){var r=this&&this.__importDefault||function(e){return e&&e.__esModule?e:{default:e}};Object.defineProperty(t,"__esModule",{value:!0}),t.Integrations=void 0;var a=r(n(67294)),o=n(49818),i=n(69412),l=n(36408);t.Integrations=function(){return a.default.createElement(l.IntegrationsWrapper,null,a.default.createElement(i.List,null),a.default.createElement(o.Outlet,null))}},35615:function(e,t,n){var r=this&&this.__importDefault||function(e){return e&&e.__esModule?e:{default:e}};Object.defineProperty(t,"__esModule",{value:!0}),t.CategorySkeleton=void 0;var a=r(n(67294)),o=r(n(84849)),i=n(80306),l=n(65937);t.CategorySkeleton=function(){return a.default.createElement(i.ThemedSkeleton,null,a.default.createElement(l.Wrapper,null,a.default.createElement(l.Label,null,a.default.createElement(o.default,{width:50})),a.default.createElement(l.ChildrenWrapper,{style:{padding:14}},[0,1,2].map((function(e){return a.default.createElement("div",{key:e,style:{display:"flex",gap:10,alignItems:"center"}},a.default.createElement(o.default,{width:20,height:20,circle:!0}),a.default.createElement("div",{style:{flexGrow:2}},a.default.createElement(o.default,{width:100,style:{top:2}})),a.default.createElement(o.default,{width:10,height:10,circle:!0,style:{top:6}}))})))))}},65937:function(e,t,n){var r=this&&this.__makeTemplateObject||function(e,t){return Object.defineProperty?Object.defineProperty(e,"raw",{value:t}):e.raw=t,e},a=this&&this.__importDefault||function(e){return e&&e.__esModule?e:{default:e}};Object.defineProperty(t,"__esModule",{value:!0}),t.ChildrenWrapper=t.Label=t.Wrapper=void 0;var o,i,l,u=n(240),d=a(n(12788));t.Wrapper=d.default.div(o||(o=r([""],[""]))),t.Label=d.default.span(i||(i=r(["\n  padding-left: ",";\n\n  font-weight: 700;\n  font-size: 11px;\n  color: ",";\n\n  text-transform: uppercase;\n"],["\n  padding-left: ",";\n\n  font-weight: 700;\n  font-size: 11px;\n  color: ",";\n\n  text-transform: uppercase;\n"])),u.spacings.md,u.colors.gray550),t.ChildrenWrapper=d.default.div(l||(l=r(["\n  display: flex;\n  flex-direction: column;\n  gap: ",";\n\n  padding: "," 0;\n"],["\n  display: flex;\n  flex-direction: column;\n  gap: ",";\n\n  padding: "," 0;\n"])),u.spacings.xs,u.spacings.xs)},84673:function(e,t,n){var r=this&&this.__assign||function(){return r=Object.assign||function(e){for(var t,n=1,r=arguments.length;n<r;n++)for(var a in t=arguments[n])Object.prototype.hasOwnProperty.call(t,a)&&(e[a]=t[a]);return e},r.apply(this,arguments)},a=this&&this.__importDefault||function(e){return e&&e.__esModule?e:{default:e}};Object.defineProperty(t,"__esModule",{value:!0}),t.Category=void 0;var o=a(n(67294)),i=n(54717),l=n(65937);t.Category=function(e){var t=e.label,n=e.children;return o.default.createElement(l.Wrapper,null,o.default.createElement(l.Label,null,t),o.default.createElement(l.ChildrenWrapper,null,n.map((function(e){return o.default.createElement(i.Integration,r({key:e.id},e))}))))}},94319:function(e,t,n){var r=this&&this.__makeTemplateObject||function(e,t){return Object.defineProperty?Object.defineProperty(e,"raw",{value:t}):e.raw=t,e},a=this&&this.__importDefault||function(e){return e&&e.__esModule?e:{default:e}};Object.defineProperty(t,"__esModule",{value:!0}),t.Status=t.Name=t.Icon=t.Wrapper=void 0;var o,i,l,u,d=n(240),c=a(n(12788));t.Wrapper=c.default.div(o||(o=r(["\n  > a {\n    display: flex;\n    align-items: center;\n    gap: ",";\n\n    padding: "," ",";\n    border-radius: ",";\n\n    color: ",";\n    font-size: 12px;\n    line-height: 12px;\n\n    transition: background-color 0.2s ease-out;\n    text-decoration: none;\n\n    &.active {\n      background-color: ",";\n    }\n\n    &:hover:not(.active) {\n      background-color: ",";\n    }\n  }\n"],["\n  > a {\n    display: flex;\n    align-items: center;\n    gap: ",";\n\n    padding: "," ",";\n    border-radius: ",";\n\n    color: ",";\n    font-size: 12px;\n    line-height: 12px;\n\n    transition: background-color 0.2s ease-out;\n    text-decoration: none;\n\n    &.active {\n      background-color: ",";\n    }\n\n    &:hover:not(.active) {\n      background-color: ",";\n    }\n  }\n"])),d.spacings.sm,d.spacings.sm,d.spacings.md,d.borderRadius.lg,d.colors.gray700,d.colors.gray200,d.colors.gray100),t.Icon=c.default.div(i||(i=r(["\n  display: block;\n  width: 20px;\n  height: 20px;\n"],["\n  display: block;\n  width: 20px;\n  height: 20px;\n"]))),t.Name=c.default.div(l||(l=r(["\n  flex-grow: 1;\n"],["\n  flex-grow: 1;\n"]))),t.Status=c.default.div(u||(u=r(["\n  content: '';\n\n  justify-self: flex-end;\n\n  width: 10px;\n  height: 10px;\n\n  border: 1px solid\n    ",";\n  border-radius: 100%;\n\n  background-color: ",";\n\n  transition: all 0.3s ease-out;\n"],["\n  content: '';\n\n  justify-self: flex-end;\n\n  width: 10px;\n  height: 10px;\n\n  border: 1px solid\n    ",";\n  border-radius: 100%;\n\n  background-color: ",";\n\n  transition: all 0.3s ease-out;\n"])),(function(e){return e.enabled?"transparent":d.colors.gray550}),(function(e){return e.enabled?d.colors.teal550:"transparent"}))},54717:function(e,t,n){var r=this&&this.__importDefault||function(e){return e&&e.__esModule?e:{default:e}};Object.defineProperty(t,"__esModule",{value:!0}),t.Integration=void 0;var a=r(n(67294)),o=n(47785),i=n(49818),l=n(68887),u=r(n(35858)),d=n(94319);t.Integration=function(e){var t=e.id,n=e.name,r=e.handle,c=e.icon,s=(0,o.useSelector)((0,l.selectIntegration)(t));return a.default.createElement(d.Wrapper,null,a.default.createElement(i.NavLink,{to:"".concat(t,"/").concat(r)},a.default.createElement(d.Icon,null,!!c&&a.default.createElement("img",{src:c}),!c&&a.default.createElement(u.default,null)),a.default.createElement(d.Name,null,n),a.default.createElement(d.Status,{enabled:s.enabled})))}},1663:function(e,t,n){var r=this&&this.__makeTemplateObject||function(e,t){return Object.defineProperty?Object.defineProperty(e,"raw",{value:t}):e.raw=t,e},a=this&&this.__importDefault||function(e){return e&&e.__esModule?e:{default:e}};Object.defineProperty(t,"__esModule",{value:!0}),t.Wrapper=void 0;var o,i=n(240),l=a(n(12788));t.Wrapper=l.default.ul(o||(o=r(["\n  display: flex;\n  flex-direction: column;\n  gap: ",";\n\n  list-style: none;\n"],["\n  display: flex;\n  flex-direction: column;\n  gap: ",";\n\n  list-style: none;\n"])),i.spacings.sm)},69412:function(e,t,n){var r=this&&this.__assign||function(){return r=Object.assign||function(e){for(var t,n=1,r=arguments.length;n<r;n++)for(var a in t=arguments[n])Object.prototype.hasOwnProperty.call(t,a)&&(e[a]=t[a]);return e},r.apply(this,arguments)},a=this&&this.__createBinding||(Object.create?function(e,t,n,r){void 0===r&&(r=n);var a=Object.getOwnPropertyDescriptor(t,n);a&&!("get"in a?!t.__esModule:a.writable||a.configurable)||(a={enumerable:!0,get:function(){return t[n]}}),Object.defineProperty(e,r,a)}:function(e,t,n,r){void 0===r&&(r=n),e[r]=t[n]}),o=this&&this.__setModuleDefault||(Object.create?function(e,t){Object.defineProperty(e,"default",{enumerable:!0,value:t})}:function(e,t){e.default=t}),i=this&&this.__importStar||function(e){if(e&&e.__esModule)return e;var t={};if(null!=e)for(var n in e)"default"!==n&&Object.prototype.hasOwnProperty.call(e,n)&&a(t,e,n);return o(t,e),t};Object.defineProperty(t,"__esModule",{value:!0}),t.List=void 0;var l=i(n(67294)),u=n(49818),d=n(66264),c=n(41648),s=n(84673),f=n(35615),p=n(1663),v={crm:"CRM",mailing_list:"Email Marketing",payment:"Payment"};t.List=function(){var e=(0,u.useParams)(),t=e.formId,n=e.id,a=(0,u.useNavigate)(),o=(0,c.useQueryFormIntegrations)(Number(null!=t?t:0)),i=o.data,h=o.isFetching;if((0,l.useEffect)((function(){if(!n&&i){var e=i.find(Boolean);e&&a("".concat(e.id,"/").concat(e.handle))}}),[n,i]),!i&&h)return l.default.createElement(d.Sidebar,null,l.default.createElement(p.Wrapper,null,l.default.createElement(f.CategorySkeleton,null)));var m={};return i.forEach((function(e){var t,n=e.type;m[n]||(m[n]={type:n,label:null!==(t=v[n])&&void 0!==t?t:"Other",children:[]}),m[n].children.push(e)})),l.default.createElement(d.Sidebar,{lean:!0},l.default.createElement(p.Wrapper,null,Object.values(m).map((function(e){return l.default.createElement(s.Category,r({key:e.type},e))}))))}},62277:function(e,t,n){var r=this&&this.__makeTemplateObject||function(e,t){return Object.defineProperty?Object.defineProperty(e,"raw",{value:t}):e.raw=t,e},a=this&&this.__importDefault||function(e){return e&&e.__esModule?e:{default:e}};Object.defineProperty(t,"__esModule",{value:!0}),t.Label=t.CellFieldWrapper=void 0;var o,i,l=a(n(12788));t.CellFieldWrapper=l.default.div(o||(o=r([""],[""]))),t.Label=l.default.label(i||(i=r([""],[""])))},53293:function(e,t,n){var r=this&&this.__importDefault||function(e){return e&&e.__esModule?e:{default:e}};Object.defineProperty(t,"__esModule",{value:!0}),t.CellField=void 0;var a=r(n(67294)),o=n(47785),i=n(85767),l=n(45323),u=n(56859),d=n(42367),c=n(62277);t.CellField=function(e){var t=e.uid,n=(0,o.useSelector)((0,u.selectField)(t)),r=(0,d.useFieldType)(null==n?void 0:n.typeClass),s=(0,i.useAppDispatch)();return void 0===(null==n?void 0:n.properties)?null:a.default.createElement(c.CellFieldWrapper,{onClick:function(){s((0,l.setFocusedItem)({type:l.FocusType.Field,uid:t}))}},a.default.createElement(c.Label,null,n.properties.label||(null==r?void 0:r.name)),a.default.createElement("div",null,a.default.createElement("input",{type:"text",className:"text fullwidth"})))}},27607:function(e,t,n){var r=this&&this.__importDefault||function(e){return e&&e.__esModule?e:{default:e}};Object.defineProperty(t,"__esModule",{value:!0}),t.CellLayout=void 0;var a=r(n(67294)),o=n(47785),i=n(1965),l=n(56206);t.CellLayout=function(e){var t=e.uid,n=(0,o.useSelector)((0,i.selectLayout)(t));return n?a.default.createElement(l.Layout,{layout:n}):null}},35917:function(e,t,n){var r=this&&this.__makeTemplateObject||function(e,t){return Object.defineProperty?Object.defineProperty(e,"raw",{value:t}):e.raw=t,e},a=this&&this.__importDefault||function(e){return e&&e.__esModule?e:{default:e}};Object.defineProperty(t,"__esModule",{value:!0}),t.Wrapper=void 0;var o,i=n(39920),l=a(n(12788));t.Wrapper=(0,l.default)(i.animated.div)(o||(o=r(["\n  //border: 3px dashed green;\n\n  flex: 1 0;\n  overflow: hidden;\n"],["\n  //border: 3px dashed green;\n\n  flex: 1 0;\n  overflow: hidden;\n"])))},26528:function(e,t,n){var r=this&&this.__importDefault||function(e){return e&&e.__esModule?e:{default:e}};Object.defineProperty(t,"__esModule",{value:!0}),t.Cell=void 0;var a=r(n(67294)),o=n(84520),i=n(11037),l=n(85767),u=n(99614),d=n(53293),c=n(27607),s=n(35917);t.Cell=function(e){var t,n=e.cell,r=e.order,f=(0,l.useAppDispatch)(),p=(0,o.useDrag)((function(){return{type:"LayoutField",item:function(){return f((0,u.setCell)(n.uid)),{}},end:function(){f((0,u.setCell)(void 0))},collect:function(e){return{isDragging:e.isDragging()}}}}),[n.uid])[1];switch(n.type){case i.CellType.Field:t=d.CellField;break;case i.CellType.Layout:t=c.CellLayout}return a.default.createElement(s.Wrapper,{ref:p,style:{order:r}},a.default.createElement(t,{uid:n.targetUid}))}},24982:function(e,t,n){var r=this&&this.__makeTemplateObject||function(e,t){return Object.defineProperty?Object.defineProperty(e,"raw",{value:t}):e.raw=t,e},a=this&&this.__importDefault||function(e){return e&&e.__esModule?e:{default:e}};Object.defineProperty(t,"__esModule",{value:!0}),t.FieldLayoutWrapper=void 0;var o,i=a(n(12788));t.FieldLayoutWrapper=i.default.div(o||(o=r(["\n  flex: 1;\n  padding: var(--padding);\n"],["\n  flex: 1;\n  padding: var(--padding);\n"])))},19149:function(e,t,n){var r=this&&this.__importDefault||function(e){return e&&e.__esModule?e:{default:e}};Object.defineProperty(t,"__esModule",{value:!0}),t.FieldLayout=void 0;var a=r(n(67294)),o=n(47785),i=n(45323),l=n(55014),u=n(29178),d=n(24982);t.FieldLayout=function(){var e=(0,o.useSelector)(i.selectCurrentPage);return a.default.createElement(d.FieldLayoutWrapper,null,a.default.createElement(u.PageTabs,null),e&&a.default.createElement(l.Page,{page:e}))}},46804:function(e,t,n){var r=this&&this.__makeTemplateObject||function(e,t){return Object.defineProperty?Object.defineProperty(e,"raw",{value:t}):e.raw=t,e},a=this&&this.__importDefault||function(e){return e&&e.__esModule?e:{default:e}};Object.defineProperty(t,"__esModule",{value:!0}),t.Wrapper=void 0;var o,i=n(240),l=a(n(12788));t.Wrapper=l.default.div(o||(o=r(["\n  //border: 3px dashed blue;\n\n  display: flex;\n  flex-direction: column;\n  gap: ",";\n"],["\n  //border: 3px dashed blue;\n\n  display: flex;\n  flex-direction: column;\n  gap: ",";\n"])),i.spacings.lg)},56206:function(e,t,n){var r=this&&this.__importDefault||function(e){return e&&e.__esModule?e:{default:e}};Object.defineProperty(t,"__esModule",{value:!0}),t.Layout=void 0;var a=r(n(67294)),o=n(47785),i=n(33314),l=n(6294),u=n(46804);t.Layout=function(e){var t=e.layout,n=(0,o.useSelector)((0,i.selectRowsInLayout)(t));return t?a.default.createElement(u.Wrapper,null,n.map((function(e){return a.default.createElement(l.Row,{row:e,key:e.uid})}))):null}},69876:function(e,t,n){var r=this&&this.__makeTemplateObject||function(e,t){return Object.defineProperty?Object.defineProperty(e,"raw",{value:t}):e.raw=t,e},a=this&&this.__importDefault||function(e){return e&&e.__esModule?e:{default:e}};Object.defineProperty(t,"__esModule",{value:!0}),t.PageTabsContainer=t.PageTabsWrapper=void 0;var o,i,l=n(240),u=a(n(12788));t.PageTabsWrapper=u.default.div(o||(o=r(["\n  display: flex;\n  justify-content: space-between;\n\n  border-bottom: 1px solid ",";\n  margin-bottom: ",";\n"],["\n  display: flex;\n  justify-content: space-between;\n\n  border-bottom: 1px solid ",";\n  margin-bottom: ",";\n"])),l.colors.gray200,l.spacings.lg),t.PageTabsContainer=u.default.div(i||(i=r(["\n  display: flex;\n"],["\n  display: flex;\n"])))},29178:function(e,t,n){var r=this&&this.__assign||function(){return r=Object.assign||function(e){for(var t,n=1,r=arguments.length;n<r;n++)for(var a in t=arguments[n])Object.prototype.hasOwnProperty.call(t,a)&&(e[a]=t[a]);return e},r.apply(this,arguments)},a=this&&this.__importDefault||function(e){return e&&e.__esModule?e:{default:e}};Object.defineProperty(t,"__esModule",{value:!0}),t.PageTabs=void 0;var o=a(n(67294)),i=n(47785),l=n(68974),u=n(78650),d=n(60859),c=n(69876);t.PageTabs=function(){var e=(0,i.useSelector)(l.selectPages);return o.default.createElement(c.PageTabsWrapper,null,o.default.createElement(c.PageTabsContainer,null,e.map((function(e){return o.default.createElement(d.Tab,r({key:e.uid},e))}))),o.default.createElement(u.NewTab,null))}},78650:function(e,t,n){var r=this&&this.__importDefault||function(e){return e&&e.__esModule?e:{default:e}};Object.defineProperty(t,"__esModule",{value:!0}),t.NewTab=void 0;var a=r(n(67294)),o=n(85767),i=n(6175),l=n(86799);t.NewTab=function(){var e=(0,o.useAppDispatch)();return a.default.createElement(l.NewTabWrapper,{onClick:function(){e((0,i.addNewPage)())}},"Add New Page")}},86799:function(e,t,n){var r=this&&this.__makeTemplateObject||function(e,t){return Object.defineProperty?Object.defineProperty(e,"raw",{value:t}):e.raw=t,e},a=this&&this.__importDefault||function(e){return e&&e.__esModule?e:{default:e}};Object.defineProperty(t,"__esModule",{value:!0}),t.NewTabWrapper=t.TabWrapper=void 0;var o,i,l=n(240),u=a(n(12788));t.TabWrapper=u.default.div(o||(o=r(["\n  box-shadow: 0 0 0 1px #cdd8e4, 0 2px 12px rgb(205 216 228 / 50%);\n\n  padding: 4px 18px;\n  border-radius: 4px 4px 0 0;\n  background: ",";\n\n  &:hover {\n    cursor: pointer;\n  }\n"],["\n  box-shadow: 0 0 0 1px #cdd8e4, 0 2px 12px rgb(205 216 228 / 50%);\n\n  padding: 4px 18px;\n  border-radius: 4px 4px 0 0;\n  background: ",";\n\n  &:hover {\n    cursor: pointer;\n  }\n"])),(function(e){return e.active?l.colors.gray050:"white"})),t.NewTabWrapper=(0,u.default)(t.TabWrapper)(i||(i=r(["\n  justify-self: flex-end;\n"],["\n  justify-self: flex-end;\n"])))},60859:function(e,t,n){var r=this&&this.__importDefault||function(e){return e&&e.__esModule?e:{default:e}};Object.defineProperty(t,"__esModule",{value:!0}),t.Tab=void 0;var a=r(n(67294)),o=n(47785),i=n(85767),l=n(45323),u=n(86799);t.Tab=function(e){var t=(0,o.useSelector)(l.selectCurrentPage).uid,n=(0,i.useAppDispatch)();return a.default.createElement(u.TabWrapper,{active:t===e.uid,onClick:function(){n((0,l.setPage)(e.uid))}},e.label)}},55014:function(e,t,n){var r=this&&this.__importDefault||function(e){return e&&e.__esModule?e:{default:e}};Object.defineProperty(t,"__esModule",{value:!0}),t.Page=void 0;var a=r(n(67294)),o=n(47785),i=n(1965),l=n(56206),u=n(19611);t.Page=function(e){var t=e.page,n=(0,o.useSelector)((0,i.selectPageLayout)(t));return a.default.createElement(u.PageWrapper,null,n&&a.default.createElement(l.Layout,{layout:n}))}},19611:function(e,t,n){var r=this&&this.__makeTemplateObject||function(e,t){return Object.defineProperty?Object.defineProperty(e,"raw",{value:t}):e.raw=t,e},a=this&&this.__importDefault||function(e){return e&&e.__esModule?e:{default:e}};Object.defineProperty(t,"__esModule",{value:!0}),t.PageWrapper=void 0;var o,i=n(240),l=a(n(12788));t.PageWrapper=l.default.div(o||(o=r(["\n  display: flex;\n  flex-direction: column;\n  gap: ",";\n"],["\n  display: flex;\n  flex-direction: column;\n  gap: ",";\n"])),i.spacings.md)},58974:function(e,t,n){Object.defineProperty(t,"__esModule",{value:!0}),t.useRowDrop=void 0;var r=n(67294),a=n(84520),o=n(47785),i=n(39920),l=n(85767),u=n(99614);t.useRowDrop=function(e,t){var n=(0,l.useAppDispatch)(),d=(0,o.useSelector)(u.selectCurrentPosition),c=(0,r.useRef)(null),s=(0,r.useRef)(d),f=(0,r.useRef)(0),p=(0,r.useState)(0),v=p[0],h=p[1],m=(0,r.useState)(0),g=m[0],y=m[1];(0,r.useEffect)((function(){var e=c.current.getBoundingClientRect();h(e.width),y(e.left)}),[t]);var _=(0,a.useDrop)((function(){return{accept:["BaseField","LayoutField"],canDrop:function(e,t){return t.isOver({shallow:!0})},drop:function(t){console.log("Row:",e.uid),console.log("Cell:",s.current,t.type)},hover:function(e,r){var a=v/(t+1);f.current!==a&&(f.current=a,console.log(a));var o=r.getClientOffset().x-g,i=Math.floor(o/a);s.current!==i&&(s.current=i,n((0,u.setPosition)(i)))},collect:function(e){return{isOver:e.isOver({shallow:!0}),canDrop:e.canDrop()}}}}),[g,v]),b=_[0].isOver;(0,_[1])(c);var x=(0,i.useSpring)({to:{background:"grey",flexGrow:1,transform:"translateX(".concat(f.current*d,"px)")}});return{dropRef:c,placeholderStyle:x,isOver:b}}},16488:function(e,t,n){var r=this&&this.__makeTemplateObject||function(e,t){return Object.defineProperty?Object.defineProperty(e,"raw",{value:t}):e.raw=t,e},a=this&&this.__importDefault||function(e){return e&&e.__esModule?e:{default:e}};Object.defineProperty(t,"__esModule",{value:!0}),t.Placeholder=t.Container=t.Wrapper=void 0;var o,i,l,u=n(39920),d=n(240),c=a(n(12788));t.Wrapper=c.default.div(o||(o=r([""],[""]))),t.Container=c.default.div(i||(i=r(["\n  //border: 3px dashed red;\n\n  position: relative;\n\n  display: flex;\n  justify-content: space-between;\n  align-items: stretch;\n  gap: ",";\n"],["\n  //border: 3px dashed red;\n\n  position: relative;\n\n  display: flex;\n  justify-content: space-between;\n  align-items: stretch;\n  gap: ",";\n"])),d.spacings.lg),t.Placeholder=(0,c.default)(u.animated.div)(l||(l=r(["\n  overflow: hidden;\n  font-size: 0;\n  line-height: 0;\n\n  border: 3px dashed red;\n"],["\n  overflow: hidden;\n  font-size: 0;\n  line-height: 0;\n\n  border: 3px dashed red;\n"])))},6294:function(e,t,n){var r=this&&this.__importDefault||function(e){return e&&e.__esModule?e:{default:e}};Object.defineProperty(t,"__esModule",{value:!0}),t.Row=void 0;var a=r(n(67294)),o=n(47785),i=n(89372),l=n(26528),u=n(58974),d=n(16488);t.Row=function(e){var t=e.row,n=(0,o.useSelector)((0,i.selectCellsInRow)(t)),r=(0,u.useRowDrop)(t,n.length),c=r.dropRef,s=r.isOver;return a.default.createElement(d.Wrapper,{ref:c},a.default.createElement(d.Container,{style:{backgroundColor:s?"pink":"white"}},n.map((function(e,t){return a.default.createElement(l.Cell,{cell:e,key:e.uid,order:10*(t+1)})}))))}},12887:function(e,t,n){var r=this&&this.__makeTemplateObject||function(e,t){return Object.defineProperty?Object.defineProperty(e,"raw",{value:t}):e.raw=t,e},a=this&&this.__importDefault||function(e){return e&&e.__esModule?e:{default:e}};Object.defineProperty(t,"__esModule",{value:!0}),t.List=t.GroupTitle=t.FieldGroupWrapper=void 0;var o,i,l,u=n(240),d=a(n(12788));t.FieldGroupWrapper=d.default.div(o||(o=r([""],[""]))),t.GroupTitle=d.default.h2(i||(i=r([""],[""]))),t.List=d.default.div(l||(l=r(["\n  display: grid;\n  grid-template-columns: 1fr 1fr;\n  gap: ",";\n\n  margin: 0;\n  padding: 0;\n"],["\n  display: grid;\n  grid-template-columns: 1fr 1fr;\n  gap: ",";\n\n  margin: 0;\n  padding: 0;\n"])),u.spacings.sm)},89829:function(e,t,n){var r=this&&this.__importDefault||function(e){return e&&e.__esModule?e:{default:e}};Object.defineProperty(t,"__esModule",{value:!0}),t.FieldGroup=void 0;var a=r(n(67294)),o=r(n(84849)),i=n(4696),l=n(42367),u=n(82344),d=n(60745),c=n(83884),s=n(12887);n(45207);t.FieldGroup=function(e){var t=e.title,n=(0,d.useSelectSearchedFields)(),r=(0,l.useFetchFieldTypes)({select:n}),f=r.data,p=r.isFetching,v=r.isError,h=r.error;return!f&&p?a.default.createElement(s.List,null,(0,u.range)(9).map((function(e){return a.default.createElement(o.default,{key:e,height:32})}))):v?a.default.createElement(i.ErrorBlock,null,h.message):a.default.createElement(s.FieldGroupWrapper,null,a.default.createElement(s.GroupTitle,null,t),a.default.createElement(s.List,null,f.map((function(e){return a.default.createElement(c.Field,{key:e.typeClass,fieldType:e})}))))}},80695:function(e,t,n){var r=this&&this.__makeTemplateObject||function(e,t){return Object.defineProperty?Object.defineProperty(e,"raw",{value:t}):e.raw=t,e},a=this&&this.__importDefault||function(e){return e&&e.__esModule?e:{default:e}};Object.defineProperty(t,"__esModule",{value:!0}),t.Icon=t.Name=t.Wrapper=void 0;var o,i,l,u=n(240),d=a(n(12788));t.Wrapper=d.default.div(o||(o=r(["\n  cursor: pointer;\n\n  display: flex;\n  gap: ",";\n  align-items: center;\n\n  height: 32px;\n\n  padding: ",";\n  overflow: hidden;\n\n  background: ",";\n  border: 1px solid ",";\n\n  font-size: 12px;\n  font-family: 'SFMono-Regular', Consolas, 'Liberation Mono', Menlo, Courier,\n    monospace;\n"],["\n  cursor: pointer;\n\n  display: flex;\n  gap: ",";\n  align-items: center;\n\n  height: 32px;\n\n  padding: ",";\n  overflow: hidden;\n\n  background: ",";\n  border: 1px solid ",";\n\n  font-size: 12px;\n  font-family: 'SFMono-Regular', Consolas, 'Liberation Mono', Menlo, Courier,\n    monospace;\n"])),u.spacings.sm,u.spacings.sm,u.colors.white,u.colors.gray200),t.Name=d.default.span(i||(i=r(["\n  flex: 1;\n  line-height: 12px;\n"],["\n  flex: 1;\n  line-height: 12px;\n"]))),t.Icon=d.default.div(l||(l=r(["\n  display: block;\n  flex-basis: 24px;\n  height: 24px;\n\n  color: ",";\n"],["\n  display: block;\n  flex-basis: 24px;\n  height: 24px;\n\n  color: ",";\n"])),u.colors.gray200)},83884:function(e,t,n){var r=this&&this.__importDefault||function(e){return e&&e.__esModule?e:{default:e}};Object.defineProperty(t,"__esModule",{value:!0}),t.Field=void 0;var a=r(n(67294)),o=n(84520),i=n(85767),l=n(73396),u=n(80695);t.Field=function(e){var t=e.fieldType,n=(0,i.useAppDispatch)(),r=(0,o.useDrag)((function(){return{type:"BaseField",item:t}}))[1];return a.default.createElement(u.Wrapper,{ref:r,onClick:function(){n((0,l.addNewField)(t))}},a.default.createElement(u.Icon,{dangerouslySetInnerHTML:{__html:t.icon}}),a.default.createElement(u.Name,null,t.name))}},84554:function(e,t,n){var r=this&&this.__makeTemplateObject||function(e,t){return Object.defineProperty?Object.defineProperty(e,"raw",{value:t}):e.raw=t,e},a=this&&this.__importDefault||function(e){return e&&e.__esModule?e:{default:e}};Object.defineProperty(t,"__esModule",{value:!0}),t.FieldListWrapper=void 0;var o,i=n(39920),l=n(240),u=a(n(12788));t.FieldListWrapper=(0,u.default)(i.animated.div)(o||(o=r(["\n  position: relative;\n  padding: ","; ;\n"],["\n  position: relative;\n  padding: ","; ;\n"])),l.spacings.lg)},18842:function(e,t,n){var r=this&&this.__importDefault||function(e){return e&&e.__esModule?e:{default:e}};Object.defineProperty(t,"__esModule",{value:!0}),t.FieldList=void 0;var a=r(n(67294)),o=n(42367),i=n(89829),l=n(65234),u=n(84554);t.FieldList=function(){return(0,o.useFetchFieldPropertySections)(),a.default.createElement(u.FieldListWrapper,null,a.default.createElement(l.Search,null),a.default.createElement(i.FieldGroup,{title:"Base Fields"}))}},60745:function(e,t,n){Object.defineProperty(t,"__esModule",{value:!0}),t.useSelectSearchedFields=void 0;var r=n(67294),a=n(47785),o=n(65673);t.useSelectSearchedFields=function(){var e=(0,a.useSelector)((0,o.selectQuery)(o.Search.Fields));return(0,r.useCallback)((function(t){return e?t.filter((function(t){return t.name.toLowerCase().includes(e)})):t}),[e])}},66528:function(e,t,n){var r=this&&this.__makeTemplateObject||function(e,t){return Object.defineProperty?Object.defineProperty(e,"raw",{value:t}):e.raw=t,e},a=this&&this.__importDefault||function(e){return e&&e.__esModule?e:{default:e}};Object.defineProperty(t,"__esModule",{value:!0}),t.ItemCheckbox=t.Item=t.Heading=t.DropDownWrapper=t.FilterIcon=void 0;var o,i,l,u,d,c=n(240),s=a(n(12788)),f=n(60755);t.FilterIcon=s.default.button(o||(o=r(["\n  cursor: pointer;\n  right: 1px;\n\n  ","\n\n  color: ",";\n  transition: all 0.2s ease-out;\n\n  background: white;\n\n  &:before {\n    content: '';\n\n    position: absolute;\n    left: 0;\n    top: 6px;\n\n    display: block;\n\n    width: 1px;\n    height: 22px;\n\n    background: ",";\n    transition: all 0.2s ease-out;\n  }\n\n  &:focus {\n    outline: none;\n  }\n\n  &:hover:before {\n    top: 0;\n    height: 32px;\n  }\n\n  &.active {\n    background: ",";\n    outline: none;\n\n    &:before {\n      top: 0;\n      height: 32px;\n    }\n  }\n\n  > svg {\n    user-select: none;\n    pointer-events: none;\n  }\n"],["\n  cursor: pointer;\n  right: 1px;\n\n  ","\n\n  color: ",";\n  transition: all 0.2s ease-out;\n\n  background: white;\n\n  &:before {\n    content: '';\n\n    position: absolute;\n    left: 0;\n    top: 6px;\n\n    display: block;\n\n    width: 1px;\n    height: 22px;\n\n    background: ",";\n    transition: all 0.2s ease-out;\n  }\n\n  &:focus {\n    outline: none;\n  }\n\n  &:hover:before {\n    top: 0;\n    height: 32px;\n  }\n\n  &.active {\n    background: ",";\n    outline: none;\n\n    &:before {\n      top: 0;\n      height: 32px;\n    }\n  }\n\n  > svg {\n    user-select: none;\n    pointer-events: none;\n  }\n"])),f.IconStyle,c.colors.gray300,c.colors.gray200,c.colors.gray050),t.DropDownWrapper=s.default.div(i||(i=r(["\n  position: absolute;\n  top: 32px;\n  right: -1px;\n\n  //display: none;\n\n  padding: "," ",";\n  white-space: nowrap;\n  text-align: left;\n\n  border: 1px solid ",";\n  background: ",";\n  box-shadow: rgba(255, 255, 255, 0.1) 0px 1px 1px 0px inset,\n    rgba(50, 50, 93, 0.25) 0px 50px 100px -20px,\n    rgba(0, 0, 0, 0.3) 0px 30px 60px -30px;\n\n  opacity: 0;\n  transform-origin: top right;\n  transform: scaleY(0);\n  transition: all 0.2s ease-out;\n\n  ",".active & {\n    display: block;\n    background: ",";\n\n    opacity: 1;\n    transform: scaleY(1);\n  }\n\n  &:before {\n    content: '';\n\n    position: absolute;\n    top: -1px;\n    right: 0px;\n\n    width: 33px;\n    height: 1px;\n\n    background: ",";\n  }\n"],["\n  position: absolute;\n  top: 32px;\n  right: -1px;\n\n  //display: none;\n\n  padding: "," ",";\n  white-space: nowrap;\n  text-align: left;\n\n  border: 1px solid ",";\n  background: ",";\n  box-shadow: rgba(255, 255, 255, 0.1) 0px 1px 1px 0px inset,\n    rgba(50, 50, 93, 0.25) 0px 50px 100px -20px,\n    rgba(0, 0, 0, 0.3) 0px 30px 60px -30px;\n\n  opacity: 0;\n  transform-origin: top right;\n  transform: scaleY(0);\n  transition: all 0.2s ease-out;\n\n  ",".active & {\n    display: block;\n    background: ",";\n\n    opacity: 1;\n    transform: scaleY(1);\n  }\n\n  &:before {\n    content: '';\n\n    position: absolute;\n    top: -1px;\n    right: 0px;\n\n    width: 33px;\n    height: 1px;\n\n    background: ",";\n  }\n"])),c.spacings.sm,c.spacings.md,c.colors.gray200,c.colors.white,t.FilterIcon,c.colors.gray050,c.colors.gray050),t.Heading=s.default.h3(l||(l=r(["\n  margin: 0;\n  padding: 0 0 ",";\n"],["\n  margin: 0;\n  padding: 0 0 ",";\n"])),c.spacings.xs),t.Item=s.default.li(u||(u=r(["\n  display: flex;\n  gap: ",";\n  text-align: left;\n"],["\n  display: flex;\n  gap: ",";\n  text-align: left;\n"])),c.spacings.sm),t.ItemCheckbox=s.default.input.attrs({type:"checkbox"})(d||(d=r([""],[""])))},14428:function(e,t,n){var r=this&&this.__createBinding||(Object.create?function(e,t,n,r){void 0===r&&(r=n);var a=Object.getOwnPropertyDescriptor(t,n);a&&!("get"in a?!t.__esModule:a.writable||a.configurable)||(a={enumerable:!0,get:function(){return t[n]}}),Object.defineProperty(e,r,a)}:function(e,t,n,r){void 0===r&&(r=n),e[r]=t[n]}),a=this&&this.__setModuleDefault||(Object.create?function(e,t){Object.defineProperty(e,"default",{enumerable:!0,value:t})}:function(e,t){e.default=t}),o=this&&this.__importStar||function(e){if(e&&e.__esModule)return e;var t={};if(null!=e)for(var n in e)"default"!==n&&Object.prototype.hasOwnProperty.call(e,n)&&r(t,e,n);return a(t,e),t},i=this&&this.__importDefault||function(e){return e&&e.__esModule?e:{default:e}};Object.defineProperty(t,"__esModule",{value:!0}),t.Filter=void 0;var l=o(n(67294)),u=n(32523),d=i(n(87351)),c=i(n(56137)),s=n(66528),f=["Favorites","Standard Fields","Form 1 fields","Form 2 fields","Form 3 fields"];t.Filter=function(){var e=(0,l.useState)(!1),t=e[0],n=e[1],r=(0,u.useClickOutside)((function(){n(!1)}),t);return l.default.createElement(l.default.Fragment,null,l.default.createElement(s.FilterIcon,{ref:r,className:t&&"active",onClick:function(e){t&&e.target===r.current&&n(!1),t||n(!0)}},l.default.createElement(c.default,null),l.default.createElement(s.DropDownWrapper,null,l.default.createElement(s.Heading,null,(0,d.default)("Search in")),l.default.createElement("ul",null,f.map((function(e){return l.default.createElement(s.Item,{key:e},l.default.createElement(s.ItemCheckbox,null),e)}))))))}},23636:function(e,t,n){Object.defineProperty(t,"__esModule",{value:!0}),t.useFieldSearch=void 0;var r=n(67294),a=n(47785),o=n(65673),i=n(25818);t.useFieldSearch=function(){var e=(0,a.useDispatch)(),t=(0,r.useState)(""),n=t[0],l=t[1],u=(0,i.useDebounce)(n,1e3);return(0,r.useEffect)((function(){e((0,o.updateQuery)({type:o.Search.Fields,query:u}))}),[u]),[n,l]}},60755:function(e,t,n){var r=this&&this.__makeTemplateObject||function(e,t){return Object.defineProperty?Object.defineProperty(e,"raw",{value:t}):e.raw=t,e},a=this&&this.__createBinding||(Object.create?function(e,t,n,r){void 0===r&&(r=n);var a=Object.getOwnPropertyDescriptor(t,n);a&&!("get"in a?!t.__esModule:a.writable||a.configurable)||(a={enumerable:!0,get:function(){return t[n]}}),Object.defineProperty(e,r,a)}:function(e,t,n,r){void 0===r&&(r=n),e[r]=t[n]}),o=this&&this.__setModuleDefault||(Object.create?function(e,t){Object.defineProperty(e,"default",{enumerable:!0,value:t})}:function(e,t){e.default=t}),i=this&&this.__importStar||function(e){if(e&&e.__esModule)return e;var t={};if(null!=e)for(var n in e)"default"!==n&&Object.prototype.hasOwnProperty.call(e,n)&&a(t,e,n);return o(t,e),t};Object.defineProperty(t,"__esModule",{value:!0}),t.SearchIcon=t.IconStyle=t.SearchBar=t.SearchBlock=t.Wrapper=void 0;var l=n(240),u=i(n(12788));t.Wrapper=u.default.div(d||(d=r(["\n  position: relative;\n  z-index: 1;\n\n  margin-bottom: ",";\n"],["\n  position: relative;\n  z-index: 1;\n\n  margin-bottom: ",";\n"])),l.spacings.lg),t.SearchBlock=u.default.div(c||(c=r(["\n  display: flex;\n"],["\n  display: flex;\n"]))),t.SearchBar=u.default.input(s||(s=r(["\n  padding: 6px 38px 6px 30px !important;\n\n  &::placeholder {\n    font-style: italic;\n    color: ",";\n  }\n"],["\n  padding: 6px 38px 6px 30px !important;\n\n  &::placeholder {\n    font-style: italic;\n    color: ",";\n  }\n"])),l.colors.gray200);var d,c,s,f,p,v="18px";t.IconStyle=(0,u.css)(f||(f=r(["\n  position: absolute;\n  top: 1px;\n  bottom: 1px;\n  z-index: 2;\n\n  display: flex;\n  flex-direction: column;\n  justify-content: center;\n\n  padding: 0 8px;\n\n  box-sizing: border-box;\n  user-select: none;\n\n  > svg {\n    width: ",";\n    height: ",";\n  }\n"],["\n  position: absolute;\n  top: 1px;\n  bottom: 1px;\n  z-index: 2;\n\n  display: flex;\n  flex-direction: column;\n  justify-content: center;\n\n  padding: 0 8px;\n\n  box-sizing: border-box;\n  user-select: none;\n\n  > svg {\n    width: ",";\n    height: ",";\n  }\n"])),v,v),t.SearchIcon=u.default.div(p||(p=r(["\n  left: 1px;\n\n  ","\n\n  color: ",";\n"],["\n  left: 1px;\n\n  ","\n\n  color: ",";\n"])),t.IconStyle,l.colors.gray100)},65234:function(e,t,n){var r=this&&this.__importDefault||function(e){return e&&e.__esModule?e:{default:e}};Object.defineProperty(t,"__esModule",{value:!0}),t.Search=void 0;var a=r(n(67294)),o=r(n(87351)),i=n(14428),l=n(23636),u=n(60755),d=r(n(65721));t.Search=function(){var e=(0,l.useFieldSearch)(),t=e[0],n=e[1];return a.default.createElement(u.Wrapper,null,a.default.createElement(u.SearchBlock,null,a.default.createElement(u.SearchIcon,null,a.default.createElement(d.default,null)),a.default.createElement(i.Filter,null),a.default.createElement(u.SearchBar,{type:"text",placeholder:(0,o.default)("Search fields..."),className:"fullwidth text",value:t,onChange:function(e){n(e.target.value)}})))}},67088:function(e,t,n){var r=this&&this.__makeTemplateObject||function(e,t){return Object.defineProperty?Object.defineProperty(e,"raw",{value:t}):e.raw=t,e},a=this&&this.__importDefault||function(e){return e&&e.__esModule?e:{default:e}};Object.defineProperty(t,"__esModule",{value:!0}),t.Grid=void 0;var o,i=a(n(12788));t.Grid=i.default.div(o||(o=r(["\n  position: relative;\n  display: flex;\n  gap: 0;\n\n  height: 100%;\n  background: #fff;\n"],["\n  position: relative;\n  display: flex;\n  gap: 0;\n\n  height: 100%;\n  background: #fff;\n"])))},86419:function(e,t,n){var r=this&&this.__importDefault||function(e){return e&&e.__esModule?e:{default:e}};Object.defineProperty(t,"__esModule",{value:!0}),t.LayoutEditor=void 0;var a=r(n(67294)),o=n(47785),i=n(4524),l=n(45323),u=n(19149),d=n(18842),c=n(24128),s=n(67088);t.LayoutEditor=function(){var e=(0,o.useSelector)(l.selectFocus).active;return a.default.createElement(s.Grid,null,a.default.createElement(i.SidebarSlider,{swiped:e},a.default.createElement(c.PropertyEditor,null),a.default.createElement(d.FieldList,null)),a.default.createElement(u.FieldLayout,null))}},91198:function(e,t,n){var r,a=this&&this.__extends||(r=function(e,t){return r=Object.setPrototypeOf||{__proto__:[]}instanceof Array&&function(e,t){e.__proto__=t}||function(e,t){for(var n in t)Object.prototype.hasOwnProperty.call(t,n)&&(e[n]=t[n])},r(e,t)},function(e,t){if("function"!=typeof t&&null!==t)throw new TypeError("Class extends value "+String(t)+" is not a constructor or null");function n(){this.constructor=e}r(e,t),e.prototype=null===t?Object.create(t):(n.prototype=t.prototype,new n)}),o=this&&this.__importDefault||function(e){return e&&e.__esModule?e:{default:e}};Object.defineProperty(t,"__esModule",{value:!0}),t.ErrorBoundary=void 0;var i=o(n(67294)),l=function(e){function t(t){var n=e.call(this,t)||this;return n.state={hasError:!1},n}return a(t,e),t.getDerivedStateFromError=function(){return{hasError:!0}},t.prototype.componentDidCatch=function(e,t){console.log(e,t)},t.prototype.render=function(){return this.state.hasError?i.default.createElement("div",null,this.props.message):this.props.children},t}(i.default.Component);t.ErrorBoundary=l},65556:function(e,t,n){var r=this&&this.__createBinding||(Object.create?function(e,t,n,r){void 0===r&&(r=n);var a=Object.getOwnPropertyDescriptor(t,n);a&&!("get"in a?!t.__esModule:a.writable||a.configurable)||(a={enumerable:!0,get:function(){return t[n]}}),Object.defineProperty(e,r,a)}:function(e,t,n,r){void 0===r&&(r=n),e[r]=t[n]}),a=this&&this.__setModuleDefault||(Object.create?function(e,t){Object.defineProperty(e,"default",{enumerable:!0,value:t})}:function(e,t){e.default=t}),o=this&&this.__importStar||function(e){if(e&&e.__esModule)return e;var t={};if(null!=e)for(var n in e)"default"!==n&&Object.prototype.hasOwnProperty.call(e,n)&&r(t,e,n);return a(t,e),t};Object.defineProperty(t,"__esModule",{value:!0}),t.EditableComponent=void 0;var i=o(n(67294)),l=n(85767),u=n(91198),d=o(n(84747));t.EditableComponent=function(e){var t=e.field,n=e.property,r=(0,l.useAppDispatch)(),a=d[n.type];return void 0===a?i.default.createElement("div",null,"...".concat(n.handle," <").concat(n.type,">")):(a.displayName="Field <".concat(n.type,">"),i.default.createElement(u.ErrorBoundary,{message:"...".concat(n.handle," <").concat(n.type,">")},i.default.createElement(i.Suspense,null,i.default.createElement(a,{field:t,property:n,dispatch:r}))))}},67230:function(e,t,n){var r=this&&this.__importDefault||function(e){return e&&e.__esModule?e:{default:e}};Object.defineProperty(t,"__esModule",{value:!0}),t.FieldProperties=void 0;var a=r(n(67294)),o=r(n(84849)),i=n(47785),l=n(56859),u=n(42367),d=n(24741),c=n(65556);t.FieldProperties=function(e){var t=e.uid,n=(0,u.useFetchFieldPropertySections)(),r=n.data,s=n.isFetching,f=(0,i.useSelector)((0,l.selectField)(t)),p=(0,u.useFieldType)(null==f?void 0:f.typeClass);return f&&p?!r&&s?a.default.createElement("div",null,a.default.createElement("h2",null,p.name),a.default.createElement(d.SectionWrapper,null,a.default.createElement(o.default,null))):a.default.createElement(a.default.Fragment,null,a.default.createElement(d.Title,null,a.default.createElement(d.Icon,{dangerouslySetInnerHTML:{__html:p.icon}}),a.default.createElement("span",null,p.name)),a.default.createElement(d.SectionWrapper,null,r.map((function(e){var t=e.handle,n=e.label;return a.default.createElement(d.SectionBlock,{label:n,key:t},p.properties.filter((function(e){return e.section===t})).sort((function(e,t){return e.order-t.order})).map((function(e){return a.default.createElement(c.EditableComponent,{key:e.handle,field:f,property:e})})))})))):a.default.createElement("div",null,"Not found")}},48170:function(e,t,n){var r=this&&this.__makeTemplateObject||function(e,t){return Object.defineProperty?Object.defineProperty(e,"raw",{value:t}):e.raw=t,e},a=this&&this.__importDefault||function(e){return e&&e.__esModule?e:{default:e}};Object.defineProperty(t,"__esModule",{value:!0}),t.CheckboxWrapper=void 0;var o,i=n(240),l=a(n(12788));t.CheckboxWrapper=l.default.div(o||(o=r(["\n  display: flex;\n  gap: ",";\n"],["\n  display: flex;\n  gap: ",";\n"])),i.spacings.sm)},26902:function(e,t,n){var r=this&&this.__importDefault||function(e){return e&&e.__esModule?e:{default:e}};Object.defineProperty(t,"__esModule",{value:!0});var a=r(n(67294)),o=n(3399),i=n(56859),l=n(48170),u=n(39452);t.default=function(e){var t,n=e.field,r=e.property,d=e.dispatch,c=r.handle,s=r.label,f=n.uid,p=null!==(t=n.properties[c])&&void 0!==t&&t;return a.default.createElement(u.ControlWrapper,null,a.default.createElement(l.CheckboxWrapper,null,a.default.createElement(o.CheckboxInput,{id:c,label:s,onClick:function(){return d((0,i.edit)({uid:f,property:r,value:!p}))}})))}},39452:function(e,t,n){var r=this&&this.__makeTemplateObject||function(e,t){return Object.defineProperty?Object.defineProperty(e,"raw",{value:t}):e.raw=t,e},a=this&&this.__importDefault||function(e){return e&&e.__esModule?e:{default:e}};Object.defineProperty(t,"__esModule",{value:!0}),t.Label=t.ControlWrapper=void 0;var o,i,l=a(n(12788));t.ControlWrapper=l.default.div(o||(o=r([""],[""]))),t.Label=l.default.label(i||(i=r(["\n  display: block;\n  font-weight: ",";\n"],["\n  display: block;\n  font-weight: ",";\n"])),(function(e){return e.regular?"normal":"bold"}))},67569:function(e,t,n){var r=this&&this.__importDefault||function(e){return e&&e.__esModule?e:{default:e}};Object.defineProperty(t,"__esModule",{value:!0}),t.Control=void 0;var a=r(n(67294)),o=n(39452);t.Control=function(e){var t=e.children,n=e.property;return a.default.createElement(o.ControlWrapper,null,a.default.createElement(o.Label,{htmlFor:n.handle},n.label),t)}},84747:function(e,t,n){var r=this&&this.__importDefault||function(e){return e&&e.__esModule?e:{default:e}};Object.defineProperty(t,"__esModule",{value:!0}),t.textarea=t.string=t.int=t.bool=void 0;var a=n(26902);Object.defineProperty(t,"bool",{enumerable:!0,get:function(){return r(a).default}});var o=n(36497);Object.defineProperty(t,"int",{enumerable:!0,get:function(){return r(o).default}});var i=n(20551);Object.defineProperty(t,"string",{enumerable:!0,get:function(){return r(i).default}});var l=n(53705);Object.defineProperty(t,"textarea",{enumerable:!0,get:function(){return r(l).default}})},36497:function(e,t,n){var r=this&&this.__importDefault||function(e){return e&&e.__esModule?e:{default:e}};Object.defineProperty(t,"__esModule",{value:!0});var a=r(n(67294)),o=n(56859),i=n(67569);t.default=function(e){var t,n=e.field,r=e.property,l=e.dispatch,u=r.handle,d=n.uid;return a.default.createElement(i.Control,{property:r},a.default.createElement("input",{id:u,type:"number",className:"text fullwidth",value:null!==(t=n.properties[u])&&void 0!==t?t:"",onChange:function(e){return l((0,o.edit)({uid:d,property:r,value:Number(e.target.value)}))}}))}},20551:function(e,t,n){var r=this&&this.__importDefault||function(e){return e&&e.__esModule?e:{default:e}};Object.defineProperty(t,"__esModule",{value:!0});var a=r(n(67294)),o=n(56859),i=r(n(89361)),l=n(67569);t.default=function(e){var t,n=e.field,r=e.property,u=e.dispatch,d=r.handle,c=n.uid,s=n.properties;return a.default.createElement(l.Control,{property:r},a.default.createElement("input",{id:d,type:"text",className:(0,i.default)("text","fullwidth",r.flags.includes("code")&&"code"),value:null!==(t=s[d])&&void 0!==t?t:"",placeholder:r.placeholder,onChange:function(e){return u((0,o.edit)({uid:c,property:r,value:e.target.value}))}}))}},53705:function(e,t,n){var r=this&&this.__importDefault||function(e){return e&&e.__esModule?e:{default:e}};Object.defineProperty(t,"__esModule",{value:!0});var a=r(n(67294)),o=n(56859),i=r(n(89361)),l=n(67569);t.default=function(e){var t,n=e.field,r=e.property,u=e.dispatch,d=r.handle,c=n.uid,s=n.properties;return a.default.createElement(l.Control,{property:r},a.default.createElement("textarea",{id:d,className:(0,i.default)("text","fullwidth",r.flags.includes("code")&&"code"),rows:2,value:null!==(t=s[d])&&void 0!==t?t:"",placeholder:r.placeholder,onChange:function(e){return u((0,o.edit)({uid:c,property:r,value:e.target.value}))}}))}},24741:function(e,t,n){var r=this&&this.__makeTemplateObject||function(e,t){return Object.defineProperty?Object.defineProperty(e,"raw",{value:t}):e.raw=t,e},a=this&&this.__importDefault||function(e){return e&&e.__esModule?e:{default:e}};Object.defineProperty(t,"__esModule",{value:!0}),t.SectionBlock=t.SectionWrapper=t.Icon=t.Title=t.CloseLink=t.PropertyEditorWrapper=void 0;var o,i,l,u,d,c,s=n(240),f=a(n(12788));t.PropertyEditorWrapper=f.default.div(o||(o=r(["\n  position: relative;\n  padding: ",";\n"],["\n  position: relative;\n  padding: ",";\n"])),s.spacings.lg),t.CloseLink=f.default.a(i||(i=r(["\n  position: absolute;\n  right: 10px;\n  top: 17px;\n\n  display: block;\n  width: 20px;\n  height: 20px;\n"],["\n  position: absolute;\n  right: 10px;\n  top: 17px;\n\n  display: block;\n  width: 20px;\n  height: 20px;\n"]))),t.Title=f.default.h3(l||(l=r(["\n  display: flex;\n  justify-content: flex-start;\n  align-items: center;\n  gap: ",";\n\n  margin: 0;\n  font-size: 16px;\n"],["\n  display: flex;\n  justify-content: flex-start;\n  align-items: center;\n  gap: ",";\n\n  margin: 0;\n  font-size: 16px;\n"])),s.spacings.sm),t.Icon=f.default.div(u||(u=r(["\n  width: 20px;\n  height: 20px;\n"],["\n  width: 20px;\n  height: 20px;\n"]))),t.SectionWrapper=f.default.div(d||(d=r(["\n  display: flex;\n  flex-direction: column;\n  gap: ",";\n"],["\n  display: flex;\n  flex-direction: column;\n  gap: ",";\n"])),s.spacings.lg),t.SectionBlock=f.default.section(c||(c=r(["\n  position: relative;\n\n  display: flex;\n  flex-direction: column;\n  gap: ",";\n\n  margin-top: ",";\n  padding-top: ",";\n\n  &:empty {\n    display: none;\n  }\n\n  &:before {\n    content: '';\n\n    position: absolute;\n    left: 0;\n    top: 0;\n    right: 0;\n\n    display: block;\n    height: 1px;\n\n    margin: 0 -18px;\n\n    box-shadow: ",";\n  }\n\n  &:after {\n    content: '","';\n\n    position: absolute;\n    left: -5px;\n    top: -7px;\n\n    display: block;\n    padding: 0 5px;\n\n    background-color: ",";\n\n    color: rgb(154 165 177 / 75%);\n    font-family: system-ui, BlinkMacSystemFont, -apple-system, Segoe UI, Roboto,\n      Oxygen, Ubuntu, Cantarell, Fira Sans, Droid Sans, Helvetica Neue,\n      sans-serif;\n    font-size: 11px;\n    font-weight: bold;\n    text-transform: uppercase;\n  }\n"],["\n  position: relative;\n\n  display: flex;\n  flex-direction: column;\n  gap: ",";\n\n  margin-top: ",";\n  padding-top: ",";\n\n  &:empty {\n    display: none;\n  }\n\n  &:before {\n    content: '';\n\n    position: absolute;\n    left: 0;\n    top: 0;\n    right: 0;\n\n    display: block;\n    height: 1px;\n\n    margin: 0 -18px;\n\n    box-shadow: ",";\n  }\n\n  &:after {\n    content: '","';\n\n    position: absolute;\n    left: -5px;\n    top: -7px;\n\n    display: block;\n    padding: 0 5px;\n\n    background-color: ",";\n\n    color: rgb(154 165 177 / 75%);\n    font-family: system-ui, BlinkMacSystemFont, -apple-system, Segoe UI, Roboto,\n      Oxygen, Ubuntu, Cantarell, Fira Sans, Droid Sans, Helvetica Neue,\n      sans-serif;\n    font-size: 11px;\n    font-weight: bold;\n    text-transform: uppercase;\n  }\n"])),s.spacings.md,s.spacings.lg,s.spacings.lg,s.shadows.bottom,(function(e){return e.label}),s.colors.gray050)},24128:function(e,t,n){var r=this&&this.__createBinding||(Object.create?function(e,t,n,r){void 0===r&&(r=n);var a=Object.getOwnPropertyDescriptor(t,n);a&&!("get"in a?!t.__esModule:a.writable||a.configurable)||(a={enumerable:!0,get:function(){return t[n]}}),Object.defineProperty(e,r,a)}:function(e,t,n,r){void 0===r&&(r=n),e[r]=t[n]}),a=this&&this.__setModuleDefault||(Object.create?function(e,t){Object.defineProperty(e,"default",{enumerable:!0,value:t})}:function(e,t){e.default=t}),o=this&&this.__importStar||function(e){if(e&&e.__esModule)return e;var t={};if(null!=e)for(var n in e)"default"!==n&&Object.prototype.hasOwnProperty.call(e,n)&&r(t,e,n);return a(t,e),t},i=this&&this.__importDefault||function(e){return e&&e.__esModule?e:{default:e}};Object.defineProperty(t,"__esModule",{value:!0}),t.PropertyEditor=void 0;var l=o(n(67294)),u=n(47785),d=n(85767),c=n(45323),s=i(n(33742)),f=n(12020),p=n(91198),v=n(67230),h=n(24741);t.PropertyEditor=function(){var e=(0,u.useSelector)(c.selectFocus),t=e.active,n=e.type,r=e.uid,a=(0,d.useAppDispatch)(),o=(0,l.useRef)(null);(0,f.useOnKeypress)({meetsCondition:t,callback:function(e){"Escape"===e.key&&a((0,c.unfocus)())}});var i=l.default.createElement(v.FieldProperties,{uid:r});return l.default.createElement(h.PropertyEditorWrapper,{ref:o},l.default.createElement(h.CloseLink,{onClick:function(){return a((0,c.unfocus)())}},l.default.createElement(s.default,null)),l.default.createElement(p.ErrorBoundary,{message:'Could not load property editor for "'.concat(n,'" type')},i))}},11037:function(e,t){Object.defineProperty(t,"__esModule",{value:!0}),t.CellType=void 0,function(e){e.Field="field",e.Layout="layout"}(t.CellType||(t.CellType={}))},70774:function(e,t,n){var r=this&&this.__makeTemplateObject||function(e,t){return Object.defineProperty?Object.defineProperty(e,"raw",{value:t}):e.raw=t,e};Object.defineProperty(t,"__esModule",{value:!0}),t.EditorGlobalStyles=void 0;var a,o=n(12788);t.EditorGlobalStyles=(0,o.createGlobalStyle)(a||(a=r(["\n  #global-header, #global-footer {\n    display: none;\n  }\n\n  #main-container {\n    #main {\n      #main-content {\n        padding: 0;\n        width: calc(100vw - 50px);\n      }\n    }\n  }\n\n  #global-sidebar {\n    width: 50px;\n    overflow: hidden;\n\n    #system-name {\n      display: none;\n    }\n\n    nav#nav {\n      margin: 8px 0 0;\n\n      > ul > li {\n        > a {\n          padding: 8px 8px;\n        }\n\n        .label {\n          display: none;\n        }\n\n        .subnav {\n          display: none;\n        }\n      }\n    }\n  }\n\n  #page-container {\n    padding-left: 50px !important;\n  }\n\n  #freeform-client-app {\n    height: 100vh;\n  }\n"],["\n  #global-header, #global-footer {\n    display: none;\n  }\n\n  #main-container {\n    #main {\n      #main-content {\n        padding: 0;\n        width: calc(100vw - 50px);\n      }\n    }\n  }\n\n  #global-sidebar {\n    width: 50px;\n    overflow: hidden;\n\n    #system-name {\n      display: none;\n    }\n\n    nav#nav {\n      margin: 8px 0 0;\n\n      > ul > li {\n        > a {\n          padding: 8px 8px;\n        }\n\n        .label {\n          display: none;\n        }\n\n        .subnav {\n          display: none;\n        }\n      }\n    }\n  }\n\n  #page-container {\n    padding-left: 50px !important;\n  }\n\n  #freeform-client-app {\n    height: 100vh;\n  }\n"])))},66444:function(e,t,n){var r=this&&this.__importDefault||function(e){return e&&e.__esModule?e:{default:e}};Object.defineProperty(t,"__esModule",{value:!0}),t.Edit=void 0;var a=r(n(67294)),o=n(49818),i=n(23493),l=n(93787);t.Edit=function(){var e=(0,o.useParams)().formId;(0,i.useQueryFormSettings)();var t=(0,i.useQuerySingleForm)(e&&Number(e)),n=t.isFetching,r=t.isError,u=t.error;return n?a.default.createElement("div",null,"Fetching ",e,"..."):r?a.default.createElement("div",null,"ERROR: ",u.message):a.default.createElement(l.Builder,null)}},33922:function(e,t,n){var r=this&&this.__importDefault||function(e){return e&&e.__esModule?e:{default:e}};Object.defineProperty(t,"__esModule",{value:!0}),t.EditProvider=void 0;var a=r(n(67294)),o=n(47785),i=n(66444),l=n(70774),u=n(85767);t.EditProvider=function(){return a.default.createElement(o.Provider,{store:u.store},a.default.createElement(l.EditorGlobalStyles,null),a.default.createElement(i.Edit,null))}},80926:function(e,t,n){Object.defineProperty(t,"__esModule",{value:!0}),t.save=void 0;var r=n(20991);t.save=(0,r.createAction)("form/save")},85767:function(e,t,n){var r=this&&this.__importDefault||function(e){return e&&e.__esModule?e:{default:e}};Object.defineProperty(t,"__esModule",{value:!0}),t.useAppSelector=t.useAppDispatch=t.store=void 0;var a=n(47785),o=n(20991),i=r(n(53894)),l=n(90531),u=r(n(89372)),d=r(n(45323)),c=r(n(99614)),s=r(n(56859)),f=r(n(61552)),p=r(n(68887)),v=r(n(1965)),h=r(n(68974)),m=r(n(33314)),g=r(n(65673));t.store=(0,o.configureStore)({middleware:[i.default,l.statePersistMiddleware],reducer:{form:f.default,fields:s.default,context:d.default,integrations:p.default,layouts:v.default,pages:h.default,rows:m.default,cells:u.default,search:g.default,drag:c.default}});t.useAppDispatch=function(){return(0,a.useDispatch)()},t.useAppSelector=a.useSelector},90531:function(e,t,n){var r=this&&this.__importDefault||function(e){return e&&e.__esModule?e:{default:e}};Object.defineProperty(t,"__esModule",{value:!0}),t.statePersistMiddleware=t.TOPIC_UPDATED=t.TOPIC_CREATED=t.TOPIC_ERRORS=t.TOPIC_SAVE=void 0;var a=r(n(9669)),o=r(n(1798)),i=n(80926);t.TOPIC_SAVE=Symbol("form.save"),t.TOPIC_ERRORS=Symbol("form.save.errors"),t.TOPIC_CREATED=Symbol("form.save.crated"),t.TOPIC_UPDATED=Symbol("form.save.updated"),o.default.clearAllSubscriptions();var l=function(e,n){o.default.publish(t.TOPIC_ERRORS,{dispatch:e,response:n})};t.statePersistMiddleware=function(e){return function(n){return function(r){if(r&&(n(r),r.type===String(i.save))){var u=e.dispatch,d={dispatch:u,state:e.getState(),persist:{}};o.default.publishSync(t.TOPIC_SAVE,d);var c=d.state.form.id;c?a.default.put("/client/api/forms/".concat(c),d.persist).then((function(e){return function(e,n){o.default.publish(t.TOPIC_UPDATED,{dispatch:e,response:n})}(u,e)})).catch((function(e){return l(u,e)})):a.default.post("/client/api/forms",d.persist).then((function(e){return function(e,n){o.default.publish(t.TOPIC_CREATED,{dispatch:e,response:n})}(u,e)})).catch((function(e){return l(u,e)}))}}}}},89372:function(e,t,n){var r,a=this&&this.__assign||function(){return a=Object.assign||function(e){for(var t,n=1,r=arguments.length;n<r;n++)for(var a in t=arguments[n])Object.prototype.hasOwnProperty.call(t,a)&&(e[a]=t[a]);return e},a.apply(this,arguments)},o=this&&this.__spreadArray||function(e,t,n){if(n||2===arguments.length)for(var r,a=0,o=t.length;a<o;a++)!r&&a in t||(r||(r=Array.prototype.slice.call(t,0,a)),r[a]=t[a]);return e.concat(r||Array.prototype.slice.call(t))};Object.defineProperty(t,"__esModule",{value:!0}),t.selectCellsInRow=t.remove=t.add=t.moveTo=t.set=t.cellsSlice=void 0;var i=n(20991);t.cellsSlice=(0,i.createSlice)({name:"cells",initialState:[],reducers:{set:function(e,t){e.splice.apply(e,o([0,e.length],t.payload,!1))},add:function(e,t){var n=Math.max.apply(Math,o([-1],e.filter((function(e){return e.rowUid===t.payload.rowUid})).map((function(e){return e.order})),!1));e.push(a(a({},t.payload),{order:n+1}))},remove:function(e,t){e=e.filter((function(e){return e.uid!==t.payload}))},moveTo:function(e,t){console.log(t)}}}),t.set=(r=t.cellsSlice.actions).set,t.moveTo=r.moveTo,t.add=r.add,t.remove=r.remove;t.selectCellsInRow=function(e){return function(t){return t.cells.filter((function(t){return t.rowUid===e.uid})).sort((function(e,t){return e.order-t.order}))}},t.default=t.cellsSlice.reducer},45323:function(e,t,n){var r,a=this&&this.__assign||function(){return a=Object.assign||function(e){for(var t,n=1,r=arguments.length;n<r;n++)for(var a in t=arguments[n])Object.prototype.hasOwnProperty.call(t,a)&&(e[a]=t[a]);return e},a.apply(this,arguments)};Object.defineProperty(t,"__esModule",{value:!0}),t.selectFocus=t.selectCurrentPage=t.unfocus=t.setFocusedItem=t.setPage=t.FocusType=void 0;var o=n(20991);!function(e){e.Page="page",e.Field="field",e.Row="row",e.Cell="cell"}(t.FocusType||(t.FocusType={}));var i=(0,o.createSlice)({name:"context",initialState:{page:null,focus:{active:!1,type:null,uid:null}},reducers:{setPage:function(e,t){var n=t.payload;e.page=n},setFocusedItem:function(e,t){var n=t.payload;e.focus=a({active:!0},n)},focus:function(e){e.focus.active=!0},unfocus:function(e){e.focus.active=!1}}});t.setPage=(r=i.actions).setPage,t.setFocusedItem=r.setFocusedItem,t.unfocus=r.unfocus;t.selectCurrentPage=function(e){var t=e.context.page;return t?e.pages.find((function(e){return e.uid===t})):e.pages.find(Boolean)};t.selectFocus=function(e){return e.context.focus},t.default=i.reducer},99614:function(e,t,n){var r;Object.defineProperty(t,"__esModule",{value:!0}),t.selectCurrentPosition=t.selectCurrentCell=t.selectCurrentRow=t.setPosition=t.setCell=t.setRow=t.dragSlice=void 0;var a=n(20991),o={cellUid:void 0,rowUid:void 0,position:void 0};t.dragSlice=(0,a.createSlice)({name:"drag",initialState:o,reducers:{setRow:function(e,t){e.rowUid=t.payload},setCell:function(e,t){e.cellUid=t.payload},setPosition:function(e,t){e.position=t.payload}}}),t.setRow=(r=t.dragSlice.actions).setRow,t.setCell=r.setCell,t.setPosition=r.setPosition;t.selectCurrentRow=function(e){return e.drag.rowUid};t.selectCurrentCell=function(e){return e.drag.cellUid};t.selectCurrentPosition=function(e){return e.drag.position},t.default=t.dragSlice.reducer},56859:function(e,t,n){var r,a=this&&this.__spreadArray||function(e,t,n){if(n||2===arguments.length)for(var r,a=0,o=t.length;a<o;a++)!r&&a in t||(r||(r=Array.prototype.slice.call(t,0,a)),r[a]=t[a]);return e.concat(r||Array.prototype.slice.call(t))},o=this&&this.__importDefault||function(e){return e&&e.__esModule?e:{default:e}};Object.defineProperty(t,"__esModule",{value:!0}),t.selectField=t.edit=t.remove=t.add=t.set=t.fieldsSlice=void 0;var i=n(90531),l=n(20991),u=o(n(1798));t.fieldsSlice=(0,l.createSlice)({name:"fields",initialState:[],reducers:{set:function(e,t){e.splice.apply(e,a([0,e.length],t.payload,!1))},add:function(e,t){var n=t.payload,r=n.uid,a=n.fieldType,o={};a.properties.forEach((function(e){return o[e.handle]=e.value})),e.push({uid:r,typeClass:a.typeClass,properties:o})},remove:function(e,t){var n=t.payload;e.splice(e.findIndex((function(e){return e.uid===n})),1)},edit:function(e,t){var n=t.payload,r=n.uid,a=n.property.handle,o=n.value;e.find((function(e){return e.uid===r})).properties[a]=o}}}),t.set=(r=t.fieldsSlice.actions).set,t.add=r.add,t.remove=r.remove,t.edit=r.edit;t.selectField=function(e){return function(t){return t.fields.find((function(t){return t.uid===e}))}},t.default=t.fieldsSlice.reducer;u.default.subscribe(i.TOPIC_SAVE,(function(e,t){var n=t.state;t.persist.fields=n.fields}))},61552:function(e,t,n){var r,a=this&&this.__importDefault||function(e){return e&&e.__esModule?e:{default:e}};Object.defineProperty(t,"__esModule",{value:!0}),t.selectFormSetting=t.selectFormSettings=t.selectForm=t.clearErrors=t.setErrors=t.removeError=t.addError=t.modifySettings=t.update=t.formSlice=void 0;var o=n(90531),i=n(90531),l=n(90531),u=n(20991),d=a(n(1798)),c={id:null,uid:(0,n(21614).v4)(),type:"Solspace\\Freeform\\Form\\Types\\Regular",name:"New Form",handle:"newForm",settings:{}};t.formSlice=(0,u.createSlice)({name:"form",initialState:c,reducers:{update:function(e,t){var n=t.payload;Object.assign(e,n)},modifySettings:function(e,t){var n=t.payload,r=n.namespace,a=n.key,o=n.value;e.settings[r]||(e.settings[r]={}),e.settings[r][a]=o},removeError:function(e,t){var n=t.payload;delete e.errors[n]},addError:function(e,t){var n=t.payload;e.errors[n.key]=n.errors},setErrors:function(e,t){var n=t.payload;e.errors=n},clearErrors:function(e){e.errors=void 0}}}),t.update=(r=t.formSlice.actions).update,t.modifySettings=r.modifySettings,t.addError=r.addError,t.removeError=r.removeError,t.setErrors=r.setErrors,t.clearErrors=r.clearErrors;t.selectForm=function(e){return e.form};t.selectFormSettings=function(e){return function(t){var n;return(null===(n=t.form.settings)||void 0===n?void 0:n[e])||{}}};t.selectFormSetting=function(e,t){return function(n){var r,a;return null===(a=null===(r=n.form.settings)||void 0===r?void 0:r[e])||void 0===a?void 0:a[t]}},t.default=t.formSlice.reducer;d.default.subscribe(l.TOPIC_SAVE,(function(e,t){var n=t.state,r=t.persist,a=n.form,o=a.id,i=a.uid,l=a.type,u=a.settings;r.form={id:o,uid:i,type:l,settings:u}})),d.default.subscribe(i.TOPIC_ERRORS,(function(e,n){var r,a=n.dispatch,o=n.response;a((0,t.clearErrors)()),a((0,t.setErrors)(null===(r=o.errors)||void 0===r?void 0:r.form))})),d.default.subscribe(o.TOPIC_CREATED,(function(e,n){var r=n.dispatch,a=n.response;r((0,t.update)({id:a.data.form.id}))}))},68887:function(e,t,n){var r,a=this&&this.__assign||function(){return a=Object.assign||function(e){for(var t,n=1,r=arguments.length;n<r;n++)for(var a in t=arguments[n])Object.prototype.hasOwnProperty.call(t,a)&&(e[a]=t[a]);return e},a.apply(this,arguments)};Object.defineProperty(t,"__esModule",{value:!0}),t.selectIntegration=t.modifyIntegrationSetting=t.toggleIntegration=t.addIntegrations=t.integrationsSlice=void 0;var o=n(20991),i=function(e,t){return e.find((function(e){return e.id===t}))};t.integrationsSlice=(0,o.createSlice)({name:"integrations",initialState:[],reducers:{addIntegrations:function(e,t){t.payload.forEach((function(t){e.push(a({updated:1,dirty:!1},t))}))},toggleIntegration:function(e,t){var n=i(e,t.payload);n.enabled=!n.enabled},modifyIntegrationSetting:function(e,t){var n=t.payload,r=n.id,a=n.key,o=n.value,l=function(e,t){return e.settings.find((function(e){return e.handle===t}))}(i(e,r),a);l.value=o}}}),t.addIntegrations=(r=t.integrationsSlice.actions).addIntegrations,t.toggleIntegration=r.toggleIntegration,t.modifyIntegrationSetting=r.modifyIntegrationSetting;t.selectIntegration=function(e){return function(t){return t.integrations.find((function(t){return t.id===e}))}},t.default=t.integrationsSlice.reducer},1965:function(e,t,n){var r,a=this&&this.__spreadArray||function(e,t,n){if(n||2===arguments.length)for(var r,a=0,o=t.length;a<o;a++)!r&&a in t||(r||(r=Array.prototype.slice.call(t,0,a)),r[a]=t[a]);return e.concat(r||Array.prototype.slice.call(t))};Object.defineProperty(t,"__esModule",{value:!0}),t.selectPageLayout=t.selectLayout=t.remove=t.add=t.set=t.layoutsSlice=void 0;var o=n(90531),i=n(20991);t.layoutsSlice=(0,i.createSlice)({name:"layouts",initialState:[{uid:"layout-uid-1"},{uid:"layout-uid-2"}],reducers:{set:function(e,t){e.splice.apply(e,a([0,e.length],t.payload,!1))},add:function(e,t){e.push(t.payload)},remove:function(e,t){e=e.filter((function(e){return e.uid!==t.payload}))}}}),t.set=(r=t.layoutsSlice.actions).set,t.add=r.add,t.remove=r.remove;t.selectLayout=function(e){return function(t){return t.layouts.find((function(t){return t.uid===e}))}};t.selectPageLayout=function(e){return function(t){return t.layouts.find((function(t){return t.uid===(null==e?void 0:e.layoutUid)}))}},t.default=t.layoutsSlice.reducer;PubSub.subscribe(o.TOPIC_SAVE,(function(e,t){var n=t.state,r=t.persist,a=n.layouts,o=n.cells,i=n.rows,l=n.pages;r.layout={pages:l,layouts:a,rows:i,cells:o}}))},68974:function(e,t,n){var r,a=this&&this.__assign||function(){return a=Object.assign||function(e){for(var t,n=1,r=arguments.length;n<r;n++)for(var a in t=arguments[n])Object.prototype.hasOwnProperty.call(t,a)&&(e[a]=t[a]);return e},a.apply(this,arguments)},o=this&&this.__spreadArray||function(e,t,n){if(n||2===arguments.length)for(var r,a=0,o=t.length;a<o;a++)!r&&a in t||(r||(r=Array.prototype.slice.call(t,0,a)),r[a]=t[a]);return e.concat(r||Array.prototype.slice.call(t))};Object.defineProperty(t,"__esModule",{value:!0}),t.selectPage=t.selectPages=t.remove=t.swap=t.add=t.set=t.pagesSlice=void 0;var i=n(20991);t.pagesSlice=(0,i.createSlice)({name:"pages",initialState:[],reducers:{set:function(e,t){e.splice.apply(e,o([0,e.length],t.payload,!1))},add:function(e,t){var n,r=null!==(n=Math.max.apply(Math,e.map((function(e){return e.order}))))&&void 0!==n?n:-1;e.push(a(a({},t.payload),{order:r+1}))},remove:function(e,t){var n=0;e=e.filter((function(e){return e.uid!==t.payload})).map((function(e){return a(a({},e),{order:n++})}))},swap:function(e,t){var n=e.find((function(e){return e.uid===t.payload.currentUid})),r=e.find((function(e){return e.uid===t.payload.targetUid})),a=n.order;n.order=r.order,r.order=a}}}),t.set=(r=t.pagesSlice.actions).set,t.add=r.add,t.swap=r.swap,t.remove=r.remove;t.selectPages=function(e){return e.pages};t.selectPage=function(e){return function(t){return t.pages.find((function(t){return t.uid===e}))}},t.default=t.pagesSlice.reducer},33314:function(e,t,n){var r,a=this&&this.__spreadArray||function(e,t,n){if(n||2===arguments.length)for(var r,a=0,o=t.length;a<o;a++)!r&&a in t||(r||(r=Array.prototype.slice.call(t,0,a)),r[a]=t[a]);return e.concat(r||Array.prototype.slice.call(t))};Object.defineProperty(t,"__esModule",{value:!0}),t.selectRowsInLayout=t.remove=t.swap=t.add=t.set=t.rowsSlice=void 0;var o=n(20991);t.rowsSlice=(0,o.createSlice)({name:"rows",initialState:[],reducers:{set:function(e,t){e.splice.apply(e,a([0,e.length],t.payload,!1))},add:function(e,t){var n=t.payload,r=n.layoutUid,o=n.uid,i=Math.max.apply(Math,a([-1],e.filter((function(e){return e.layoutUid===r})).map((function(e){return e.order})),!1));e.push({uid:o,order:i+1,layoutUid:r})},remove:function(e,t){e=e.filter((function(e){return e.uid!==t.payload}))},swap:function(e,t){var n=e.find((function(e){return e.uid===t.payload.currentUid})),r=e.find((function(e){return e.uid===t.payload.targetUid})),a=n.order;n.order=r.order,r.order=a}}}),t.set=(r=t.rowsSlice.actions).set,t.add=r.add,t.swap=r.swap,t.remove=r.remove;t.selectRowsInLayout=function(e){return function(t){return e?t.rows.filter((function(t){return t.layoutUid===e.uid})).sort((function(e,t){return e.order-t.order})):[]}},t.default=t.rowsSlice.reducer},65673:function(e,t,n){var r;Object.defineProperty(t,"__esModule",{value:!0}),t.selectQuery=t.clearQuery=t.updateQuery=t.searchSlice=t.Search=void 0;var a=n(20991);!function(e){e.Fields="fields"}(t.Search||(t.Search={}));t.searchSlice=(0,a.createSlice)({name:"search",initialState:{fields:""},reducers:{updateQuery:function(e,t){e[t.payload.type]=t.payload.query},clearQuery:function(e,t){e[t.payload]=""}}}),t.updateQuery=(r=t.searchSlice.actions).updateQuery,t.clearQuery=r.clearQuery;t.selectQuery=function(e){return function(t){return t.search[e]}},t.default=t.searchSlice.reducer},73396:function(e,t,n){Object.defineProperty(t,"__esModule",{value:!0}),t.addNewField=void 0;var r=n(11037),a=n(89372),o=n(45323),i=n(56859),l=n(33314),u=n(21614);t.addNewField=function(e){return function(t,n){var d=(0,u.v4)(),c=(0,u.v4)(),s=(0,u.v4)(),f=n(),p=(0,o.selectCurrentPage)(f);if(!p)throw new Error("No pages present");t((0,i.add)({fieldType:e,uid:d})),t((0,l.add)({layoutUid:p.layoutUid,uid:s})),t((0,a.add)({type:r.CellType.Field,rowUid:s,targetUid:d,uid:c}))}}},6175:function(e,t,n){Object.defineProperty(t,"__esModule",{value:!0}),t.addNewPage=void 0;var r=n(1965),a=n(68974),o=n(21614);t.addNewPage=function(){return function(e,t){var n=(0,o.v4)(),i=(0,o.v4)(),l=t().pages.length+1;e((0,r.add)({uid:i})),e((0,a.add)({uid:n,label:"Page ".concat(l),layoutUid:i,order:l}))}}},89141:function(e,t,n){Object.defineProperty(t,"__esModule",{value:!0}),t.Forms=t.Form=void 0;var r=n(33922);Object.defineProperty(t,"Form",{enumerable:!0,get:function(){return r.EditProvider}});var a=n(85581);Object.defineProperty(t,"Forms",{enumerable:!0,get:function(){return a.List}})},62819:function(e,t,n){var r=this&&this.__makeTemplateObject||function(e,t){return Object.defineProperty?Object.defineProperty(e,"raw",{value:t}):e.raw=t,e},a=this&&this.__importDefault||function(e){return e&&e.__esModule?e:{default:e}};Object.defineProperty(t,"__esModule",{value:!0}),t.Subtitle=t.Title=t.Card=t.Wrapper=void 0;var o,i,l,u,d=a(n(12788));t.Wrapper=d.default.ul(o||(o=r(["\n  display: grid;\n  grid-template-columns: repeat(4, 1fr);\n  gap: var(--l);\n\n  margin-top: var(--l);\n"],["\n  display: grid;\n  grid-template-columns: repeat(4, 1fr);\n  gap: var(--l);\n\n  margin-top: var(--l);\n"]))),t.Card=d.default.li(i||(i=r(["\n  padding: var(--xl);\n\n  background: #fff;\n  box-shadow: 0 0 0 1px #cdd8e4, 0 2px 12px rgb(205 216 228 / 50%);\n  border-radius: var(--large-border-radius);\n"],["\n  padding: var(--xl);\n\n  background: #fff;\n  box-shadow: 0 0 0 1px #cdd8e4, 0 2px 12px rgb(205 216 228 / 50%);\n  border-radius: var(--large-border-radius);\n"]))),t.Title=d.default.h3(l||(l=r(["\n  margin: 0;\n\n  font-size: 22px;\n  font-weight: normal;\n  text-align: center;\n"],["\n  margin: 0;\n\n  font-size: 22px;\n  font-weight: normal;\n  text-align: center;\n"]))),t.Subtitle=d.default.small(u||(u=r(["\n  display: block;\n\n  color: grey;\n  font-size: 12px;\n  font-family: monospace;\n  text-align: center;\n"],["\n  display: block;\n\n  color: grey;\n  font-size: 12px;\n  font-family: monospace;\n  text-align: center;\n"])))},85581:function(e,t,n){var r=this&&this.__importDefault||function(e){return e&&e.__esModule?e:{default:e}};Object.defineProperty(t,"__esModule",{value:!0}),t.List=void 0;var a=r(n(67294)),o=n(49818),i=n(23493),l=n(62819);t.List=function(){var e=(0,i.useQueryForms)(),t=e.data,n=e.isFetching,r=e.isError,u=e.error;return!t&&n?a.default.createElement("div",null,"fetching forms..."):r?a.default.createElement("div",null,"ERROR ",u.message):a.default.createElement("div",null,a.default.createElement("h1",null,"Forms",n&&a.default.createElement("span",null,"is fetching")),a.default.createElement(l.Wrapper,null,a.default.createElement(l.Card,null,a.default.createElement(l.Title,null,a.default.createElement(o.Link,{to:"new"},"Create new Form")),a.default.createElement(l.Subtitle,null,"click me")),t.map((function(e){return a.default.createElement(l.Card,{key:e.id},a.default.createElement(l.Title,null,a.default.createElement(o.Link,{to:"".concat(e.id)},e.name)),a.default.createElement(l.Subtitle,null,e.handle))}))))}},98320:function(e,t,n){var r=this&&this.__createBinding||(Object.create?function(e,t,n,r){void 0===r&&(r=n);var a=Object.getOwnPropertyDescriptor(t,n);a&&!("get"in a?!t.__esModule:a.writable||a.configurable)||(a={enumerable:!0,get:function(){return t[n]}}),Object.defineProperty(e,r,a)}:function(e,t,n,r){void 0===r&&(r=n),e[r]=t[n]}),a=this&&this.__setModuleDefault||(Object.create?function(e,t){Object.defineProperty(e,"default",{enumerable:!0,value:t})}:function(e,t){e.default=t}),o=this&&this.__importStar||function(e){if(e&&e.__esModule)return e;var t={};if(null!=e)for(var n in e)"default"!==n&&Object.prototype.hasOwnProperty.call(e,n)&&r(t,e,n);return a(t,e),t};Object.defineProperty(t,"__esModule",{value:!0}),t.Settings=void 0;var i=o(n(67294));t.Settings=function(){return(0,i.useEffect)((function(){return console.log("+++ settings"),function(){console.log("--- settings")}}),[]),i.default.createElement("div",null,"Settings")}},32523:function(e,t,n){Object.defineProperty(t,"__esModule",{value:!0}),t.useClickOutside=void 0;var r=n(67294);t.useClickOutside=function(e,t){var n=(0,r.useRef)();return(0,r.useEffect)((function(){var r=function(r){t&&n.current&&!n.current.contains(r.target)&&"function"==typeof e&&e()};return document.addEventListener("click",r,!0),function(){document.removeEventListener("click",r,!0)}}),[n,t]),n}},25818:function(e,t,n){Object.defineProperty(t,"__esModule",{value:!0}),t.useDebounce=void 0;var r=n(67294);t.useDebounce=function(e,t){void 0===t&&(t=500);var n=(0,r.useState)(e),a=n[0],o=n[1];return(0,r.useEffect)((function(){var n=setTimeout((function(){return o(e)}),t);return function(){return clearTimeout(n)}}),[e,t]),a}},12020:function(e,t,n){Object.defineProperty(t,"__esModule",{value:!0}),t.useOnKeypress=void 0;var r=n(67294);t.useOnKeypress=function(e){var t=e.meetsCondition,n=e.callback,a=e.type,o=void 0===a?"keyup":a;(0,r.useEffect)((function(){return(void 0===t||t)&&document.addEventListener(o,n),!1===t&&document.removeEventListener(o,n),function(){document.removeEventListener(o,n)}}),[t])}},92629:function(e,t,n){var r=this&&this.__importDefault||function(e){return e&&e.__esModule?e:{default:e}};Object.defineProperty(t,"__esModule",{value:!0});var a=r(n(67294)),o=r(n(20745)),i=n(88767),l=n(20938),u=n(49818);n(87662);var d=n(63438),c=n(16583),s=n(33658),f=n(89141),p=n(98320),v=n(48101),h=r(n(38957)),m=document.getElementById("freeform-client");o.default.createRoot(m).render(a.default.createElement(u.BrowserRouter,{basename:(0,v.generateUrl)("/client",!1)},a.default.createElement(i.QueryClientProvider,{client:d.queryClient},a.default.createElement(l.ReactQueryDevtools,null),a.default.createElement(c.CpNavigation,null),a.default.createElement(u.Routes,null,a.default.createElement(u.Route,{path:"/",element:a.default.createElement(h.default,null)},a.default.createElement(u.Route,{index:!0,element:a.default.createElement(s.Dashboard,null)}),a.default.createElement(u.Route,{path:"forms"},a.default.createElement(u.Route,{path:"new/*",element:a.default.createElement(f.Form,null)}),a.default.createElement(u.Route,{path:":formId/*",element:a.default.createElement(f.Form,null)}),a.default.createElement(u.Route,{index:!0,element:a.default.createElement(f.Forms,null)})),a.default.createElement(u.Route,{path:"settings",element:a.default.createElement(p.Settings,null)}))))))},42367:function(e,t,n){var r=this&&this.__importDefault||function(e){return e&&e.__esModule?e:{default:e}};Object.defineProperty(t,"__esModule",{value:!0}),t.useFieldType=t.useFetchFieldPropertySections=t.useFetchFieldTypes=void 0;var a=n(88767),o=r(n(9669));t.useFetchFieldTypes=function(e){var t=(void 0===e?{}:e).select;return(0,a.useQuery)("field-types",(function(){return o.default.get("/client/api/fields/types").then((function(e){return e.data}))}),{staleTime:1/0,select:t})};t.useFetchFieldPropertySections=function(){return(0,a.useQuery)(["field-types","property-sections"],(function(){return o.default.get("/client/api/fields/types/sections").then((function(e){return e.data})).then((function(e){return e.sort((function(e,t){return e.order-t.order}))}))}),{staleTime:1/0})};t.useFieldType=function(e){var n=(0,t.useFetchFieldTypes)().data;if(n)return n.find((function(t){return t.typeClass===e}))}},23493:function(e,t,n){var r=this&&this.__importDefault||function(e){return e&&e.__esModule?e:{default:e}};Object.defineProperty(t,"__esModule",{value:!0}),t.useQueryFormSettings=t.useQuerySingleForm=t.useQueryForms=void 0;var a=n(88767),o=n(85767),i=n(89372),l=n(45323),u=n(56859),d=n(61552),c=n(1965),s=n(68974),f=n(33314),p=r(n(9669));t.useQueryForms=function(){return(0,a.useQuery)("forms",(function(){return p.default.get("/client/api/forms").then((function(e){return e.data}))}))};t.useQuerySingleForm=function(e){var t=(0,o.useAppDispatch)();return(0,a.useQuery)(["forms",e],(function(){return p.default.get("/client/api/forms/".concat(e)).then((function(e){return e.data}))}),{staleTime:1/0,enabled:!!e,onSuccess:function(e){var n,r=e.layout,a=r.fields,o=r.pages,p=r.layouts,v=r.rows,h=r.cells;t((0,d.update)(e)),t((0,u.set)(a)),t((0,s.set)(o)),t((0,c.set)(p)),t((0,f.set)(v)),t((0,i.set)(h)),t((0,l.setPage)(null===(n=o.find(Boolean))||void 0===n?void 0:n.uid))}})};t.useQueryFormSettings=function(){return(0,a.useQuery)(["forms","settings"],(function(){return p.default.get("/client/api/forms/settings").then((function(e){return e.data}))}))}},41648:function(e,t,n){var r=this&&this.__importDefault||function(e){return e&&e.__esModule?e:{default:e}};Object.defineProperty(t,"__esModule",{value:!0}),t.useQuerySingleFormIntegration=t.useQueryFormIntegrations=t.useQueryIntegrations=void 0;var a=n(88767),o=n(47785),i=n(68887),l=r(n(9669));t.useQueryIntegrations=function(){return(0,a.useQuery)(["integrations"],(function(){return l.default.get("/client/api/integrations").then((function(e){return e.data}))}),{staleTime:1/0})};t.useQueryFormIntegrations=function(e){var t=(0,o.useDispatch)();return(0,a.useQuery)(["form-integrations"],(function(){return l.default.get("/client/api/forms/".concat(e,"/integrations")).then((function(e){return e.data}))}),{staleTime:1/0,cacheTime:1/0,onSuccess:function(e){t((0,i.addIntegrations)(e))}})};t.useQuerySingleFormIntegration=function(e,t){return(0,a.useQuery)(["form-integrations",t],(function(){return l.default.get("/client/api/forms/".concat(e,"/integrations/").concat(t)).then((function(e){return e.data}))}),{staleTime:1/0})}},240:function(e,t,n){var r=this&&this.__makeTemplateObject||function(e,t){return Object.defineProperty?Object.defineProperty(e,"raw",{value:t}):e.raw=t,e},a=this&&this.__importDefault||function(e){return e&&e.__esModule?e:{default:e}};Object.defineProperty(t,"__esModule",{value:!0}),t.colors=t.shadows=t.borderRadius=t.spacings=t.BoxShadow=void 0;var o,i=a(n(12788));t.BoxShadow=i.default.div(o||(o=r(["\n  box-shadow: 0 0 0 1px #cdd8e4, 0 2px 12px rgb(205 216 228 / 50%); ;\n"],["\n  box-shadow: 0 0 0 1px #cdd8e4, 0 2px 12px rgb(205 216 228 / 50%); ;\n"]))),t.spacings={xs:"var(--xs)",sm:"var(--s)",md:"var(--m)",lg:"var(--l)",xl:"var(--xl)"},t.borderRadius={sm:"var(--small-border-radius)",md:"var(--medium-border-radius)",lg:"var(--large-border-radius)"},t.shadows={box:"0 0 0 1px #cdd8e4, 0 2px 12px rgb(205 216 228 / 50%)",bottom:"inset 0 -1px 0 0 rgb(154 165 177 / 25%)",top:"inset 0 1px 0 0 rgb(154 165 177 / 25%)",left:"inset 1px 0 0 0 rgb(154 165 177 / 25%)",right:"inset -1px 0 0 0 rgb(154 165 177 / 25%)"},t.colors={error:"var(--error-color)",warning:"var(--warning-color)",success:"var(--success-color)",notice:"var(--notice-color)",enabled:"var(--enabled-color)",pending:"var(--pending-color)",disabled:"var(--disabled-color)",white:"var(--white)",black:"var(--black)",gray050:"var(--gray-050)",gray100:"var(--gray-100)",gray200:"var(--gray-200)",gray300:"var(--gray-300)",gray350:"var(--gray-350)",gray400:"var(--gray-400)",gray500:"var(--gray-500)",gray550:"var(--gray-550)",gray600:"var(--gray-600)",gray700:"var(--gray-700)",gray800:"var(--gray-800)",gray900:"var(--gray-900)",gray1000:"var(--gray-1000)",blue050:"var(--blue-050)",blue100:"var(--blue-100)",blue200:"var(--blue-200)",blue300:"var(--blue-300)",blue400:"var(--blue-400)",blue500:"var(--blue-500)",blue600:"var(--blue-600)",blue700:"var(--blue-700)",blue800:"var(--blue-800)",blue900:"var(--blue-900)",cyan050:"var(--cyan-050)",cyan100:"var(--cyan-100)",cyan200:"var(--cyan-200)",cyan300:"var(--cyan-300)",cyan400:"var(--cyan-400)",cyan500:"var(--cyan-500)",cyan600:"var(--cyan-600)",cyan700:"var(--cyan-700)",cyan800:"var(--cyan-800)",cyan900:"var(--cyan-900)",pink050:"var(--pink-050)",pink100:"var(--pink-100)",pink200:"var(--pink-200)",pink300:"var(--pink-300)",pink400:"var(--pink-400)",pink500:"var(--pink-500)",pink600:"var(--pink-600)",pink700:"var(--pink-700)",pink800:"var(--pink-800)",pink900:"var(--pink-900)",red050:"var(--red-050)",red100:"var(--red-100)",red200:"var(--red-200)",red300:"var(--red-300)",red400:"var(--red-400)",red500:"var(--red-500)",red600:"var(--red-600)",red700:"var(--red-700)",red800:"var(--red-800)",red900:"var(--red-900)",yellow050:"var(--yellow-050)",yellow100:"var(--yellow-100)",yellow200:"var(--yellow-200)",yellow300:"var(--yellow-300)",yellow400:"var(--yellow-400)",yellow500:"var(--yellow-500)",yellow600:"var(--yellow-600)",yellow700:"var(--yellow-700)",yellow800:"var(--yellow-800)",yellow900:"var(--yellow-900)",teal050:"var(--teal-050)",teal100:"var(--teal-100)",teal200:"var(--teal-200)",teal300:"var(--teal-300)",teal400:"var(--teal-400)",teal500:"var(--teal-500)",teal550:"var(--teal-550)",teal600:"var(--teal-600)",teal700:"var(--teal-700)",teal800:"var(--teal-800)",teal900:"var(--teal-900)"}},88454:function(e,t){var n,r=this&&this.__extends||(n=function(e,t){return n=Object.setPrototypeOf||{__proto__:[]}instanceof Array&&function(e,t){e.__proto__=t}||function(e,t){for(var n in t)Object.prototype.hasOwnProperty.call(t,n)&&(e[n]=t[n])},n(e,t)},function(e,t){if("function"!=typeof t&&null!==t)throw new TypeError("Class extends value "+String(t)+" is not a constructor or null");function r(){this.constructor=e}n(e,t),e.prototype=null===t?Object.create(t):(r.prototype=t.prototype,new r)});Object.defineProperty(t,"__esModule",{value:!0}),t.APIError=t.API_ERROR=void 0,t.API_ERROR="api_error";var a=function(e){function n(n,r){var a=e.call(this,n)||this;return a.errors={},a.name=t.API_ERROR,a.errors=r,a}return r(n,e),n}(Error);t.APIError=a},82344:function(e,t){Object.defineProperty(t,"__esModule",{value:!0}),t.range=void 0;t.range=function(e,t){void 0===t&&(e>1?(t=e,e=1):(t=e,e=0));for(var n=[],r=e;r<=t;r++)n.push(r);return n}},89361:function(e,t){Object.defineProperty(t,"__esModule",{value:!0}),t.classes=void 0;t.classes=function(){for(var e=[],t=0;t<arguments.length;t++)e[t]=arguments[t];return e.map((function(e){return"string"==typeof e&&(e=e.trim()),e})).filter((function(e){return!!e})).join(" ")},t.default=t.classes},87351:function(e,t){Object.defineProperty(t,"__esModule",{value:!0}),t.translate=t.replace=void 0;t.replace=function(e,t){void 0===t&&(t={});for(var n=0,r=Object.entries(t);n<r.length;n++){var a=r[n],o=a[0],i=a[1],l=new RegExp("\\{"+o+"\\}","g");e=e.replace(l,i.toString())}return e};t.translate=function(e,n){return void 0===n&&(n={}),"undefined"!=typeof Craft?Craft.t("freeform",e,n):(0,t.replace)(e,n)},t.default=t.translate},48101:function(e,t,n){var r=this&&this.__importDefault||function(e){return e&&e.__esModule?e:{default:e}};Object.defineProperty(t,"__esModule",{value:!0}),t.generateUrl=t.baseUrl=void 0;var a=r(n(84564));t.baseUrl=window.location.href.replace(/(.*\/freeform).*/i,"$1");t.generateUrl=function(e,n){void 0===n&&(n=!0),e=(e=(null!=e?e:"").replace(/\/+/g,"/").replace(/^\/(.*)/,"$1").replace(/\/$/,"")).length?"/".concat(e):"";var r=(0,a.default)("".concat(t.baseUrl).concat(e));return n?r.href:r.pathname}}},a={};function o(e){var t=a[e];if(void 0!==t)return t.exports;var n=a[e]={id:e,loaded:!1,exports:{}};return r[e].call(n.exports,n,n.exports,o),n.loaded=!0,n.exports}o.m=r,e=[],o.O=function(t,n,r,a){if(!n){var i=1/0;for(c=0;c<e.length;c++){n=e[c][0],r=e[c][1],a=e[c][2];for(var l=!0,u=0;u<n.length;u++)(!1&a||i>=a)&&Object.keys(o.O).every((function(e){return o.O[e](n[u])}))?n.splice(u--,1):(l=!1,a<i&&(i=a));if(l){e.splice(c--,1);var d=r();void 0!==d&&(t=d)}}return t}a=a||0;for(var c=e.length;c>0&&e[c-1][2]>a;c--)e[c]=e[c-1];e[c]=[n,r,a]},o.n=function(e){var t=e&&e.__esModule?function(){return e.default}:function(){return e};return o.d(t,{a:t}),t},n=Object.getPrototypeOf?function(e){return Object.getPrototypeOf(e)}:function(e){return e.__proto__},o.t=function(e,r){if(1&r&&(e=this(e)),8&r)return e;if("object"==typeof e&&e){if(4&r&&e.__esModule)return e;if(16&r&&"function"==typeof e.then)return e}var a=Object.create(null);o.r(a);var i={};t=t||[null,n({}),n([]),n(n)];for(var l=2&r&&e;"object"==typeof l&&!~t.indexOf(l);l=n(l))Object.getOwnPropertyNames(l).forEach((function(t){i[t]=function(){return e[t]}}));return i.default=function(){return e},o.d(a,i),a},o.d=function(e,t){for(var n in t)o.o(t,n)&&!o.o(e,n)&&Object.defineProperty(e,n,{enumerable:!0,get:t[n]})},o.g=function(){if("object"==typeof globalThis)return globalThis;try{return this||new Function("return this")()}catch(e){if("object"==typeof window)return window}}(),o.o=function(e,t){return Object.prototype.hasOwnProperty.call(e,t)},o.r=function(e){"undefined"!=typeof Symbol&&Symbol.toStringTag&&Object.defineProperty(e,Symbol.toStringTag,{value:"Module"}),Object.defineProperty(e,"__esModule",{value:!0})},o.nmd=function(e){return e.paths=[],e.children||(e.children=[]),e},function(){var e={47:0};o.O.j=function(t){return 0===e[t]};var t=function(t,n){var r,a,i=n[0],l=n[1],u=n[2],d=0;if(i.some((function(t){return 0!==e[t]}))){for(r in l)o.o(l,r)&&(o.m[r]=l[r]);if(u)var c=u(o)}for(t&&t(n);d<i.length;d++)a=i[d],o.o(e,a)&&e[a]&&e[a][0](),e[a]=0;return o.O(c)},n=self.webpackChunk_ff_client=self.webpackChunk_ff_client||[];n.forEach(t.bind(null,0)),n.push=t.bind(null,n.push.bind(n))}(),o.nc=void 0;var i=o.O(void 0,[736],(function(){return o(92629)}));i=o.O(i)}();
+!function(){"use strict";var e,t,n,r={35858:function(e,t,n){n.r(t);var r,a=n(67294);function o(){return o=Object.assign?Object.assign.bind():function(e){for(var t=1;t<arguments.length;t++){var n=arguments[t];for(var r in n)Object.prototype.hasOwnProperty.call(n,r)&&(e[r]=n[r])}return e},o.apply(this,arguments)}t.default=function(e){return a.createElement("svg",o({xmlns:"http://www.w3.org/2000/svg",viewBox:"0 0 512 512"},e),r||(r=a.createElement("path",{d:"M495.9 166.6c3.2 8.7.5 18.4-6.4 24.6l-43.3 39.4c1.1 8.3 1.7 16.8 1.7 25.4s-.6 17.1-1.7 25.4l43.3 39.4c6.9 6.2 9.6 15.9 6.4 24.6-4.4 11.9-9.7 23.3-15.8 34.3l-4.7 8.1c-6.6 11-14 21.4-22.1 31.2-5.9 7.2-15.7 9.6-24.5 6.8l-55.7-17.7c-13.4 10.3-28.2 18.9-44 25.4l-12.5 57.1c-2 9.1-9 16.3-18.2 17.8-13.8 2.3-28 3.5-42.5 3.5s-28.7-1.2-42.5-3.5c-9.2-1.5-16.2-8.7-18.2-17.8l-12.5-57.1c-15.8-6.5-30.6-15.1-44-25.4l-55.6 17.8c-8.8 2.8-18.6.3-24.5-6.8-8.1-9.8-15.5-20.2-22.1-31.2l-4.7-8.1c-6.1-11-11.4-22.4-15.8-34.3-3.2-8.7-.5-18.4 6.4-24.6l43.3-39.4c-1.1-8.4-1.7-16.9-1.7-25.5s.6-17.1 1.7-25.4l-43.3-39.4c-6.9-6.2-9.6-15.9-6.4-24.6 4.4-11.9 9.7-23.3 15.8-34.3l4.7-8.1c6.6-11 14-21.4 22.1-31.2 5.9-7.2 15.7-9.6 24.5-6.8l55.7 17.7c13.4-10.3 28.2-18.9 44-25.4l12.5-57.1c2-9.1 9-16.3 18.2-17.8C227.3 1.2 241.5 0 256 0s28.7 1.2 42.5 3.5c9.2 1.5 16.2 8.7 18.2 17.8l12.5 57.1c15.8 6.5 30.6 15.1 44 25.4l55.7-17.7c8.8-2.8 18.6-.3 24.5 6.8 8.1 9.8 15.5 20.2 22.1 31.2l4.7 8.1c6.1 11 11.4 22.4 15.8 34.3zM256 336c44.2 0 80-35.8 80-80s-35.8-80-80-80-80 35.8-80 80 35.8 80 80 80z"})))}},65721:function(e,t,n){n.r(t);var r,a=n(67294);function o(){return o=Object.assign?Object.assign.bind():function(e){for(var t=1;t<arguments.length;t++){var n=arguments[t];for(var r in n)Object.prototype.hasOwnProperty.call(n,r)&&(e[r]=n[r])}return e},o.apply(this,arguments)}t.default=function(e){return a.createElement("svg",o({xmlns:"http://www.w3.org/2000/svg",viewBox:"0 0 512 512"},e),r||(r=a.createElement("path",{fill:"currentColor",d:"M416 208c0 45.9-14.9 88.3-40 122.7l126.6 126.7c12.5 12.5 12.5 32.8 0 45.3s-32.8 12.5-45.3 0L330.7 376c-34.4 25.2-76.8 40-122.7 40C93.1 416 0 322.9 0 208S93.1 0 208 0s208 93.1 208 208zM208 352c79.5 0 144-64.5 144-144S287.5 64 208 64 64 128.5 64 208s64.5 144 144 144z"})))}},56137:function(e,t,n){n.r(t);var r,a=n(67294);function o(){return o=Object.assign?Object.assign.bind():function(e){for(var t=1;t<arguments.length;t++){var n=arguments[t];for(var r in n)Object.prototype.hasOwnProperty.call(n,r)&&(e[r]=n[r])}return e},o.apply(this,arguments)}t.default=function(e){return a.createElement("svg",o({xmlns:"http://www.w3.org/2000/svg",viewBox:"0 0 512 512"},e),r||(r=a.createElement("path",{fill:"currentColor",d:"M0 416c0-17.7 14.3-32 32-32h54.7c12.3-28.3 40.5-48 73.3-48s61 19.7 73.3 48H480c17.7 0 32 14.3 32 32s-14.3 32-32 32H233.3c-12.3 28.3-40.5 48-73.3 48s-61-19.7-73.3-48H32c-17.7 0-32-14.3-32-32zm192 0c0-17.7-14.3-32-32-32s-32 14.3-32 32 14.3 32 32 32 32-14.3 32-32zm192-160c0-17.7-14.3-32-32-32s-32 14.3-32 32 14.3 32 32 32 32-14.3 32-32zm-32-80c32.8 0 61 19.7 73.3 48H480c17.7 0 32 14.3 32 32s-14.3 32-32 32h-54.7c-12.3 28.3-40.5 48-73.3 48s-61-19.7-73.3-48H32c-17.7 0-32-14.3-32-32s14.3-32 32-32h246.7c12.3-28.3 40.5-48 73.3-48zM192 64c-17.7 0-32 14.3-32 32s14.3 32 32 32 32-14.3 32-32-14.3-32-32-32zm73.3 0H480c17.7 0 32 14.3 32 32s-14.3 32-32 32H265.3c-12.3 28.3-40.5 48-73.3 48s-61-19.7-73.3-48H32c-17.7 0-32-14.3-32-32s14.3-32 32-32h86.7C131 35.7 159.2 16 192 16s61 19.7 73.3 48z"})))}},27796:function(e,t,n){n.r(t);var r,a=n(67294);function o(){return o=Object.assign?Object.assign.bind():function(e){for(var t=1;t<arguments.length;t++){var n=arguments[t];for(var r in n)Object.prototype.hasOwnProperty.call(n,r)&&(e[r]=n[r])}return e},o.apply(this,arguments)}t.default=function(e){return a.createElement("svg",o({xmlns:"http://www.w3.org/2000/svg",viewBox:"0 0 384 512"},e),r||(r=a.createElement("path",{fill:"currentColor",d:"M41.4 233.4c-12.5 12.5-12.5 32.8 0 45.3l192 192c12.5 12.5 32.8 12.5 45.3 0s12.5-32.8 0-45.3L109.3 256 278.6 86.6c12.5-12.5 12.5-32.8 0-45.3s-32.8-12.5-45.3 0l-192 192z"})))}},33742:function(e,t,n){n.r(t);var r,a=n(67294);function o(){return o=Object.assign?Object.assign.bind():function(e){for(var t=1;t<arguments.length;t++){var n=arguments[t];for(var r in n)Object.prototype.hasOwnProperty.call(n,r)&&(e[r]=n[r])}return e},o.apply(this,arguments)}t.default=function(e){return a.createElement("svg",o({xmlns:"http://www.w3.org/2000/svg",viewBox:"0 0 512 512"},e),r||(r=a.createElement("path",{d:"M256 512c141.4 0 256-114.6 256-256S397.4 0 256 0 0 114.6 0 256s114.6 256 256 256zm-81-337c9.4-9.4 24.6-9.4 33.9 0l47 47 47-47c9.4-9.4 24.6-9.4 33.9 0s9.4 24.6 0 33.9l-47 47 47 47c9.4 9.4 9.4 24.6 0 33.9s-24.6 9.4-33.9 0l-47-47-47 47c-9.4 9.4-24.6 9.4-33.9 0s-9.4-24.6 0-33.9l47-47-47-47c-9.4-9.4-9.4-24.6 0-33.9z"})))}},80755:function(e,t,n){var r=this&&this.__importDefault||function(e){return e&&e.__esModule?e:{default:e}};Object.defineProperty(t,"__esModule",{value:!0});var a=n(88454),o=n(48101),i=r(n(9669));i.default.defaults.baseURL=(0,o.generateUrl)("/"),i.default.defaults.headers.get&&(i.default.defaults.headers.get.Accept="application/json"),i.default.defaults.headers.post&&(i.default.defaults.headers.post.Accept="application/json"),i.default.interceptors.request.use((function(e){return["post","put","patch","delete"].includes(e.method)&&e.data&&void 0!==n.g.Craft&&(e.data[Craft.csrfTokenName]=Craft.csrfTokenValue),e})),i.default.interceptors.response.use(null,(function(e){return e.response.data.error&&(e.message=e.response.data.error),e.response.data.errors?Promise.reject(new a.APIError(e.message,e.response.data.errors)):Promise.reject(e)}))},87662:function(e,t,n){Object.defineProperty(t,"__esModule",{value:!0}),n(80755)},63438:function(e,t,n){Object.defineProperty(t,"__esModule",{value:!0}),t.queryClient=void 0;var r=n(88767);t.queryClient=new r.QueryClient({defaultOptions:{queries:{cacheTime:6e5,retry:!1,refetchOnWindowFocus:!1}}})},72094:function(e,t,n){var r=this&&this.__makeTemplateObject||function(e,t){return Object.defineProperty?Object.defineProperty(e,"raw",{value:t}):e.raw=t,e},a=this&&this.__importDefault||function(e){return e&&e.__esModule?e:{default:e}};Object.defineProperty(t,"__esModule",{value:!0}),t.AppWrapper=void 0;var o,i=a(n(12788));t.AppWrapper=i.default.div(o||(o=r([""],[""])))},38957:function(e,t,n){var r=this&&this.__importDefault||function(e){return e&&e.__esModule?e:{default:e}};Object.defineProperty(t,"__esModule",{value:!0});var a=r(n(67294)),o=n(49818),i=n(72094);t.default=function(){return a.default.createElement(i.AppWrapper,{id:"freeform-client-app"},a.default.createElement(o.Outlet,null))}},16583:function(e,t,n){var r=this&&this.__createBinding||(Object.create?function(e,t,n,r){void 0===r&&(r=n);var a=Object.getOwnPropertyDescriptor(t,n);a&&!("get"in a?!t.__esModule:a.writable||a.configurable)||(a={enumerable:!0,get:function(){return t[n]}}),Object.defineProperty(e,r,a)}:function(e,t,n,r){void 0===r&&(r=n),e[r]=t[n]}),a=this&&this.__setModuleDefault||(Object.create?function(e,t){Object.defineProperty(e,"default",{enumerable:!0,value:t})}:function(e,t){e.default=t}),o=this&&this.__importStar||function(e){if(e&&e.__esModule)return e;var t={};if(null!=e)for(var n in e)"default"!==n&&Object.prototype.hasOwnProperty.call(e,n)&&r(t,e,n);return a(t,e),t},i=this&&this.__importDefault||function(e){return e&&e.__esModule?e:{default:e}};Object.defineProperty(t,"__esModule",{value:!0}),t.CpNavigation=void 0;var l=o(n(67294)),u=n(73935),d=n(49818),c=i(n(87351)),s=function(e){return e.isActive?"sel":void 0};t.CpNavigation=function(){return(0,l.useEffect)((function(){var e;null===(e=document.querySelector("#nav-freeform > a.sel"))||void 0===e||e.classList.remove("sel")}),[]),(0,u.createPortal)(l.default.createElement("ul",{className:"subnav react"},l.default.createElement("li",null,l.default.createElement(d.NavLink,{to:"",className:s},(0,c.default)("Dashboard"))),l.default.createElement("li",null,l.default.createElement(d.NavLink,{to:"forms",className:s},(0,c.default)("Forms"))),l.default.createElement("li",null,l.default.createElement(d.NavLink,{to:"settings",className:s},(0,c.default)("Settings")))),document.getElementById("nav-freeform"))}},47170:function(e,t,n){var r,a=this&&this.__extends||(r=function(e,t){return r=Object.setPrototypeOf||{__proto__:[]}instanceof Array&&function(e,t){e.__proto__=t}||function(e,t){for(var n in t)Object.prototype.hasOwnProperty.call(t,n)&&(e[n]=t[n])},r(e,t)},function(e,t){if("function"!=typeof t&&null!==t)throw new TypeError("Class extends value "+String(t)+" is not a constructor or null");function n(){this.constructor=e}r(e,t),e.prototype=null===t?Object.create(t):(n.prototype=t.prototype,new n)}),o=this&&this.__importDefault||function(e){return e&&e.__esModule?e:{default:e}};Object.defineProperty(t,"__esModule",{value:!0}),t.ErrorBoundary=void 0;var i=o(n(67294)),l=function(e){function t(t){var n=e.call(this,t)||this;return n.state={hasError:!1},n}return a(t,e),t.getDerivedStateFromError=function(){return{hasError:!0}},t.prototype.componentDidCatch=function(e,t){console.log(e,t)},t.prototype.render=function(){return this.state.hasError?i.default.createElement("div",null,this.props.message):this.props.children},t}(i.default.Component);t.ErrorBoundary=l},92617:function(e,t,n){var r=this&&this.__makeTemplateObject||function(e,t){return Object.defineProperty?Object.defineProperty(e,"raw",{value:t}):e.raw=t,e},a=this&&this.__importDefault||function(e){return e&&e.__esModule?e:{default:e}};Object.defineProperty(t,"__esModule",{value:!0}),t.FieldInstructions=t.FieldHeading=t.Field=void 0;var o,i,l,u=a(n(12788));t.Field=u.default.div(o||(o=r(["\n  margin: 0;\n  padding: 0;\n  width: 100%;\n  display: block;\n"],["\n  margin: 0;\n  padding: 0;\n  width: 100%;\n  display: block;\n"]))),t.FieldHeading=u.default.div(i||(i=r(["\n  margin: 0;\n  font-weight: bold;\n  padding: 0 0 10px 0;\n"],["\n  margin: 0;\n  font-weight: bold;\n  padding: 0 0 10px 0;\n"]))),t.FieldInstructions=u.default.div(l||(l=r(["\n  margin: 0;\n  font-size: 90%;\n  font-style: italic;\n  padding: 0 0 10px 0;\n"],["\n  margin: 0;\n  font-size: 90%;\n  font-style: italic;\n  padding: 0 0 10px 0;\n"])))},71967:function(e,t,n){var r=this&&this.__importDefault||function(e){return e&&e.__esModule?e:{default:e}};Object.defineProperty(t,"__esModule",{value:!0}),t.BaseControl=void 0;var a=r(n(67294)),o=n(92617);t.BaseControl=function(e){var t=e.property,n=e.children,r=t.label,i=t.handle,l=t.instructions;return a.default.createElement(o.Field,{className:"field"},!!r&&a.default.createElement(o.FieldHeading,null,a.default.createElement("label",{htmlFor:i},r)),!!l&&a.default.createElement(o.FieldInstructions,null,l),a.default.createElement("div",{className:"input"},n))}},42922:function(e,t,n){var r=this&&this.__importDefault||function(e){return e&&e.__esModule?e:{default:e}};Object.defineProperty(t,"__esModule",{value:!0});var a=r(n(67294)),o=n(32969),i=n(30440),l=n(71967);t.default=function(e){var t=e.value,n=e.property,r=e.onUpdateValue;return a.default.createElement(l.BaseControl,{property:n},a.default.createElement(i.InputWrapper,null,a.default.createElement(o.LightSwitchInput,{id:n.handle,enabled:t,onClick:function(){return r(!t)}})))}},81218:function(e,t,n){var r=this&&this.__makeTemplateObject||function(e,t){return Object.defineProperty?Object.defineProperty(e,"raw",{value:t}):e.raw=t,e},a=this&&this.__importDefault||function(e){return e&&e.__esModule?e:{default:e}};Object.defineProperty(t,"__esModule",{value:!0}),t.Overlay=t.Popover=t.SelectedColor=t.Swatch=void 0;var o,i,l,u,d=a(n(12788));t.Swatch=d.default.div(o||(o=r(["\n  margin: 0;\n  padding: 5px;\n  cursor: pointer;\n  border-radius: 1px;\n  display: inline-block;\n  background-color: #ffffff;\n  box-shadow: 0 0 0 1px rgba(0, 0, 0, 0.1);,\n"],["\n  margin: 0;\n  padding: 5px;\n  cursor: pointer;\n  border-radius: 1px;\n  display: inline-block;\n  background-color: #ffffff;\n  box-shadow: 0 0 0 1px rgba(0, 0, 0, 0.1);,\n"]))),t.SelectedColor=d.default.div(i||(i=r(["\n  margin: 0;\n  padding: 0;\n  width: 36px;\n  height: 14px;\n  border-radius: 2px;\n"],["\n  margin: 0;\n  padding: 0;\n  width: 36px;\n  height: 14px;\n  border-radius: 2px;\n"]))),t.Popover=d.default.div(l||(l=r(["\n  z-index: 2;\n  position: absolute;\n"],["\n  z-index: 2;\n  position: absolute;\n"]))),t.Overlay=d.default.div(u||(u=r(["\n  top: 0;\n  left: 0;\n  right: 0;\n  bottom: 0;\n  position: fixed;\n"],["\n  top: 0;\n  left: 0;\n  right: 0;\n  bottom: 0;\n  position: fixed;\n"])))},25857:function(e,t,n){var r=this&&this.__createBinding||(Object.create?function(e,t,n,r){void 0===r&&(r=n);var a=Object.getOwnPropertyDescriptor(t,n);a&&!("get"in a?!t.__esModule:a.writable||a.configurable)||(a={enumerable:!0,get:function(){return t[n]}}),Object.defineProperty(e,r,a)}:function(e,t,n,r){void 0===r&&(r=n),e[r]=t[n]}),a=this&&this.__setModuleDefault||(Object.create?function(e,t){Object.defineProperty(e,"default",{enumerable:!0,value:t})}:function(e,t){e.default=t}),o=this&&this.__importStar||function(e){if(e&&e.__esModule)return e;var t={};if(null!=e)for(var n in e)"default"!==n&&Object.prototype.hasOwnProperty.call(e,n)&&r(t,e,n);return a(t,e),t};Object.defineProperty(t,"__esModule",{value:!0});var i=o(n(67294)),l=n(93565),u=n(81218),d=n(71967);t.default=function(e){var t=e.value,n=e.property,r=e.onUpdateValue,a=(0,i.useState)(!1),o=a[0],c=a[1];return i.default.createElement(d.BaseControl,{property:n},i.default.createElement(u.Swatch,{onClick:function(){return c(!o)}},i.default.createElement(u.SelectedColor,{style:{backgroundColor:t}})),o&&i.default.createElement(u.Popover,null,i.default.createElement(u.Overlay,{onClick:function(){return c(!1)}}),i.default.createElement(l.SketchPicker,{color:t,onChangeComplete:function(e){var t=e.hex;return r(t)}})))}},42250:function(e,t,n){var r=this&&this.__importDefault||function(e){return e&&e.__esModule?e:{default:e}};Object.defineProperty(t,"__esModule",{value:!0});var a=r(n(67294)),o=r(n(9198)),i=r(n(89361)),l=r(n(49546)),u=r(n(14372)),d=n(71967);n(59339);var c="MM/dd/yyyy";t.default=function(e){var t=e.value,n=e.property,r=e.onUpdateValue,s=new Date,f=t||(0,l.default)(s,c),p=(0,u.default)(f,c,s);return a.default.createElement(d.BaseControl,{property:n},a.default.createElement(o.default,{id:n.handle,minDate:s,selected:p,dateFormat:c,className:(0,i.default)("text","fullwidth"),onChange:function(e){return r((0,l.default)(e,c))}}))}},458:function(e,t,n){var r=this&&this.__importDefault||function(e){return e&&e.__esModule?e:{default:e}};Object.defineProperty(t,"__esModule",{value:!0}),t.textarea=t.string=t.select=t.int=t.datePicker=t.colorPicker=t.bool=void 0;var a=n(42922);Object.defineProperty(t,"bool",{enumerable:!0,get:function(){return r(a).default}});var o=n(25857);Object.defineProperty(t,"colorPicker",{enumerable:!0,get:function(){return r(o).default}});var i=n(42250);Object.defineProperty(t,"datePicker",{enumerable:!0,get:function(){return r(i).default}});var l=n(37280);Object.defineProperty(t,"int",{enumerable:!0,get:function(){return r(l).default}});var u=n(15509);Object.defineProperty(t,"select",{enumerable:!0,get:function(){return r(u).default}});var d=n(21969);Object.defineProperty(t,"string",{enumerable:!0,get:function(){return r(d).default}});var c=n(62553);Object.defineProperty(t,"textarea",{enumerable:!0,get:function(){return r(c).default}})},37280:function(e,t,n){var r=this&&this.__importDefault||function(e){return e&&e.__esModule?e:{default:e}};Object.defineProperty(t,"__esModule",{value:!0});var a=r(n(67294)),o=r(n(89361)),i=n(71967);t.default=function(e){var t=e.value,n=e.property,r=e.onUpdateValue,l=n.handle,u=n.placeholder;return a.default.createElement(i.BaseControl,{property:n},a.default.createElement("input",{id:l,type:"number",placeholder:u,className:(0,o.default)("text","fullwidth"),value:t.toString(),onChange:function(e){return r(Number(e.target.value))}}))}},15509:function(e,t,n){var r=this&&this.__importDefault||function(e){return e&&e.__esModule?e:{default:e}};Object.defineProperty(t,"__esModule",{value:!0});var a=r(n(67294)),o=n(71967);t.default=function(e){var t=e.value,n=e.property,r=e.onUpdateValue,i=n.options;return a.default.createElement(o.BaseControl,{property:n},a.default.createElement("select",{id:n.handle,value:t||"",className:"text fullwidth",onChange:function(e){return r(e.target.value)}},i.map((function(e,t){var n=e.value,r=e.label;return a.default.createElement("option",{key:t,value:n,label:r})}))))}},21969:function(e,t,n){var r=this&&this.__importDefault||function(e){return e&&e.__esModule?e:{default:e}};Object.defineProperty(t,"__esModule",{value:!0});var a=r(n(67294)),o=r(n(89361)),i=n(71967);t.default=function(e){var t=e.value,n=e.property,r=e.onUpdateValue,l=n.handle,u=n.placeholder;return a.default.createElement(i.BaseControl,{property:n},a.default.createElement("input",{id:l,type:"text",placeholder:u,className:(0,o.default)("text","fullwidth",n.flags.includes("code")&&"code"),value:t||"",onChange:function(e){return r(e.target.value)}}))}},62553:function(e,t,n){var r=this&&this.__importDefault||function(e){return e&&e.__esModule?e:{default:e}};Object.defineProperty(t,"__esModule",{value:!0});var a=r(n(67294)),o=r(n(89361)),i=n(71967);t.default=function(e){var t=e.value,n=e.property,r=e.onUpdateValue,l=n.handle,u=n.placeholder,d=n.rows;return a.default.createElement(i.BaseControl,{property:n},a.default.createElement("textarea",{id:l,rows:d,placeholder:u,className:(0,o.default)("text","fullwidth",n.flags.includes("code")&&"code"),value:t||"",onChange:function(e){return r(e.target.value)}}))}},82384:function(e,t,n){var r=this&&this.__createBinding||(Object.create?function(e,t,n,r){void 0===r&&(r=n);var a=Object.getOwnPropertyDescriptor(t,n);a&&!("get"in a?!t.__esModule:a.writable||a.configurable)||(a={enumerable:!0,get:function(){return t[n]}}),Object.defineProperty(e,r,a)}:function(e,t,n,r){void 0===r&&(r=n),e[r]=t[n]}),a=this&&this.__setModuleDefault||(Object.create?function(e,t){Object.defineProperty(e,"default",{enumerable:!0,value:t})}:function(e,t){e.default=t}),o=this&&this.__importStar||function(e){if(e&&e.__esModule)return e;var t={};if(null!=e)for(var n in e)"default"!==n&&Object.prototype.hasOwnProperty.call(e,n)&&r(t,e,n);return a(t,e),t},i=this&&this.__importDefault||function(e){return e&&e.__esModule?e:{default:e}};Object.defineProperty(t,"__esModule",{value:!0}),t.FormControlGenerator=void 0;var l=o(n(67294)),u=n(47785),d=n(61552),c=i(n(76884)),s=n(47170),f=o(n(458));t.FormControlGenerator=function(e){var t=e.onValueUpdate,n=e.namespace,r=e.property,a=(0,u.useSelector)((0,d.selectFormSetting)(n,r.handle)),o=(0,c.default)(r.type),i=f[o];return void 0===i?l.default.createElement("div",null,"...".concat(r.handle," <").concat(r.type," [").concat(o,"]>")):(i.displayName="Setting <".concat(r.type,">"),l.default.createElement(s.ErrorBoundary,{message:"...".concat(r.handle," <").concat(r.type,">")},l.default.createElement(l.Suspense,null,l.default.createElement(i,{value:a,property:r,onUpdateValue:t}))))}},3399:function(e,t,n){var r=this&&this.__importDefault||function(e){return e&&e.__esModule?e:{default:e}};Object.defineProperty(t,"__esModule",{value:!0}),t.CheckboxInput=void 0;var a=r(n(67294));t.CheckboxInput=function(e){var t=e.id,n=e.label,r=e.onClick;return a.default.createElement("div",{className:"checkbox-wrapper"},a.default.createElement("input",{id:t,type:"checkbox",className:"checkbox",onClick:r}),a.default.createElement("label",{htmlFor:t},n))}},30440:function(e,t,n){var r=this&&this.__makeTemplateObject||function(e,t){return Object.defineProperty?Object.defineProperty(e,"raw",{value:t}):e.raw=t,e},a=this&&this.__createBinding||(Object.create?function(e,t,n,r){void 0===r&&(r=n);var a=Object.getOwnPropertyDescriptor(t,n);a&&!("get"in a?!t.__esModule:a.writable||a.configurable)||(a={enumerable:!0,get:function(){return t[n]}}),Object.defineProperty(e,r,a)}:function(e,t,n,r){void 0===r&&(r=n),e[r]=t[n]}),o=this&&this.__setModuleDefault||(Object.create?function(e,t){Object.defineProperty(e,"default",{enumerable:!0,value:t})}:function(e,t){e.default=t}),i=this&&this.__importStar||function(e){if(e&&e.__esModule)return e;var t={};if(null!=e)for(var n in e)"default"!==n&&Object.prototype.hasOwnProperty.call(e,n)&&a(t,e,n);return o(t,e),t};Object.defineProperty(t,"__esModule",{value:!0}),t.Wrapper=t.Handle=t.InputWrapper=void 0;var l=n(240),u=i(n(12788));t.InputWrapper=u.default.div(d||(d=r(["\n  padding: 6px 0;\n"],["\n  padding: 6px 0;\n"]))),t.Handle=u.default.div(c||(c=r(["\n  position: absolute;\n  top: 1px;\n  left: 1px;\n\n  width: 20px;\n  height: 20px;\n\n  border-radius: 10px;\n  background-color: ",";\n\n  transition: left 0.1s ease-out;\n"],["\n  position: absolute;\n  top: 1px;\n  left: 1px;\n\n  width: 20px;\n  height: 20px;\n\n  border-radius: 10px;\n  background-color: ",";\n\n  transition: left 0.1s ease-out;\n"])),l.colors.white);var d,c,s,f,p=(0,u.css)(s||(s=r(["\n  background-image: linear-gradient(\n    to right,\n    ",",\n    ","\n  );\n\n  > "," {\n    left: calc(100% - 21px);\n  }\n"],["\n  background-image: linear-gradient(\n    to right,\n    ",",\n    ","\n  );\n\n  > "," {\n    left: calc(100% - 21px);\n  }\n"])),l.colors.teal550,l.colors.teal550,t.Handle);t.Wrapper=u.default.button(f||(f=r(["\n  cursor: pointer;\n  position: relative;\n\n  display: block;\n  overflow: hidden;\n\n  width: 34px;\n  height: 22px;\n\n  border: none;\n  border-radius: 11px;\n\n  user-select: none;\n  transition: background-image 0.1s linear;\n  background-image: linear-gradient(\n    to right,\n    ",",\n    ","\n  );\n\n  ","\n"],["\n  cursor: pointer;\n  position: relative;\n\n  display: block;\n  overflow: hidden;\n\n  width: 34px;\n  height: 22px;\n\n  border: none;\n  border-radius: 11px;\n\n  user-select: none;\n  transition: background-image 0.1s linear;\n  background-image: linear-gradient(\n    to right,\n    ",",\n    ","\n  );\n\n  ","\n"])),l.colors.gray400,l.colors.gray400,(function(e){return e.enabled&&p}))},32969:function(e,t,n){var r=this&&this.__importDefault||function(e){return e&&e.__esModule?e:{default:e}};Object.defineProperty(t,"__esModule",{value:!0}),t.LightSwitchInput=void 0;var a=r(n(67294)),o=n(30440);t.LightSwitchInput=function(e){var t=e.enabled,n=e.onClick;return a.default.createElement(o.Wrapper,{enabled:t,onClick:function(){return n&&n(!t)}},a.default.createElement(o.Handle,null))}},1847:function(e,t,n){var r=this&&this.__makeTemplateObject||function(e,t){return Object.defineProperty?Object.defineProperty(e,"raw",{value:t}):e.raw=t,e},a=this&&this.__importDefault||function(e){return e&&e.__esModule?e:{default:e}};Object.defineProperty(t,"__esModule",{value:!0}),t.Space=void 0;var o,i=n(240),l=a(n(12788));t.Space=l.default.div(o||(o=r(["\n  width: 100%;\n  height: 0;\n  margin-bottom: ",";\n"],["\n  width: 100%;\n  height: 0;\n  margin-bottom: ",";\n"])),i.spacings.xl)},4524:function(e,t,n){var r=this&&this.__makeTemplateObject||function(e,t){return Object.defineProperty?Object.defineProperty(e,"raw",{value:t}):e.raw=t,e},a=this&&this.__createBinding||(Object.create?function(e,t,n,r){void 0===r&&(r=n);var a=Object.getOwnPropertyDescriptor(t,n);a&&!("get"in a?!t.__esModule:a.writable||a.configurable)||(a={enumerable:!0,get:function(){return t[n]}}),Object.defineProperty(e,r,a)}:function(e,t,n,r){void 0===r&&(r=n),e[r]=t[n]}),o=this&&this.__setModuleDefault||(Object.create?function(e,t){Object.defineProperty(e,"default",{enumerable:!0,value:t})}:function(e,t){e.default=t}),i=this&&this.__importStar||function(e){if(e&&e.__esModule)return e;var t={};if(null!=e)for(var n in e)"default"!==n&&Object.prototype.hasOwnProperty.call(e,n)&&a(t,e,n);return o(t,e),t},l=this&&this.__importDefault||function(e){return e&&e.__esModule?e:{default:e}};Object.defineProperty(t,"__esModule",{value:!0}),t.SidebarSlider=void 0;var u,d,c=i(n(67294)),s=n(39920),f=n(240),p=l(n(12788)),v=n(66264),h=p.default.div(u||(u=r(["\n  position: absolute !important;\n  left: 0;\n  right: 0;\n  top: 0;\n  bottom: 0;\n"],["\n  position: absolute !important;\n  left: 0;\n  right: 0;\n  top: 0;\n  bottom: 0;\n"]))),m=(0,p.default)(s.animated.div)(d||(d=r(["\n  position: absolute;\n  top: 0;\n  left: 0;\n  right: 0;\n  bottom: 0;\n\n  box-shadow: ",";\n"],["\n  position: absolute;\n  top: 0;\n  left: 0;\n  right: 0;\n  bottom: 0;\n\n  box-shadow: ",";\n"])),f.shadows.left);t.SidebarSlider=function(e){var t=e.swiped,n=e.children,r=(0,c.useRef)(null),a=(0,s.useSprings)(n.length,n.map((function(e,n){return{x:n?t?-300:0:t?0:300,config:{mass:1,tension:open?200:350,friction:26}}})));return c.default.createElement(v.Sidebar,{style:{overflow:"hidden"},ref:r},c.default.createElement(h,null,n.map((function(e,t){return c.default.createElement(m,{style:a[t],key:t},e)}))))}},66264:function(e,t,n){var r=this&&this.__makeTemplateObject||function(e,t){return Object.defineProperty?Object.defineProperty(e,"raw",{value:t}):e.raw=t,e},a=this&&this.__importDefault||function(e){return e&&e.__esModule?e:{default:e}};Object.defineProperty(t,"__esModule",{value:!0}),t.Sidebar=void 0;var o,i=n(240),l=a(n(12788));t.Sidebar=l.default.div(o||(o=r(["\n  position: relative;\n\n  flex-basis: 300px;\n  width: 300px;\n  padding: ",";\n  box-sizing: border-box;\n\n  border-bottom-left-radius: ",";\n  box-shadow: inset -1px 0 0 0 rgb(154 165 177 / 25%);\n  background: ",";\n"],["\n  position: relative;\n\n  flex-basis: 300px;\n  width: 300px;\n  padding: ",";\n  box-sizing: border-box;\n\n  border-bottom-left-radius: ",";\n  box-shadow: inset -1px 0 0 0 rgb(154 165 177 / 25%);\n  background: ",";\n"])),(function(e){return e.lean?i.spacings.sm:i.spacings.lg}),i.borderRadius.lg,i.colors.gray050)},59518:function(e,t,n){var r=this&&this.__importDefault||function(e){return e&&e.__esModule?e:{default:e}};Object.defineProperty(t,"__esModule",{value:!0});var a=r(n(76884));t.default=function(e,t){return(void 0===t?[!1]:t)[0]&&(e=(0,a.default)(e)),e.replace(/[^a-zA-Z0-9\-_]/g,"")}},99834:function(e,t,n){var r=this&&this.__importDefault||function(e){return e&&e.__esModule?e:{default:e}};Object.defineProperty(t,"__esModule",{value:!0}),t.handle=void 0;var a=n(59518);Object.defineProperty(t,"handle",{enumerable:!0,get:function(){return r(a).default}})},16430:function(e,t,n){var r=this&&this.__createBinding||(Object.create?function(e,t,n,r){void 0===r&&(r=n);var a=Object.getOwnPropertyDescriptor(t,n);a&&!("get"in a?!t.__esModule:a.writable||a.configurable)||(a={enumerable:!0,get:function(){return t[n]}}),Object.defineProperty(e,r,a)}:function(e,t,n,r){void 0===r&&(r=n),e[r]=t[n]}),a=this&&this.__setModuleDefault||(Object.create?function(e,t){Object.defineProperty(e,"default",{enumerable:!0,value:t})}:function(e,t){e.default=t}),o=this&&this.__importStar||function(e){if(e&&e.__esModule)return e;var t={};if(null!=e)for(var n in e)"default"!==n&&Object.prototype.hasOwnProperty.call(e,n)&&r(t,e,n);return a(t,e),t};Object.defineProperty(t,"__esModule",{value:!0}),t.applyMiddleware=void 0;var i=o(n(99834));t.applyMiddleware=function(e,t,n){var r=e;return t.forEach((function(t){var a=t[0],o=t[1];i[a]&&(r=i[a](e,o,n))})),r}},54981:function(e,t,n){var r=this&&this.__makeTemplateObject||function(e,t){return Object.defineProperty?Object.defineProperty(e,"raw",{value:t}):e.raw=t,e},a=this&&this.__importDefault||function(e){return e&&e.__esModule?e:{default:e}};Object.defineProperty(t,"__esModule",{value:!0}),t.Wrapper=void 0;var o,i=a(n(12788));t.Wrapper=i.default.div(o||(o=r(["\n  color: white;\n  background: red;\n  border: 1px solid darkred;\n"],["\n  color: white;\n  background: red;\n  border: 1px solid darkred;\n"])))},4696:function(e,t,n){var r=this&&this.__importDefault||function(e){return e&&e.__esModule?e:{default:e}};Object.defineProperty(t,"__esModule",{value:!0}),t.ErrorBlock=void 0;var a=r(n(67294)),o=n(54981);t.ErrorBlock=function(e){var t=e.children;return a.default.createElement(o.Wrapper,null,t)}},80306:function(e,t,n){var r=this&&this.__importDefault||function(e){return e&&e.__esModule?e:{default:e}};Object.defineProperty(t,"__esModule",{value:!0}),t.ThemedSkeleton=void 0;var a=r(n(67294)),o=n(84849);t.ThemedSkeleton=function(e){var t=e.children;return a.default.createElement(o.SkeletonTheme,{baseColor:"#e6eaee",highlightColor:"#ced1d4"},t)}},33658:function(e,t,n){var r=this&&this.__importDefault||function(e){return e&&e.__esModule?e:{default:e}};Object.defineProperty(t,"__esModule",{value:!0}),t.Dashboard=void 0;var a=r(n(67294));t.Dashboard=function(){return a.default.createElement("div",null,"Dashboard")}},73803:function(e,t,n){var r=this&&this.__makeTemplateObject||function(e,t){return Object.defineProperty?Object.defineProperty(e,"raw",{value:t}):e.raw=t,e},a=this&&this.__importDefault||function(e){return e&&e.__esModule?e:{default:e}};Object.defineProperty(t,"__esModule",{value:!0}),t.BuilderContent=t.BuilderWrapper=void 0;var o,i,l=n(240),u=a(n(12788));t.BuilderWrapper=u.default.div(o||(o=r(["\n  display: flex;\n  flex-direction: column;\n  height: 100%;\n\n  box-shadow: 0 0 0 1px ",", 0 2px 12px rgb(205 216 228 / 50%);\n  border-radius: ",";\n"],["\n  display: flex;\n  flex-direction: column;\n  height: 100%;\n\n  box-shadow: 0 0 0 1px ",", 0 2px 12px rgb(205 216 228 / 50%);\n  border-radius: ",";\n"])),l.colors.gray200,l.borderRadius.lg),t.BuilderContent=u.default.div(i||(i=r(["\n  flex-grow: 1;\n"],["\n  flex-grow: 1;\n"])))},93787:function(e,t,n){var r=this&&this.__importDefault||function(e){return e&&e.__esModule?e:{default:e}};Object.defineProperty(t,"__esModule",{value:!0}),t.Builder=void 0;var a=r(n(67294)),o=n(84520),i=n(27318),l=n(49818),u=n(39130),d=n(678),c=n(26678),s=n(70793),f=n(86419),p=n(73803),v=n(19179);t.Builder=function(){return a.default.createElement(o.DndProvider,{backend:i.HTML5Backend},a.default.createElement(p.BuilderWrapper,null,a.default.createElement(v.Tabs,null),a.default.createElement(p.BuilderContent,null,a.default.createElement(l.Routes,null,a.default.createElement(l.Route,{index:!0,element:a.default.createElement(f.LayoutEditor,null)}),a.default.createElement(l.Route,{path:"notifications",element:a.default.createElement("div",null,"Notifications")}),a.default.createElement(l.Route,{path:"integrations",element:a.default.createElement(s.Integrations,null)},a.default.createElement(l.Route,{index:!0,element:a.default.createElement(c.EmptyEditor,null)}),a.default.createElement(l.Route,{path:":id/:handle",element:a.default.createElement(d.Editor,null)})),a.default.createElement(l.Route,{path:"rules",element:a.default.createElement("div",null,"Rules")}),a.default.createElement(l.Route,{path:":namespace",element:a.default.createElement(u.FormSettings,null)})))))}},98104:function(e,t,n){var r=this&&this.__makeTemplateObject||function(e,t){return Object.defineProperty?Object.defineProperty(e,"raw",{value:t}):e.raw=t,e},a=this&&this.__importDefault||function(e){return e&&e.__esModule?e:{default:e}};Object.defineProperty(t,"__esModule",{value:!0}),t.GroupsCollection=t.FormSettingsContainer=t.FormSettingsWrapper=void 0;var o,i,l,u=n(240),d=a(n(12788));t.FormSettingsWrapper=d.default.div(o||(o=r(["\n  display: flex;\n  height: 100%;\n  background: ",";\n"],["\n  display: flex;\n  height: 100%;\n  background: ",";\n"])),u.colors.white),t.FormSettingsContainer=d.default.div(i||(i=r(["\n  width: calc(100% - 300px);\n\n  padding: ",";\n"],["\n  width: calc(100% - 300px);\n\n  padding: ",";\n"])),u.spacings.lg),t.GroupsCollection=d.default.div(l||(l=r(["\n  display: flex;\n  flex-direction: column;\n  gap: ",";\n"],["\n  display: flex;\n  flex-direction: column;\n  gap: ",";\n"])),u.spacings.xl)},39130:function(e,t,n){var r=this&&this.__spreadArray||function(e,t,n){if(n||2===arguments.length)for(var r,a=0,o=t.length;a<o;a++)!r&&a in t||(r||(r=Array.prototype.slice.call(t,0,a)),r[a]=t[a]);return e.concat(r||Array.prototype.slice.call(t))},a=this&&this.__importDefault||function(e){return e&&e.__esModule?e:{default:e}};Object.defineProperty(t,"__esModule",{value:!0}),t.FormSettings=void 0;var o=a(n(67294)),i=n(49818),l=n(82384),u=n(66264),d=n(23493),c=n(62762),s=n(98104),f=n(62054);t.FormSettings=function(){var e=(0,i.useParams)().namespace,t=(0,d.useQueryFormSettings)(),n=t.data,a=t.isFetching,p=(0,f.useValueUpdateGenerator)(e);if(!n&&a)return o.default.createElement("div",null,"Loading...");var v=n.find((function(t){return t.handle===e}));if(!v)return null;var h=v.groups,m=v.properties;return o.default.createElement(s.FormSettingsWrapper,null,o.default.createElement(u.Sidebar,null),o.default.createElement(s.FormSettingsContainer,null,o.default.createElement(s.GroupsCollection,null,r(r([],h,!0),[null],!1).map((function(t,n){var r=m.filter((function(e){return!t&&!e.group||(null==t?void 0:t.handle)===e.group})).sort((function(e,t){return e.order-t.order}));return r.length?o.default.createElement(c.Group,{key:n,group:t},r.map((function(t){return o.default.createElement(l.FormControlGenerator,{key:t.handle,namespace:e,property:t,onValueUpdate:p(t)})}))):null})))))}},28741:function(e,t,n){var r=this&&this.__makeTemplateObject||function(e,t){return Object.defineProperty?Object.defineProperty(e,"raw",{value:t}):e.raw=t,e},a=this&&this.__importDefault||function(e){return e&&e.__esModule?e:{default:e}};Object.defineProperty(t,"__esModule",{value:!0}),t.GroupHeader=t.GroupContainer=t.GroupWrapper=void 0;var o,i,l,u=n(240),d=a(n(12788));t.GroupWrapper=d.default.div(o||(o=r([""],[""]))),t.GroupContainer=d.default.div(i||(i=r(["\n  display: grid;\n  gap: ",";\n  grid-template-columns: repeat(2, 1fr);\n"],["\n  display: grid;\n  gap: ",";\n  grid-template-columns: repeat(2, 1fr);\n"])),u.spacings.xl),t.GroupHeader=d.default.h2(l||(l=r(["\n  padding: 0 0 ",";\n  margin: 0;\n"],["\n  padding: 0 0 ",";\n  margin: 0;\n"])),u.spacings.lg)},62762:function(e,t,n){var r=this&&this.__importDefault||function(e){return e&&e.__esModule?e:{default:e}};Object.defineProperty(t,"__esModule",{value:!0}),t.Group=void 0;var a=r(n(67294)),o=r(n(87351)),i=n(28741);t.Group=function(e){var t=e.group,n=e.children;return a.default.createElement(i.GroupWrapper,null,!!(null==t?void 0:t.label)&&a.default.createElement(i.GroupHeader,null,(0,o.default)(t.label)),a.default.createElement(i.GroupContainer,null,n))}},62054:function(e,t,n){Object.defineProperty(t,"__esModule",{value:!0}),t.useValueUpdateGenerator=void 0;var r=n(67294),a=n(16430),o=n(85767),i=n(61552);t.useValueUpdateGenerator=function(e){var t=(0,o.useAppDispatch)(),n=(0,r.useCallback)((function(n){return function(r){t((function(t,o){t((0,i.modifySettings)({namespace:e,key:n.handle,value:(0,a.applyMiddleware)(r,n.middleware,o)}))}))}}),[e,t]);return n}},80968:function(e,t,n){var r=this&&this.__makeTemplateObject||function(e,t){return Object.defineProperty?Object.defineProperty(e,"raw",{value:t}):e.raw=t,e},a=this&&this.__importDefault||function(e){return e&&e.__esModule?e:{default:e}};Object.defineProperty(t,"__esModule",{value:!0}),t.SaveButton=t.SaveButtonWrapper=t.TabsWrapper=t.FormName=t.Heading=t.TabWrapper=void 0;var o,i,l,u,d,c,s=n(240),f=a(n(12788));t.TabWrapper=f.default.nav(o||(o=r(["\n  position: relative;\n\n  display: grid;\n  grid-template-columns: 300px auto min-content;\n  align-items: center;\n\n  height: 50px;\n  flex: 0 0 50px;\n\n  box-sizing: border-box;\n  overflow-x: hidden;\n\n  background: ",";\n  box-shadow: inset 0 -1px 0 0 rgb(154 165 177 / 25%);\n"],["\n  position: relative;\n\n  display: grid;\n  grid-template-columns: 300px auto min-content;\n  align-items: center;\n\n  height: 50px;\n  flex: 0 0 50px;\n\n  box-sizing: border-box;\n  overflow-x: hidden;\n\n  background: ",";\n  box-shadow: inset 0 -1px 0 0 rgb(154 165 177 / 25%);\n"])),s.colors.gray050),t.Heading=f.default.h1(i||(i=r(["\n  position: relative;\n\n  display: flex;\n  align-items: center;\n  gap: ",";\n\n  margin: 0;\n  padding: 0 ",";\n\n  a {\n    svg {\n      position: relative;\n\n      width: 18px;\n      height: 18px;\n      color: ",";\n\n      transition: all 0.2s ease-out;\n    }\n\n    &:hover svg {\n      color: ",";\n      transform: translateX(-5px);\n    }\n  }\n"],["\n  position: relative;\n\n  display: flex;\n  align-items: center;\n  gap: ",";\n\n  margin: 0;\n  padding: 0 ",";\n\n  a {\n    svg {\n      position: relative;\n\n      width: 18px;\n      height: 18px;\n      color: ",";\n\n      transition: all 0.2s ease-out;\n    }\n\n    &:hover svg {\n      color: ",";\n      transform: translateX(-5px);\n    }\n  }\n"])),s.spacings.sm,s.spacings.lg,s.colors.gray200,s.colors.gray700),t.FormName=f.default.span(l||(l=r(["\n  font-size: 18px;\n  font-weight: 700;\n  line-height: 1.2;\n  color: ",";\n"],["\n  font-size: 18px;\n  font-weight: 700;\n  line-height: 1.2;\n  color: ",";\n"])),s.colors.gray700),t.TabsWrapper=f.default.div(u||(u=r(["\n  display: flex;\n  align-self: flex-end;\n\n  a {\n    display: flex;\n    align-items: center;\n\n    height: 42px;\n    padding: 0 12px;\n\n    white-space: nowrap;\n\n    color: var(--light-text-color);\n    border-radius: "," "," 0 0;\n\n    &:hover {\n      text-decoration: none;\n      background-color: rgba(154, 165, 177, 0.15);\n    }\n\n    &.active {\n      background: ",";\n\n      box-shadow: 0 0 0 1px ",", 0 2px 12px rgb(205 216 228 / 50%) !important;\n\n      color: ",";\n    }\n  }\n"],["\n  display: flex;\n  align-self: flex-end;\n\n  a {\n    display: flex;\n    align-items: center;\n\n    height: 42px;\n    padding: 0 12px;\n\n    white-space: nowrap;\n\n    color: var(--light-text-color);\n    border-radius: "," "," 0 0;\n\n    &:hover {\n      text-decoration: none;\n      background-color: rgba(154, 165, 177, 0.15);\n    }\n\n    &.active {\n      background: ",";\n\n      box-shadow: 0 0 0 1px ",", 0 2px 12px rgb(205 216 228 / 50%) !important;\n\n      color: ",";\n    }\n  }\n"])),s.borderRadius.md,s.borderRadius.md,s.colors.white,s.colors.gray200,s.colors.gray700),t.SaveButtonWrapper=f.default.div(d||(d=r(["\n  padding: 0 ",";\n"],["\n  padding: 0 ",";\n"])),s.spacings.lg),t.SaveButton=f.default.button.attrs({className:"btn submit"})(c||(c=r([""],[""])))},19179:function(e,t,n){var r=this&&this.__importDefault||function(e){return e&&e.__esModule?e:{default:e}};Object.defineProperty(t,"__esModule",{value:!0}),t.Tabs=void 0;var a=r(n(67294)),o=n(49818),i=n(85767),l=n(80926),u=r(n(27796)),d=n(12020),c=n(23493),s=r(n(87351)),f=n(80968);t.Tabs=function(){var e=(0,i.useAppDispatch)(),t=(0,c.useQueryFormSettings)().data,n=function(){e((0,l.save)())};return(0,d.useOnKeypress)({callback:function(e){if("s"===e.key){var t=window.navigator.platform.match(/Mac/);if(t&&!e.metaKey)return;if(!t&&!e.ctrlKey)return;return e.preventDefault(),n(),!1}},type:"keydown"}),a.default.createElement(f.TabWrapper,null,a.default.createElement(f.Heading,null,a.default.createElement(o.Link,{to:"..",title:(0,s.default)("Back to form list")},a.default.createElement(u.default,null)),a.default.createElement(f.FormName,null,"My form name")),a.default.createElement(f.TabsWrapper,null,a.default.createElement(o.NavLink,{to:"",end:!0},(0,s.default)("Layout")),a.default.createElement(o.NavLink,{to:"notifications"},(0,s.default)("Notifications")),a.default.createElement(o.NavLink,{to:"rules"},(0,s.default)("Rules")),a.default.createElement(o.NavLink,{to:"integrations"},(0,s.default)("Integrations")),(t||[]).map((function(e){return a.default.createElement(o.NavLink,{key:e.handle,to:e.handle},(0,s.default)(e.label))}))),a.default.createElement(f.SaveButtonWrapper,null,a.default.createElement(f.SaveButton,{onClick:n},(0,s.default)("Save"))))}},72467:function(e,t,n){var r=this&&this.__makeTemplateObject||function(e,t){return Object.defineProperty?Object.defineProperty(e,"raw",{value:t}):e.raw=t,e},a=this&&this.__importDefault||function(e){return e&&e.__esModule?e:{default:e}};Object.defineProperty(t,"__esModule",{value:!0}),t.SettingsWrapper=t.EditorWrapper=void 0;var o,i,l=n(240),u=a(n(12788));t.EditorWrapper=u.default.div(o||(o=r(["\n  flex: 1;\n\n  background: ",";\n  padding: ",";\n"],["\n  flex: 1;\n\n  background: ",";\n  padding: ",";\n"])),l.colors.white,l.spacings.md),t.SettingsWrapper=u.default.div(i||(i=r(["\n  display: flex;\n  flex-direction: column;\n  gap: ",";\n"],["\n  display: flex;\n  flex-direction: column;\n  gap: ",";\n"])),l.spacings.xl)},678:function(e,t,n){var r=this&&this.__importDefault||function(e){return e&&e.__esModule?e:{default:e}};Object.defineProperty(t,"__esModule",{value:!0}),t.Editor=void 0;var a=r(n(67294)),o=n(47785),i=n(49818),l=n(1847),u=n(68887),d=n(32156),c=n(72467),s=n(26678);t.Editor=function(){var e=(0,i.useParams)().id,t=((0,o.useDispatch)(),(0,o.useSelector)((0,u.selectIntegration)(Number(e))));if(!t)return a.default.createElement(s.EmptyEditor,null);var n=t.id,r=t.handle,f=t.name,p=t.description,v=(t.enabled,t.settings);return a.default.createElement(c.EditorWrapper,null,a.default.createElement("h1",{title:r},f),!!p&&a.default.createElement("p",null,p),a.default.createElement(l.Space,null),a.default.createElement(c.SettingsWrapper,null,v.map((function(e){return a.default.createElement(d.Setting,{key:e.handle,id:n,setting:e})}))))}},26678:function(e,t,n){var r=this&&this.__importDefault||function(e){return e&&e.__esModule?e:{default:e}};Object.defineProperty(t,"__esModule",{value:!0}),t.EmptyEditor=void 0;var a=r(n(67294)),o=r(n(84849)),i=n(80306),l=n(72467);t.EmptyEditor=function(){return a.default.createElement(l.EditorWrapper,null,a.default.createElement(i.ThemedSkeleton,null,a.default.createElement(o.default,{width:120,height:20}),a.default.createElement("br",null),a.default.createElement(o.default,{width:100,height:10}),a.default.createElement(o.default,{width:50,height:20}),a.default.createElement("br",null),a.default.createElement(o.default,{width:200,height:10}),a.default.createElement(o.default,{width:500,height:10}),a.default.createElement(o.default,{height:30}),a.default.createElement("br",null),a.default.createElement(o.default,{width:150,height:10}),a.default.createElement(o.default,{width:300,height:10}),a.default.createElement(o.default,{height:30})))}},32156:function(e,t){Object.defineProperty(t,"__esModule",{value:!0}),t.Setting=void 0;t.Setting=function(e){e.id,e.setting;return null}},36408:function(e,t,n){var r=this&&this.__makeTemplateObject||function(e,t){return Object.defineProperty?Object.defineProperty(e,"raw",{value:t}):e.raw=t,e},a=this&&this.__importDefault||function(e){return e&&e.__esModule?e:{default:e}};Object.defineProperty(t,"__esModule",{value:!0}),t.IntegrationsWrapper=void 0;var o,i=a(n(12788));t.IntegrationsWrapper=i.default.div(o||(o=r(["\n  display: flex;\n  height: 100%;\n"],["\n  display: flex;\n  height: 100%;\n"])))},70793:function(e,t,n){var r=this&&this.__importDefault||function(e){return e&&e.__esModule?e:{default:e}};Object.defineProperty(t,"__esModule",{value:!0}),t.Integrations=void 0;var a=r(n(67294)),o=n(49818),i=n(69412),l=n(36408);t.Integrations=function(){return a.default.createElement(l.IntegrationsWrapper,null,a.default.createElement(i.List,null),a.default.createElement(o.Outlet,null))}},35615:function(e,t,n){var r=this&&this.__importDefault||function(e){return e&&e.__esModule?e:{default:e}};Object.defineProperty(t,"__esModule",{value:!0}),t.CategorySkeleton=void 0;var a=r(n(67294)),o=r(n(84849)),i=n(80306),l=n(65937);t.CategorySkeleton=function(){return a.default.createElement(i.ThemedSkeleton,null,a.default.createElement(l.Wrapper,null,a.default.createElement(l.Label,null,a.default.createElement(o.default,{width:50})),a.default.createElement(l.ChildrenWrapper,{style:{padding:14}},[0,1,2].map((function(e){return a.default.createElement("div",{key:e,style:{display:"flex",gap:10,alignItems:"center"}},a.default.createElement(o.default,{width:20,height:20,circle:!0}),a.default.createElement("div",{style:{flexGrow:2}},a.default.createElement(o.default,{width:100,style:{top:2}})),a.default.createElement(o.default,{width:10,height:10,circle:!0,style:{top:6}}))})))))}},65937:function(e,t,n){var r=this&&this.__makeTemplateObject||function(e,t){return Object.defineProperty?Object.defineProperty(e,"raw",{value:t}):e.raw=t,e},a=this&&this.__importDefault||function(e){return e&&e.__esModule?e:{default:e}};Object.defineProperty(t,"__esModule",{value:!0}),t.ChildrenWrapper=t.Label=t.Wrapper=void 0;var o,i,l,u=n(240),d=a(n(12788));t.Wrapper=d.default.div(o||(o=r([""],[""]))),t.Label=d.default.span(i||(i=r(["\n  padding-left: ",";\n\n  font-weight: 700;\n  font-size: 11px;\n  color: ",";\n\n  text-transform: uppercase;\n"],["\n  padding-left: ",";\n\n  font-weight: 700;\n  font-size: 11px;\n  color: ",";\n\n  text-transform: uppercase;\n"])),u.spacings.md,u.colors.gray550),t.ChildrenWrapper=d.default.div(l||(l=r(["\n  display: flex;\n  flex-direction: column;\n  gap: ",";\n\n  padding: "," 0;\n"],["\n  display: flex;\n  flex-direction: column;\n  gap: ",";\n\n  padding: "," 0;\n"])),u.spacings.xs,u.spacings.xs)},84673:function(e,t,n){var r=this&&this.__assign||function(){return r=Object.assign||function(e){for(var t,n=1,r=arguments.length;n<r;n++)for(var a in t=arguments[n])Object.prototype.hasOwnProperty.call(t,a)&&(e[a]=t[a]);return e},r.apply(this,arguments)},a=this&&this.__importDefault||function(e){return e&&e.__esModule?e:{default:e}};Object.defineProperty(t,"__esModule",{value:!0}),t.Category=void 0;var o=a(n(67294)),i=n(54717),l=n(65937);t.Category=function(e){var t=e.label,n=e.children;return o.default.createElement(l.Wrapper,null,o.default.createElement(l.Label,null,t),o.default.createElement(l.ChildrenWrapper,null,n.map((function(e){return o.default.createElement(i.Integration,r({key:e.id},e))}))))}},94319:function(e,t,n){var r=this&&this.__makeTemplateObject||function(e,t){return Object.defineProperty?Object.defineProperty(e,"raw",{value:t}):e.raw=t,e},a=this&&this.__importDefault||function(e){return e&&e.__esModule?e:{default:e}};Object.defineProperty(t,"__esModule",{value:!0}),t.Status=t.Name=t.Icon=t.Wrapper=void 0;var o,i,l,u,d=n(240),c=a(n(12788));t.Wrapper=c.default.div(o||(o=r(["\n  > a {\n    display: flex;\n    align-items: center;\n    gap: ",";\n\n    padding: "," ",";\n    border-radius: ",";\n\n    color: ",";\n    font-size: 12px;\n    line-height: 12px;\n\n    transition: background-color 0.2s ease-out;\n    text-decoration: none;\n\n    &.active {\n      background-color: ",";\n    }\n\n    &:hover:not(.active) {\n      background-color: ",";\n    }\n  }\n"],["\n  > a {\n    display: flex;\n    align-items: center;\n    gap: ",";\n\n    padding: "," ",";\n    border-radius: ",";\n\n    color: ",";\n    font-size: 12px;\n    line-height: 12px;\n\n    transition: background-color 0.2s ease-out;\n    text-decoration: none;\n\n    &.active {\n      background-color: ",";\n    }\n\n    &:hover:not(.active) {\n      background-color: ",";\n    }\n  }\n"])),d.spacings.sm,d.spacings.sm,d.spacings.md,d.borderRadius.lg,d.colors.gray700,d.colors.gray200,d.colors.gray100),t.Icon=c.default.div(i||(i=r(["\n  display: block;\n  width: 20px;\n  height: 20px;\n"],["\n  display: block;\n  width: 20px;\n  height: 20px;\n"]))),t.Name=c.default.div(l||(l=r(["\n  flex-grow: 1;\n"],["\n  flex-grow: 1;\n"]))),t.Status=c.default.div(u||(u=r(["\n  content: '';\n\n  justify-self: flex-end;\n\n  width: 10px;\n  height: 10px;\n\n  border: 1px solid\n    ",";\n  border-radius: 100%;\n\n  background-color: ",";\n\n  transition: all 0.3s ease-out;\n"],["\n  content: '';\n\n  justify-self: flex-end;\n\n  width: 10px;\n  height: 10px;\n\n  border: 1px solid\n    ",";\n  border-radius: 100%;\n\n  background-color: ",";\n\n  transition: all 0.3s ease-out;\n"])),(function(e){return e.enabled?"transparent":d.colors.gray550}),(function(e){return e.enabled?d.colors.teal550:"transparent"}))},54717:function(e,t,n){var r=this&&this.__importDefault||function(e){return e&&e.__esModule?e:{default:e}};Object.defineProperty(t,"__esModule",{value:!0}),t.Integration=void 0;var a=r(n(67294)),o=n(47785),i=n(49818),l=n(68887),u=r(n(35858)),d=n(94319);t.Integration=function(e){var t=e.id,n=e.name,r=e.handle,c=e.icon,s=(0,o.useSelector)((0,l.selectIntegration)(t));return a.default.createElement(d.Wrapper,null,a.default.createElement(i.NavLink,{to:"".concat(t,"/").concat(r)},a.default.createElement(d.Icon,null,!!c&&a.default.createElement("img",{src:c}),!c&&a.default.createElement(u.default,null)),a.default.createElement(d.Name,null,n),a.default.createElement(d.Status,{enabled:s.enabled})))}},1663:function(e,t,n){var r=this&&this.__makeTemplateObject||function(e,t){return Object.defineProperty?Object.defineProperty(e,"raw",{value:t}):e.raw=t,e},a=this&&this.__importDefault||function(e){return e&&e.__esModule?e:{default:e}};Object.defineProperty(t,"__esModule",{value:!0}),t.Wrapper=void 0;var o,i=n(240),l=a(n(12788));t.Wrapper=l.default.ul(o||(o=r(["\n  display: flex;\n  flex-direction: column;\n  gap: ",";\n\n  list-style: none;\n"],["\n  display: flex;\n  flex-direction: column;\n  gap: ",";\n\n  list-style: none;\n"])),i.spacings.sm)},69412:function(e,t,n){var r=this&&this.__assign||function(){return r=Object.assign||function(e){for(var t,n=1,r=arguments.length;n<r;n++)for(var a in t=arguments[n])Object.prototype.hasOwnProperty.call(t,a)&&(e[a]=t[a]);return e},r.apply(this,arguments)},a=this&&this.__createBinding||(Object.create?function(e,t,n,r){void 0===r&&(r=n);var a=Object.getOwnPropertyDescriptor(t,n);a&&!("get"in a?!t.__esModule:a.writable||a.configurable)||(a={enumerable:!0,get:function(){return t[n]}}),Object.defineProperty(e,r,a)}:function(e,t,n,r){void 0===r&&(r=n),e[r]=t[n]}),o=this&&this.__setModuleDefault||(Object.create?function(e,t){Object.defineProperty(e,"default",{enumerable:!0,value:t})}:function(e,t){e.default=t}),i=this&&this.__importStar||function(e){if(e&&e.__esModule)return e;var t={};if(null!=e)for(var n in e)"default"!==n&&Object.prototype.hasOwnProperty.call(e,n)&&a(t,e,n);return o(t,e),t};Object.defineProperty(t,"__esModule",{value:!0}),t.List=void 0;var l=i(n(67294)),u=n(49818),d=n(66264),c=n(41648),s=n(84673),f=n(35615),p=n(1663),v={crm:"CRM",mailing_list:"Email Marketing",payment:"Payment"};t.List=function(){var e=(0,u.useParams)(),t=e.formId,n=e.id,a=(0,u.useNavigate)(),o=(0,c.useQueryFormIntegrations)(Number(null!=t?t:0)),i=o.data,h=o.isFetching;if((0,l.useEffect)((function(){if(!n&&i){var e=i.find(Boolean);e&&a("".concat(e.id,"/").concat(e.handle))}}),[n,i]),!i&&h)return l.default.createElement(d.Sidebar,null,l.default.createElement(p.Wrapper,null,l.default.createElement(f.CategorySkeleton,null)));var m={};return i.forEach((function(e){var t,n=e.type;m[n]||(m[n]={type:n,label:null!==(t=v[n])&&void 0!==t?t:"Other",children:[]}),m[n].children.push(e)})),l.default.createElement(d.Sidebar,{lean:!0},l.default.createElement(p.Wrapper,null,Object.values(m).map((function(e){return l.default.createElement(s.Category,r({key:e.type},e))}))))}},62277:function(e,t,n){var r=this&&this.__makeTemplateObject||function(e,t){return Object.defineProperty?Object.defineProperty(e,"raw",{value:t}):e.raw=t,e},a=this&&this.__importDefault||function(e){return e&&e.__esModule?e:{default:e}};Object.defineProperty(t,"__esModule",{value:!0}),t.Label=t.CellFieldWrapper=void 0;var o,i,l=a(n(12788));t.CellFieldWrapper=l.default.div(o||(o=r([""],[""]))),t.Label=l.default.label(i||(i=r([""],[""])))},53293:function(e,t,n){var r=this&&this.__importDefault||function(e){return e&&e.__esModule?e:{default:e}};Object.defineProperty(t,"__esModule",{value:!0}),t.CellField=void 0;var a=r(n(67294)),o=n(47785),i=n(85767),l=n(45323),u=n(56859),d=n(42367),c=n(62277);t.CellField=function(e){var t=e.uid,n=(0,o.useSelector)((0,u.selectField)(t)),r=(0,d.useFieldType)(null==n?void 0:n.typeClass),s=(0,i.useAppDispatch)();return void 0===(null==n?void 0:n.properties)?null:a.default.createElement(c.CellFieldWrapper,{onClick:function(){s((0,l.setFocusedItem)({type:l.FocusType.Field,uid:t}))}},a.default.createElement(c.Label,null,n.properties.label||(null==r?void 0:r.name)),a.default.createElement("div",null,a.default.createElement("input",{type:"text",className:"text fullwidth"})))}},27607:function(e,t,n){var r=this&&this.__importDefault||function(e){return e&&e.__esModule?e:{default:e}};Object.defineProperty(t,"__esModule",{value:!0}),t.CellLayout=void 0;var a=r(n(67294)),o=n(47785),i=n(1965),l=n(56206);t.CellLayout=function(e){var t=e.uid,n=(0,o.useSelector)((0,i.selectLayout)(t));return n?a.default.createElement(l.Layout,{layout:n}):null}},35917:function(e,t,n){var r=this&&this.__makeTemplateObject||function(e,t){return Object.defineProperty?Object.defineProperty(e,"raw",{value:t}):e.raw=t,e},a=this&&this.__importDefault||function(e){return e&&e.__esModule?e:{default:e}};Object.defineProperty(t,"__esModule",{value:!0}),t.Wrapper=void 0;var o,i=n(39920),l=a(n(12788));t.Wrapper=(0,l.default)(i.animated.div)(o||(o=r(["\n  //border: 3px dashed green;\n\n  flex: 1 0;\n  overflow: hidden;\n"],["\n  //border: 3px dashed green;\n\n  flex: 1 0;\n  overflow: hidden;\n"])))},26528:function(e,t,n){var r=this&&this.__importDefault||function(e){return e&&e.__esModule?e:{default:e}};Object.defineProperty(t,"__esModule",{value:!0}),t.Cell=void 0;var a=r(n(67294)),o=n(84520),i=n(11037),l=n(85767),u=n(99614),d=n(53293),c=n(27607),s=n(35917);t.Cell=function(e){var t,n=e.cell,r=e.order,f=(0,l.useAppDispatch)(),p=(0,o.useDrag)((function(){return{type:"LayoutField",item:function(){return f((0,u.setCell)(n.uid)),{}},end:function(){f((0,u.setCell)(void 0))},collect:function(e){return{isDragging:e.isDragging()}}}}),[n.uid])[1];switch(n.type){case i.CellType.Field:t=d.CellField;break;case i.CellType.Layout:t=c.CellLayout}return a.default.createElement(s.Wrapper,{ref:p,style:{order:r}},a.default.createElement(t,{uid:n.targetUid}))}},24982:function(e,t,n){var r=this&&this.__makeTemplateObject||function(e,t){return Object.defineProperty?Object.defineProperty(e,"raw",{value:t}):e.raw=t,e},a=this&&this.__importDefault||function(e){return e&&e.__esModule?e:{default:e}};Object.defineProperty(t,"__esModule",{value:!0}),t.FieldLayoutWrapper=void 0;var o,i=a(n(12788));t.FieldLayoutWrapper=i.default.div(o||(o=r(["\n  flex: 1;\n  padding: var(--padding);\n"],["\n  flex: 1;\n  padding: var(--padding);\n"])))},19149:function(e,t,n){var r=this&&this.__importDefault||function(e){return e&&e.__esModule?e:{default:e}};Object.defineProperty(t,"__esModule",{value:!0}),t.FieldLayout=void 0;var a=r(n(67294)),o=n(47785),i=n(45323),l=n(55014),u=n(29178),d=n(24982);t.FieldLayout=function(){var e=(0,o.useSelector)(i.selectCurrentPage);return a.default.createElement(d.FieldLayoutWrapper,null,a.default.createElement(u.PageTabs,null),e&&a.default.createElement(l.Page,{page:e}))}},46804:function(e,t,n){var r=this&&this.__makeTemplateObject||function(e,t){return Object.defineProperty?Object.defineProperty(e,"raw",{value:t}):e.raw=t,e},a=this&&this.__importDefault||function(e){return e&&e.__esModule?e:{default:e}};Object.defineProperty(t,"__esModule",{value:!0}),t.Wrapper=void 0;var o,i=n(240),l=a(n(12788));t.Wrapper=l.default.div(o||(o=r(["\n  //border: 3px dashed blue;\n\n  display: flex;\n  flex-direction: column;\n  gap: ",";\n"],["\n  //border: 3px dashed blue;\n\n  display: flex;\n  flex-direction: column;\n  gap: ",";\n"])),i.spacings.lg)},56206:function(e,t,n){var r=this&&this.__importDefault||function(e){return e&&e.__esModule?e:{default:e}};Object.defineProperty(t,"__esModule",{value:!0}),t.Layout=void 0;var a=r(n(67294)),o=n(47785),i=n(33314),l=n(6294),u=n(46804);t.Layout=function(e){var t=e.layout,n=(0,o.useSelector)((0,i.selectRowsInLayout)(t));return t?a.default.createElement(u.Wrapper,null,n.map((function(e){return a.default.createElement(l.Row,{row:e,key:e.uid})}))):null}},69876:function(e,t,n){var r=this&&this.__makeTemplateObject||function(e,t){return Object.defineProperty?Object.defineProperty(e,"raw",{value:t}):e.raw=t,e},a=this&&this.__importDefault||function(e){return e&&e.__esModule?e:{default:e}};Object.defineProperty(t,"__esModule",{value:!0}),t.PageTabsContainer=t.PageTabsWrapper=void 0;var o,i,l=n(240),u=a(n(12788));t.PageTabsWrapper=u.default.div(o||(o=r(["\n  display: flex;\n  justify-content: space-between;\n\n  border-bottom: 1px solid ",";\n  margin-bottom: ",";\n"],["\n  display: flex;\n  justify-content: space-between;\n\n  border-bottom: 1px solid ",";\n  margin-bottom: ",";\n"])),l.colors.gray200,l.spacings.lg),t.PageTabsContainer=u.default.div(i||(i=r(["\n  display: flex;\n"],["\n  display: flex;\n"])))},29178:function(e,t,n){var r=this&&this.__assign||function(){return r=Object.assign||function(e){for(var t,n=1,r=arguments.length;n<r;n++)for(var a in t=arguments[n])Object.prototype.hasOwnProperty.call(t,a)&&(e[a]=t[a]);return e},r.apply(this,arguments)},a=this&&this.__importDefault||function(e){return e&&e.__esModule?e:{default:e}};Object.defineProperty(t,"__esModule",{value:!0}),t.PageTabs=void 0;var o=a(n(67294)),i=n(47785),l=n(68974),u=n(78650),d=n(60859),c=n(69876);t.PageTabs=function(){var e=(0,i.useSelector)(l.selectPages);return o.default.createElement(c.PageTabsWrapper,null,o.default.createElement(c.PageTabsContainer,null,e.map((function(e){return o.default.createElement(d.Tab,r({key:e.uid},e))}))),o.default.createElement(u.NewTab,null))}},78650:function(e,t,n){var r=this&&this.__importDefault||function(e){return e&&e.__esModule?e:{default:e}};Object.defineProperty(t,"__esModule",{value:!0}),t.NewTab=void 0;var a=r(n(67294)),o=n(85767),i=n(6175),l=n(86799);t.NewTab=function(){var e=(0,o.useAppDispatch)();return a.default.createElement(l.NewTabWrapper,{onClick:function(){e((0,i.addNewPage)())}},"Add New Page")}},86799:function(e,t,n){var r=this&&this.__makeTemplateObject||function(e,t){return Object.defineProperty?Object.defineProperty(e,"raw",{value:t}):e.raw=t,e},a=this&&this.__importDefault||function(e){return e&&e.__esModule?e:{default:e}};Object.defineProperty(t,"__esModule",{value:!0}),t.NewTabWrapper=t.TabWrapper=void 0;var o,i,l=n(240),u=a(n(12788));t.TabWrapper=u.default.div(o||(o=r(["\n  box-shadow: 0 0 0 1px #cdd8e4, 0 2px 12px rgb(205 216 228 / 50%);\n\n  padding: 4px 18px;\n  border-radius: 4px 4px 0 0;\n  background: ",";\n\n  &:hover {\n    cursor: pointer;\n  }\n"],["\n  box-shadow: 0 0 0 1px #cdd8e4, 0 2px 12px rgb(205 216 228 / 50%);\n\n  padding: 4px 18px;\n  border-radius: 4px 4px 0 0;\n  background: ",";\n\n  &:hover {\n    cursor: pointer;\n  }\n"])),(function(e){return e.active?l.colors.gray050:"white"})),t.NewTabWrapper=(0,u.default)(t.TabWrapper)(i||(i=r(["\n  justify-self: flex-end;\n"],["\n  justify-self: flex-end;\n"])))},60859:function(e,t,n){var r=this&&this.__importDefault||function(e){return e&&e.__esModule?e:{default:e}};Object.defineProperty(t,"__esModule",{value:!0}),t.Tab=void 0;var a=r(n(67294)),o=n(47785),i=n(85767),l=n(45323),u=n(86799);t.Tab=function(e){var t=(0,o.useSelector)(l.selectCurrentPage).uid,n=(0,i.useAppDispatch)();return a.default.createElement(u.TabWrapper,{active:t===e.uid,onClick:function(){n((0,l.setPage)(e.uid))}},e.label)}},55014:function(e,t,n){var r=this&&this.__importDefault||function(e){return e&&e.__esModule?e:{default:e}};Object.defineProperty(t,"__esModule",{value:!0}),t.Page=void 0;var a=r(n(67294)),o=n(47785),i=n(1965),l=n(56206),u=n(19611);t.Page=function(e){var t=e.page,n=(0,o.useSelector)((0,i.selectPageLayout)(t));return a.default.createElement(u.PageWrapper,null,n&&a.default.createElement(l.Layout,{layout:n}))}},19611:function(e,t,n){var r=this&&this.__makeTemplateObject||function(e,t){return Object.defineProperty?Object.defineProperty(e,"raw",{value:t}):e.raw=t,e},a=this&&this.__importDefault||function(e){return e&&e.__esModule?e:{default:e}};Object.defineProperty(t,"__esModule",{value:!0}),t.PageWrapper=void 0;var o,i=n(240),l=a(n(12788));t.PageWrapper=l.default.div(o||(o=r(["\n  display: flex;\n  flex-direction: column;\n  gap: ",";\n"],["\n  display: flex;\n  flex-direction: column;\n  gap: ",";\n"])),i.spacings.md)},58974:function(e,t,n){Object.defineProperty(t,"__esModule",{value:!0}),t.useRowDrop=void 0;var r=n(67294),a=n(84520),o=n(47785),i=n(39920),l=n(85767),u=n(99614);t.useRowDrop=function(e,t){var n=(0,l.useAppDispatch)(),d=(0,o.useSelector)(u.selectCurrentPosition),c=(0,r.useRef)(null),s=(0,r.useRef)(d),f=(0,r.useRef)(0),p=(0,r.useState)(0),v=p[0],h=p[1],m=(0,r.useState)(0),g=m[0],y=m[1];(0,r.useEffect)((function(){var e=c.current.getBoundingClientRect();h(e.width),y(e.left)}),[t]);var _=(0,a.useDrop)((function(){return{accept:["BaseField","LayoutField"],canDrop:function(e,t){return t.isOver({shallow:!0})},drop:function(t){console.log("Row:",e.uid),console.log("Cell:",s.current,t.type)},hover:function(e,r){var a=v/(t+1);f.current!==a&&(f.current=a,console.log(a));var o=r.getClientOffset().x-g,i=Math.floor(o/a);s.current!==i&&(s.current=i,n((0,u.setPosition)(i)))},collect:function(e){return{isOver:e.isOver({shallow:!0}),canDrop:e.canDrop()}}}}),[g,v]),b=_[0].isOver;(0,_[1])(c);var x=(0,i.useSpring)({to:{background:"grey",flexGrow:1,transform:"translateX(".concat(f.current*d,"px)")}});return{dropRef:c,placeholderStyle:x,isOver:b}}},16488:function(e,t,n){var r=this&&this.__makeTemplateObject||function(e,t){return Object.defineProperty?Object.defineProperty(e,"raw",{value:t}):e.raw=t,e},a=this&&this.__importDefault||function(e){return e&&e.__esModule?e:{default:e}};Object.defineProperty(t,"__esModule",{value:!0}),t.Placeholder=t.Container=t.Wrapper=void 0;var o,i,l,u=n(39920),d=n(240),c=a(n(12788));t.Wrapper=c.default.div(o||(o=r([""],[""]))),t.Container=c.default.div(i||(i=r(["\n  //border: 3px dashed red;\n\n  position: relative;\n\n  display: flex;\n  justify-content: space-between;\n  align-items: stretch;\n  gap: ",";\n"],["\n  //border: 3px dashed red;\n\n  position: relative;\n\n  display: flex;\n  justify-content: space-between;\n  align-items: stretch;\n  gap: ",";\n"])),d.spacings.lg),t.Placeholder=(0,c.default)(u.animated.div)(l||(l=r(["\n  overflow: hidden;\n  font-size: 0;\n  line-height: 0;\n\n  border: 3px dashed red;\n"],["\n  overflow: hidden;\n  font-size: 0;\n  line-height: 0;\n\n  border: 3px dashed red;\n"])))},6294:function(e,t,n){var r=this&&this.__importDefault||function(e){return e&&e.__esModule?e:{default:e}};Object.defineProperty(t,"__esModule",{value:!0}),t.Row=void 0;var a=r(n(67294)),o=n(47785),i=n(89372),l=n(26528),u=n(58974),d=n(16488);t.Row=function(e){var t=e.row,n=(0,o.useSelector)((0,i.selectCellsInRow)(t)),r=(0,u.useRowDrop)(t,n.length),c=r.dropRef,s=r.isOver;return a.default.createElement(d.Wrapper,{ref:c},a.default.createElement(d.Container,{style:{backgroundColor:s?"pink":"white"}},n.map((function(e,t){return a.default.createElement(l.Cell,{cell:e,key:e.uid,order:10*(t+1)})}))))}},12887:function(e,t,n){var r=this&&this.__makeTemplateObject||function(e,t){return Object.defineProperty?Object.defineProperty(e,"raw",{value:t}):e.raw=t,e},a=this&&this.__importDefault||function(e){return e&&e.__esModule?e:{default:e}};Object.defineProperty(t,"__esModule",{value:!0}),t.List=t.GroupTitle=t.FieldGroupWrapper=void 0;var o,i,l,u=n(240),d=a(n(12788));t.FieldGroupWrapper=d.default.div(o||(o=r([""],[""]))),t.GroupTitle=d.default.h2(i||(i=r([""],[""]))),t.List=d.default.div(l||(l=r(["\n  display: grid;\n  grid-template-columns: 1fr 1fr;\n  gap: ",";\n\n  margin: 0;\n  padding: 0;\n"],["\n  display: grid;\n  grid-template-columns: 1fr 1fr;\n  gap: ",";\n\n  margin: 0;\n  padding: 0;\n"])),u.spacings.sm)},89829:function(e,t,n){var r=this&&this.__importDefault||function(e){return e&&e.__esModule?e:{default:e}};Object.defineProperty(t,"__esModule",{value:!0}),t.FieldGroup=void 0;var a=r(n(67294)),o=r(n(84849)),i=n(4696),l=n(42367),u=n(82344),d=n(60745),c=n(83884),s=n(12887);n(45207);t.FieldGroup=function(e){var t=e.title,n=(0,d.useSelectSearchedFields)(),r=(0,l.useFetchFieldTypes)({select:n}),f=r.data,p=r.isFetching,v=r.isError,h=r.error;return!f&&p?a.default.createElement(s.List,null,(0,u.range)(9).map((function(e){return a.default.createElement(o.default,{key:e,height:32})}))):v?a.default.createElement(i.ErrorBlock,null,h.message):a.default.createElement(s.FieldGroupWrapper,null,a.default.createElement(s.GroupTitle,null,t),a.default.createElement(s.List,null,f.map((function(e){return a.default.createElement(c.Field,{key:e.typeClass,fieldType:e})}))))}},80695:function(e,t,n){var r=this&&this.__makeTemplateObject||function(e,t){return Object.defineProperty?Object.defineProperty(e,"raw",{value:t}):e.raw=t,e},a=this&&this.__importDefault||function(e){return e&&e.__esModule?e:{default:e}};Object.defineProperty(t,"__esModule",{value:!0}),t.Icon=t.Name=t.Wrapper=void 0;var o,i,l,u=n(240),d=a(n(12788));t.Wrapper=d.default.div(o||(o=r(["\n  cursor: pointer;\n\n  display: flex;\n  gap: ",";\n  align-items: center;\n\n  height: 32px;\n\n  padding: ",";\n  overflow: hidden;\n\n  background: ",";\n  border: 1px solid ",";\n\n  font-size: 12px;\n  font-family: 'SFMono-Regular', Consolas, 'Liberation Mono', Menlo, Courier,\n    monospace;\n"],["\n  cursor: pointer;\n\n  display: flex;\n  gap: ",";\n  align-items: center;\n\n  height: 32px;\n\n  padding: ",";\n  overflow: hidden;\n\n  background: ",";\n  border: 1px solid ",";\n\n  font-size: 12px;\n  font-family: 'SFMono-Regular', Consolas, 'Liberation Mono', Menlo, Courier,\n    monospace;\n"])),u.spacings.sm,u.spacings.sm,u.colors.white,u.colors.gray200),t.Name=d.default.span(i||(i=r(["\n  flex: 1;\n  line-height: 12px;\n"],["\n  flex: 1;\n  line-height: 12px;\n"]))),t.Icon=d.default.div(l||(l=r(["\n  display: block;\n  flex-basis: 24px;\n  height: 24px;\n\n  color: ",";\n"],["\n  display: block;\n  flex-basis: 24px;\n  height: 24px;\n\n  color: ",";\n"])),u.colors.gray200)},83884:function(e,t,n){var r=this&&this.__importDefault||function(e){return e&&e.__esModule?e:{default:e}};Object.defineProperty(t,"__esModule",{value:!0}),t.Field=void 0;var a=r(n(67294)),o=n(84520),i=n(85767),l=n(73396),u=n(80695);t.Field=function(e){var t=e.fieldType,n=(0,i.useAppDispatch)(),r=(0,o.useDrag)((function(){return{type:"BaseField",item:t}}))[1];return a.default.createElement(u.Wrapper,{ref:r,onClick:function(){n((0,l.addNewField)(t))}},a.default.createElement(u.Icon,{dangerouslySetInnerHTML:{__html:t.icon}}),a.default.createElement(u.Name,null,t.name))}},84554:function(e,t,n){var r=this&&this.__makeTemplateObject||function(e,t){return Object.defineProperty?Object.defineProperty(e,"raw",{value:t}):e.raw=t,e},a=this&&this.__importDefault||function(e){return e&&e.__esModule?e:{default:e}};Object.defineProperty(t,"__esModule",{value:!0}),t.FieldListWrapper=void 0;var o,i=n(39920),l=n(240),u=a(n(12788));t.FieldListWrapper=(0,u.default)(i.animated.div)(o||(o=r(["\n  position: relative;\n  padding: ","; ;\n"],["\n  position: relative;\n  padding: ","; ;\n"])),l.spacings.lg)},18842:function(e,t,n){var r=this&&this.__importDefault||function(e){return e&&e.__esModule?e:{default:e}};Object.defineProperty(t,"__esModule",{value:!0}),t.FieldList=void 0;var a=r(n(67294)),o=n(42367),i=n(89829),l=n(65234),u=n(84554);t.FieldList=function(){return(0,o.useFetchFieldPropertySections)(),a.default.createElement(u.FieldListWrapper,null,a.default.createElement(l.Search,null),a.default.createElement(i.FieldGroup,{title:"Base Fields"}))}},60745:function(e,t,n){Object.defineProperty(t,"__esModule",{value:!0}),t.useSelectSearchedFields=void 0;var r=n(67294),a=n(47785),o=n(65673);t.useSelectSearchedFields=function(){var e=(0,a.useSelector)((0,o.selectQuery)(o.Search.Fields));return(0,r.useCallback)((function(t){return e?t.filter((function(t){return t.name.toLowerCase().includes(e)})):t}),[e])}},66528:function(e,t,n){var r=this&&this.__makeTemplateObject||function(e,t){return Object.defineProperty?Object.defineProperty(e,"raw",{value:t}):e.raw=t,e},a=this&&this.__importDefault||function(e){return e&&e.__esModule?e:{default:e}};Object.defineProperty(t,"__esModule",{value:!0}),t.ItemCheckbox=t.Item=t.Heading=t.DropDownWrapper=t.FilterIcon=void 0;var o,i,l,u,d,c=n(240),s=a(n(12788)),f=n(60755);t.FilterIcon=s.default.button(o||(o=r(["\n  cursor: pointer;\n  right: 1px;\n\n  ","\n\n  color: ",";\n  transition: all 0.2s ease-out;\n\n  background: white;\n\n  &:before {\n    content: '';\n\n    position: absolute;\n    left: 0;\n    top: 6px;\n\n    display: block;\n\n    width: 1px;\n    height: 22px;\n\n    background: ",";\n    transition: all 0.2s ease-out;\n  }\n\n  &:focus {\n    outline: none;\n  }\n\n  &:hover:before {\n    top: 0;\n    height: 32px;\n  }\n\n  &.active {\n    background: ",";\n    outline: none;\n\n    &:before {\n      top: 0;\n      height: 32px;\n    }\n  }\n\n  > svg {\n    user-select: none;\n    pointer-events: none;\n  }\n"],["\n  cursor: pointer;\n  right: 1px;\n\n  ","\n\n  color: ",";\n  transition: all 0.2s ease-out;\n\n  background: white;\n\n  &:before {\n    content: '';\n\n    position: absolute;\n    left: 0;\n    top: 6px;\n\n    display: block;\n\n    width: 1px;\n    height: 22px;\n\n    background: ",";\n    transition: all 0.2s ease-out;\n  }\n\n  &:focus {\n    outline: none;\n  }\n\n  &:hover:before {\n    top: 0;\n    height: 32px;\n  }\n\n  &.active {\n    background: ",";\n    outline: none;\n\n    &:before {\n      top: 0;\n      height: 32px;\n    }\n  }\n\n  > svg {\n    user-select: none;\n    pointer-events: none;\n  }\n"])),f.IconStyle,c.colors.gray300,c.colors.gray200,c.colors.gray050),t.DropDownWrapper=s.default.div(i||(i=r(["\n  position: absolute;\n  top: 32px;\n  right: -1px;\n\n  //display: none;\n\n  padding: "," ",";\n  white-space: nowrap;\n  text-align: left;\n\n  border: 1px solid ",";\n  background: ",";\n  box-shadow: rgba(255, 255, 255, 0.1) 0px 1px 1px 0px inset,\n    rgba(50, 50, 93, 0.25) 0px 50px 100px -20px,\n    rgba(0, 0, 0, 0.3) 0px 30px 60px -30px;\n\n  opacity: 0;\n  transform-origin: top right;\n  transform: scaleY(0);\n  transition: all 0.2s ease-out;\n\n  ",".active & {\n    display: block;\n    background: ",";\n\n    opacity: 1;\n    transform: scaleY(1);\n  }\n\n  &:before {\n    content: '';\n\n    position: absolute;\n    top: -1px;\n    right: 0px;\n\n    width: 33px;\n    height: 1px;\n\n    background: ",";\n  }\n"],["\n  position: absolute;\n  top: 32px;\n  right: -1px;\n\n  //display: none;\n\n  padding: "," ",";\n  white-space: nowrap;\n  text-align: left;\n\n  border: 1px solid ",";\n  background: ",";\n  box-shadow: rgba(255, 255, 255, 0.1) 0px 1px 1px 0px inset,\n    rgba(50, 50, 93, 0.25) 0px 50px 100px -20px,\n    rgba(0, 0, 0, 0.3) 0px 30px 60px -30px;\n\n  opacity: 0;\n  transform-origin: top right;\n  transform: scaleY(0);\n  transition: all 0.2s ease-out;\n\n  ",".active & {\n    display: block;\n    background: ",";\n\n    opacity: 1;\n    transform: scaleY(1);\n  }\n\n  &:before {\n    content: '';\n\n    position: absolute;\n    top: -1px;\n    right: 0px;\n\n    width: 33px;\n    height: 1px;\n\n    background: ",";\n  }\n"])),c.spacings.sm,c.spacings.md,c.colors.gray200,c.colors.white,t.FilterIcon,c.colors.gray050,c.colors.gray050),t.Heading=s.default.h3(l||(l=r(["\n  margin: 0;\n  padding: 0 0 ",";\n"],["\n  margin: 0;\n  padding: 0 0 ",";\n"])),c.spacings.xs),t.Item=s.default.li(u||(u=r(["\n  display: flex;\n  gap: ",";\n  text-align: left;\n"],["\n  display: flex;\n  gap: ",";\n  text-align: left;\n"])),c.spacings.sm),t.ItemCheckbox=s.default.input.attrs({type:"checkbox"})(d||(d=r([""],[""])))},14428:function(e,t,n){var r=this&&this.__createBinding||(Object.create?function(e,t,n,r){void 0===r&&(r=n);var a=Object.getOwnPropertyDescriptor(t,n);a&&!("get"in a?!t.__esModule:a.writable||a.configurable)||(a={enumerable:!0,get:function(){return t[n]}}),Object.defineProperty(e,r,a)}:function(e,t,n,r){void 0===r&&(r=n),e[r]=t[n]}),a=this&&this.__setModuleDefault||(Object.create?function(e,t){Object.defineProperty(e,"default",{enumerable:!0,value:t})}:function(e,t){e.default=t}),o=this&&this.__importStar||function(e){if(e&&e.__esModule)return e;var t={};if(null!=e)for(var n in e)"default"!==n&&Object.prototype.hasOwnProperty.call(e,n)&&r(t,e,n);return a(t,e),t},i=this&&this.__importDefault||function(e){return e&&e.__esModule?e:{default:e}};Object.defineProperty(t,"__esModule",{value:!0}),t.Filter=void 0;var l=o(n(67294)),u=n(32523),d=i(n(87351)),c=i(n(56137)),s=n(66528),f=["Favorites","Standard Fields","Form 1 fields","Form 2 fields","Form 3 fields"];t.Filter=function(){var e=(0,l.useState)(!1),t=e[0],n=e[1],r=(0,u.useClickOutside)((function(){n(!1)}),t);return l.default.createElement(l.default.Fragment,null,l.default.createElement(s.FilterIcon,{ref:r,className:t&&"active",onClick:function(e){t&&e.target===r.current&&n(!1),t||n(!0)}},l.default.createElement(c.default,null),l.default.createElement(s.DropDownWrapper,null,l.default.createElement(s.Heading,null,(0,d.default)("Search in")),l.default.createElement("ul",null,f.map((function(e){return l.default.createElement(s.Item,{key:e},l.default.createElement(s.ItemCheckbox,null),e)}))))))}},23636:function(e,t,n){Object.defineProperty(t,"__esModule",{value:!0}),t.useFieldSearch=void 0;var r=n(67294),a=n(47785),o=n(65673),i=n(25818);t.useFieldSearch=function(){var e=(0,a.useDispatch)(),t=(0,r.useState)(""),n=t[0],l=t[1],u=(0,i.useDebounce)(n,1e3);return(0,r.useEffect)((function(){e((0,o.updateQuery)({type:o.Search.Fields,query:u}))}),[u]),[n,l]}},60755:function(e,t,n){var r=this&&this.__makeTemplateObject||function(e,t){return Object.defineProperty?Object.defineProperty(e,"raw",{value:t}):e.raw=t,e},a=this&&this.__createBinding||(Object.create?function(e,t,n,r){void 0===r&&(r=n);var a=Object.getOwnPropertyDescriptor(t,n);a&&!("get"in a?!t.__esModule:a.writable||a.configurable)||(a={enumerable:!0,get:function(){return t[n]}}),Object.defineProperty(e,r,a)}:function(e,t,n,r){void 0===r&&(r=n),e[r]=t[n]}),o=this&&this.__setModuleDefault||(Object.create?function(e,t){Object.defineProperty(e,"default",{enumerable:!0,value:t})}:function(e,t){e.default=t}),i=this&&this.__importStar||function(e){if(e&&e.__esModule)return e;var t={};if(null!=e)for(var n in e)"default"!==n&&Object.prototype.hasOwnProperty.call(e,n)&&a(t,e,n);return o(t,e),t};Object.defineProperty(t,"__esModule",{value:!0}),t.SearchIcon=t.IconStyle=t.SearchBar=t.SearchBlock=t.Wrapper=void 0;var l=n(240),u=i(n(12788));t.Wrapper=u.default.div(d||(d=r(["\n  position: relative;\n  z-index: 1;\n\n  margin-bottom: ",";\n"],["\n  position: relative;\n  z-index: 1;\n\n  margin-bottom: ",";\n"])),l.spacings.lg),t.SearchBlock=u.default.div(c||(c=r(["\n  display: flex;\n"],["\n  display: flex;\n"]))),t.SearchBar=u.default.input(s||(s=r(["\n  padding: 6px 38px 6px 30px !important;\n\n  &::placeholder {\n    font-style: italic;\n    color: ",";\n  }\n"],["\n  padding: 6px 38px 6px 30px !important;\n\n  &::placeholder {\n    font-style: italic;\n    color: ",";\n  }\n"])),l.colors.gray200);var d,c,s,f,p,v="18px";t.IconStyle=(0,u.css)(f||(f=r(["\n  position: absolute;\n  top: 1px;\n  bottom: 1px;\n  z-index: 2;\n\n  display: flex;\n  flex-direction: column;\n  justify-content: center;\n\n  padding: 0 8px;\n\n  box-sizing: border-box;\n  user-select: none;\n\n  > svg {\n    width: ",";\n    height: ",";\n  }\n"],["\n  position: absolute;\n  top: 1px;\n  bottom: 1px;\n  z-index: 2;\n\n  display: flex;\n  flex-direction: column;\n  justify-content: center;\n\n  padding: 0 8px;\n\n  box-sizing: border-box;\n  user-select: none;\n\n  > svg {\n    width: ",";\n    height: ",";\n  }\n"])),v,v),t.SearchIcon=u.default.div(p||(p=r(["\n  left: 1px;\n\n  ","\n\n  color: ",";\n"],["\n  left: 1px;\n\n  ","\n\n  color: ",";\n"])),t.IconStyle,l.colors.gray100)},65234:function(e,t,n){var r=this&&this.__importDefault||function(e){return e&&e.__esModule?e:{default:e}};Object.defineProperty(t,"__esModule",{value:!0}),t.Search=void 0;var a=r(n(67294)),o=r(n(87351)),i=n(14428),l=n(23636),u=n(60755),d=r(n(65721));t.Search=function(){var e=(0,l.useFieldSearch)(),t=e[0],n=e[1];return a.default.createElement(u.Wrapper,null,a.default.createElement(u.SearchBlock,null,a.default.createElement(u.SearchIcon,null,a.default.createElement(d.default,null)),a.default.createElement(i.Filter,null),a.default.createElement(u.SearchBar,{type:"text",placeholder:(0,o.default)("Search fields..."),className:"fullwidth text",value:t,onChange:function(e){n(e.target.value)}})))}},67088:function(e,t,n){var r=this&&this.__makeTemplateObject||function(e,t){return Object.defineProperty?Object.defineProperty(e,"raw",{value:t}):e.raw=t,e},a=this&&this.__importDefault||function(e){return e&&e.__esModule?e:{default:e}};Object.defineProperty(t,"__esModule",{value:!0}),t.Grid=void 0;var o,i=a(n(12788));t.Grid=i.default.div(o||(o=r(["\n  position: relative;\n  display: flex;\n  gap: 0;\n\n  height: 100%;\n  background: #fff;\n"],["\n  position: relative;\n  display: flex;\n  gap: 0;\n\n  height: 100%;\n  background: #fff;\n"])))},86419:function(e,t,n){var r=this&&this.__importDefault||function(e){return e&&e.__esModule?e:{default:e}};Object.defineProperty(t,"__esModule",{value:!0}),t.LayoutEditor=void 0;var a=r(n(67294)),o=n(47785),i=n(4524),l=n(45323),u=n(19149),d=n(18842),c=n(24128),s=n(67088);t.LayoutEditor=function(){var e=(0,o.useSelector)(l.selectFocus).active;return a.default.createElement(s.Grid,null,a.default.createElement(i.SidebarSlider,{swiped:e},a.default.createElement(c.PropertyEditor,null),a.default.createElement(d.FieldList,null)),a.default.createElement(u.FieldLayout,null))}},91198:function(e,t,n){var r,a=this&&this.__extends||(r=function(e,t){return r=Object.setPrototypeOf||{__proto__:[]}instanceof Array&&function(e,t){e.__proto__=t}||function(e,t){for(var n in t)Object.prototype.hasOwnProperty.call(t,n)&&(e[n]=t[n])},r(e,t)},function(e,t){if("function"!=typeof t&&null!==t)throw new TypeError("Class extends value "+String(t)+" is not a constructor or null");function n(){this.constructor=e}r(e,t),e.prototype=null===t?Object.create(t):(n.prototype=t.prototype,new n)}),o=this&&this.__importDefault||function(e){return e&&e.__esModule?e:{default:e}};Object.defineProperty(t,"__esModule",{value:!0}),t.ErrorBoundary=void 0;var i=o(n(67294)),l=function(e){function t(t){var n=e.call(this,t)||this;return n.state={hasError:!1},n}return a(t,e),t.getDerivedStateFromError=function(){return{hasError:!0}},t.prototype.componentDidCatch=function(e,t){console.log(e,t)},t.prototype.render=function(){return this.state.hasError?i.default.createElement("div",null,this.props.message):this.props.children},t}(i.default.Component);t.ErrorBoundary=l},65556:function(e,t,n){var r=this&&this.__createBinding||(Object.create?function(e,t,n,r){void 0===r&&(r=n);var a=Object.getOwnPropertyDescriptor(t,n);a&&!("get"in a?!t.__esModule:a.writable||a.configurable)||(a={enumerable:!0,get:function(){return t[n]}}),Object.defineProperty(e,r,a)}:function(e,t,n,r){void 0===r&&(r=n),e[r]=t[n]}),a=this&&this.__setModuleDefault||(Object.create?function(e,t){Object.defineProperty(e,"default",{enumerable:!0,value:t})}:function(e,t){e.default=t}),o=this&&this.__importStar||function(e){if(e&&e.__esModule)return e;var t={};if(null!=e)for(var n in e)"default"!==n&&Object.prototype.hasOwnProperty.call(e,n)&&r(t,e,n);return a(t,e),t};Object.defineProperty(t,"__esModule",{value:!0}),t.EditableComponent=void 0;var i=o(n(67294)),l=n(85767),u=n(91198),d=o(n(84747));t.EditableComponent=function(e){var t=e.field,n=e.property,r=(0,l.useAppDispatch)(),a=d[n.type];return void 0===a?i.default.createElement("div",null,"...".concat(n.handle," <").concat(n.type,">")):(a.displayName="Field <".concat(n.type,">"),i.default.createElement(u.ErrorBoundary,{message:"...".concat(n.handle," <").concat(n.type,">")},i.default.createElement(i.Suspense,null,i.default.createElement(a,{field:t,property:n,dispatch:r}))))}},67230:function(e,t,n){var r=this&&this.__importDefault||function(e){return e&&e.__esModule?e:{default:e}};Object.defineProperty(t,"__esModule",{value:!0}),t.FieldProperties=void 0;var a=r(n(67294)),o=r(n(84849)),i=n(47785),l=n(56859),u=n(42367),d=n(24741),c=n(65556);t.FieldProperties=function(e){var t=e.uid,n=(0,u.useFetchFieldPropertySections)(),r=n.data,s=n.isFetching,f=(0,i.useSelector)((0,l.selectField)(t)),p=(0,u.useFieldType)(null==f?void 0:f.typeClass);return f&&p?!r&&s?a.default.createElement("div",null,a.default.createElement("h2",null,p.name),a.default.createElement(d.SectionWrapper,null,a.default.createElement(o.default,null))):a.default.createElement(a.default.Fragment,null,a.default.createElement(d.Title,null,a.default.createElement(d.Icon,{dangerouslySetInnerHTML:{__html:p.icon}}),a.default.createElement("span",null,p.name)),a.default.createElement(d.SectionWrapper,null,r.map((function(e){var t=e.handle,n=e.label;return a.default.createElement(d.SectionBlock,{label:n,key:t},p.properties.filter((function(e){return e.section===t})).sort((function(e,t){return e.order-t.order})).map((function(e){return a.default.createElement(c.EditableComponent,{key:e.handle,field:f,property:e})})))})))):a.default.createElement("div",null,"Not found")}},48170:function(e,t,n){var r=this&&this.__makeTemplateObject||function(e,t){return Object.defineProperty?Object.defineProperty(e,"raw",{value:t}):e.raw=t,e},a=this&&this.__importDefault||function(e){return e&&e.__esModule?e:{default:e}};Object.defineProperty(t,"__esModule",{value:!0}),t.CheckboxWrapper=void 0;var o,i=n(240),l=a(n(12788));t.CheckboxWrapper=l.default.div(o||(o=r(["\n  display: flex;\n  gap: ",";\n"],["\n  display: flex;\n  gap: ",";\n"])),i.spacings.sm)},26902:function(e,t,n){var r=this&&this.__importDefault||function(e){return e&&e.__esModule?e:{default:e}};Object.defineProperty(t,"__esModule",{value:!0});var a=r(n(67294)),o=n(3399),i=n(56859),l=n(48170),u=n(39452);t.default=function(e){var t,n=e.field,r=e.property,d=e.dispatch,c=r.handle,s=r.label,f=n.uid,p=null!==(t=n.properties[c])&&void 0!==t&&t;return a.default.createElement(u.ControlWrapper,null,a.default.createElement(l.CheckboxWrapper,null,a.default.createElement(o.CheckboxInput,{id:c,label:s,onClick:function(){return d((0,i.edit)({uid:f,property:r,value:!p}))}})))}},39452:function(e,t,n){var r=this&&this.__makeTemplateObject||function(e,t){return Object.defineProperty?Object.defineProperty(e,"raw",{value:t}):e.raw=t,e},a=this&&this.__importDefault||function(e){return e&&e.__esModule?e:{default:e}};Object.defineProperty(t,"__esModule",{value:!0}),t.Label=t.ControlWrapper=void 0;var o,i,l=a(n(12788));t.ControlWrapper=l.default.div(o||(o=r([""],[""]))),t.Label=l.default.label(i||(i=r(["\n  display: block;\n  font-weight: ",";\n"],["\n  display: block;\n  font-weight: ",";\n"])),(function(e){return e.regular?"normal":"bold"}))},67569:function(e,t,n){var r=this&&this.__importDefault||function(e){return e&&e.__esModule?e:{default:e}};Object.defineProperty(t,"__esModule",{value:!0}),t.Control=void 0;var a=r(n(67294)),o=n(39452);t.Control=function(e){var t=e.children,n=e.property;return a.default.createElement(o.ControlWrapper,null,a.default.createElement(o.Label,{htmlFor:n.handle},n.label),t)}},84747:function(e,t,n){var r=this&&this.__importDefault||function(e){return e&&e.__esModule?e:{default:e}};Object.defineProperty(t,"__esModule",{value:!0}),t.textarea=t.string=t.int=t.bool=void 0;var a=n(26902);Object.defineProperty(t,"bool",{enumerable:!0,get:function(){return r(a).default}});var o=n(36497);Object.defineProperty(t,"int",{enumerable:!0,get:function(){return r(o).default}});var i=n(20551);Object.defineProperty(t,"string",{enumerable:!0,get:function(){return r(i).default}});var l=n(53705);Object.defineProperty(t,"textarea",{enumerable:!0,get:function(){return r(l).default}})},36497:function(e,t,n){var r=this&&this.__importDefault||function(e){return e&&e.__esModule?e:{default:e}};Object.defineProperty(t,"__esModule",{value:!0});var a=r(n(67294)),o=n(56859),i=n(67569);t.default=function(e){var t,n=e.field,r=e.property,l=e.dispatch,u=r.handle,d=n.uid;return a.default.createElement(i.Control,{property:r},a.default.createElement("input",{id:u,type:"number",className:"text fullwidth",value:null!==(t=n.properties[u])&&void 0!==t?t:"",onChange:function(e){return l((0,o.edit)({uid:d,property:r,value:Number(e.target.value)}))}}))}},20551:function(e,t,n){var r=this&&this.__importDefault||function(e){return e&&e.__esModule?e:{default:e}};Object.defineProperty(t,"__esModule",{value:!0});var a=r(n(67294)),o=n(56859),i=r(n(89361)),l=n(67569);t.default=function(e){var t,n=e.field,r=e.property,u=e.dispatch,d=r.handle,c=n.uid,s=n.properties;return a.default.createElement(l.Control,{property:r},a.default.createElement("input",{id:d,type:"text",className:(0,i.default)("text","fullwidth",r.flags.includes("code")&&"code"),value:null!==(t=s[d])&&void 0!==t?t:"",placeholder:r.placeholder,onChange:function(e){return u((0,o.edit)({uid:c,property:r,value:e.target.value}))}}))}},53705:function(e,t,n){var r=this&&this.__importDefault||function(e){return e&&e.__esModule?e:{default:e}};Object.defineProperty(t,"__esModule",{value:!0});var a=r(n(67294)),o=n(56859),i=r(n(89361)),l=n(67569);t.default=function(e){var t,n=e.field,r=e.property,u=e.dispatch,d=r.handle,c=n.uid,s=n.properties;return a.default.createElement(l.Control,{property:r},a.default.createElement("textarea",{id:d,className:(0,i.default)("text","fullwidth",r.flags.includes("code")&&"code"),rows:2,value:null!==(t=s[d])&&void 0!==t?t:"",placeholder:r.placeholder,onChange:function(e){return u((0,o.edit)({uid:c,property:r,value:e.target.value}))}}))}},24741:function(e,t,n){var r=this&&this.__makeTemplateObject||function(e,t){return Object.defineProperty?Object.defineProperty(e,"raw",{value:t}):e.raw=t,e},a=this&&this.__importDefault||function(e){return e&&e.__esModule?e:{default:e}};Object.defineProperty(t,"__esModule",{value:!0}),t.SectionBlock=t.SectionWrapper=t.Icon=t.Title=t.CloseLink=t.PropertyEditorWrapper=void 0;var o,i,l,u,d,c,s=n(240),f=a(n(12788));t.PropertyEditorWrapper=f.default.div(o||(o=r(["\n  position: relative;\n  padding: ",";\n"],["\n  position: relative;\n  padding: ",";\n"])),s.spacings.lg),t.CloseLink=f.default.a(i||(i=r(["\n  position: absolute;\n  right: 10px;\n  top: 17px;\n\n  display: block;\n  width: 20px;\n  height: 20px;\n"],["\n  position: absolute;\n  right: 10px;\n  top: 17px;\n\n  display: block;\n  width: 20px;\n  height: 20px;\n"]))),t.Title=f.default.h3(l||(l=r(["\n  display: flex;\n  justify-content: flex-start;\n  align-items: center;\n  gap: ",";\n\n  margin: 0;\n  font-size: 16px;\n"],["\n  display: flex;\n  justify-content: flex-start;\n  align-items: center;\n  gap: ",";\n\n  margin: 0;\n  font-size: 16px;\n"])),s.spacings.sm),t.Icon=f.default.div(u||(u=r(["\n  width: 20px;\n  height: 20px;\n"],["\n  width: 20px;\n  height: 20px;\n"]))),t.SectionWrapper=f.default.div(d||(d=r(["\n  display: flex;\n  flex-direction: column;\n  gap: ",";\n"],["\n  display: flex;\n  flex-direction: column;\n  gap: ",";\n"])),s.spacings.lg),t.SectionBlock=f.default.section(c||(c=r(["\n  position: relative;\n\n  display: flex;\n  flex-direction: column;\n  gap: ",";\n\n  margin-top: ",";\n  padding-top: ",";\n\n  &:empty {\n    display: none;\n  }\n\n  &:before {\n    content: '';\n\n    position: absolute;\n    left: 0;\n    top: 0;\n    right: 0;\n\n    display: block;\n    height: 1px;\n\n    margin: 0 -18px;\n\n    box-shadow: ",";\n  }\n\n  &:after {\n    content: '","';\n\n    position: absolute;\n    left: -5px;\n    top: -7px;\n\n    display: block;\n    padding: 0 5px;\n\n    background-color: ",";\n\n    color: rgb(154 165 177 / 75%);\n    font-family: system-ui, BlinkMacSystemFont, -apple-system, Segoe UI, Roboto,\n      Oxygen, Ubuntu, Cantarell, Fira Sans, Droid Sans, Helvetica Neue,\n      sans-serif;\n    font-size: 11px;\n    font-weight: bold;\n    text-transform: uppercase;\n  }\n"],["\n  position: relative;\n\n  display: flex;\n  flex-direction: column;\n  gap: ",";\n\n  margin-top: ",";\n  padding-top: ",";\n\n  &:empty {\n    display: none;\n  }\n\n  &:before {\n    content: '';\n\n    position: absolute;\n    left: 0;\n    top: 0;\n    right: 0;\n\n    display: block;\n    height: 1px;\n\n    margin: 0 -18px;\n\n    box-shadow: ",";\n  }\n\n  &:after {\n    content: '","';\n\n    position: absolute;\n    left: -5px;\n    top: -7px;\n\n    display: block;\n    padding: 0 5px;\n\n    background-color: ",";\n\n    color: rgb(154 165 177 / 75%);\n    font-family: system-ui, BlinkMacSystemFont, -apple-system, Segoe UI, Roboto,\n      Oxygen, Ubuntu, Cantarell, Fira Sans, Droid Sans, Helvetica Neue,\n      sans-serif;\n    font-size: 11px;\n    font-weight: bold;\n    text-transform: uppercase;\n  }\n"])),s.spacings.md,s.spacings.lg,s.spacings.lg,s.shadows.bottom,(function(e){return e.label}),s.colors.gray050)},24128:function(e,t,n){var r=this&&this.__createBinding||(Object.create?function(e,t,n,r){void 0===r&&(r=n);var a=Object.getOwnPropertyDescriptor(t,n);a&&!("get"in a?!t.__esModule:a.writable||a.configurable)||(a={enumerable:!0,get:function(){return t[n]}}),Object.defineProperty(e,r,a)}:function(e,t,n,r){void 0===r&&(r=n),e[r]=t[n]}),a=this&&this.__setModuleDefault||(Object.create?function(e,t){Object.defineProperty(e,"default",{enumerable:!0,value:t})}:function(e,t){e.default=t}),o=this&&this.__importStar||function(e){if(e&&e.__esModule)return e;var t={};if(null!=e)for(var n in e)"default"!==n&&Object.prototype.hasOwnProperty.call(e,n)&&r(t,e,n);return a(t,e),t},i=this&&this.__importDefault||function(e){return e&&e.__esModule?e:{default:e}};Object.defineProperty(t,"__esModule",{value:!0}),t.PropertyEditor=void 0;var l=o(n(67294)),u=n(47785),d=n(85767),c=n(45323),s=i(n(33742)),f=n(12020),p=n(91198),v=n(67230),h=n(24741);t.PropertyEditor=function(){var e=(0,u.useSelector)(c.selectFocus),t=e.active,n=e.type,r=e.uid,a=(0,d.useAppDispatch)(),o=(0,l.useRef)(null);(0,f.useOnKeypress)({meetsCondition:t,callback:function(e){"Escape"===e.key&&a((0,c.unfocus)())}});var i=l.default.createElement(v.FieldProperties,{uid:r});return l.default.createElement(h.PropertyEditorWrapper,{ref:o},l.default.createElement(h.CloseLink,{onClick:function(){return a((0,c.unfocus)())}},l.default.createElement(s.default,null)),l.default.createElement(p.ErrorBoundary,{message:'Could not load property editor for "'.concat(n,'" type')},i))}},11037:function(e,t){Object.defineProperty(t,"__esModule",{value:!0}),t.CellType=void 0,function(e){e.Field="field",e.Layout="layout"}(t.CellType||(t.CellType={}))},70774:function(e,t,n){var r=this&&this.__makeTemplateObject||function(e,t){return Object.defineProperty?Object.defineProperty(e,"raw",{value:t}):e.raw=t,e};Object.defineProperty(t,"__esModule",{value:!0}),t.EditorGlobalStyles=void 0;var a,o=n(12788);t.EditorGlobalStyles=(0,o.createGlobalStyle)(a||(a=r(["\n  #global-header, #global-footer {\n    display: none;\n  }\n\n  #main-container {\n    #main {\n      #main-content {\n        padding: 0;\n        width: calc(100vw - 50px);\n      }\n    }\n  }\n\n  #global-sidebar {\n    width: 50px;\n    overflow: hidden;\n\n    #system-name {\n      display: none;\n    }\n\n    nav#nav {\n      margin: 8px 0 0;\n\n      > ul > li {\n        > a {\n          padding: 8px 8px;\n        }\n\n        .label {\n          display: none;\n        }\n\n        .subnav {\n          display: none;\n        }\n      }\n    }\n  }\n\n  #page-container {\n    padding-left: 50px !important;\n  }\n\n  #freeform-client-app {\n    height: 100vh;\n  }\n"],["\n  #global-header, #global-footer {\n    display: none;\n  }\n\n  #main-container {\n    #main {\n      #main-content {\n        padding: 0;\n        width: calc(100vw - 50px);\n      }\n    }\n  }\n\n  #global-sidebar {\n    width: 50px;\n    overflow: hidden;\n\n    #system-name {\n      display: none;\n    }\n\n    nav#nav {\n      margin: 8px 0 0;\n\n      > ul > li {\n        > a {\n          padding: 8px 8px;\n        }\n\n        .label {\n          display: none;\n        }\n\n        .subnav {\n          display: none;\n        }\n      }\n    }\n  }\n\n  #page-container {\n    padding-left: 50px !important;\n  }\n\n  #freeform-client-app {\n    height: 100vh;\n  }\n"])))},66444:function(e,t,n){var r=this&&this.__importDefault||function(e){return e&&e.__esModule?e:{default:e}};Object.defineProperty(t,"__esModule",{value:!0}),t.Edit=void 0;var a=r(n(67294)),o=n(49818),i=n(23493),l=n(93787);t.Edit=function(){var e=(0,o.useParams)().formId;(0,i.useQueryFormSettings)();var t=(0,i.useQuerySingleForm)(e&&Number(e)),n=t.isFetching,r=t.isError,u=t.error;return n?a.default.createElement("div",null,"Fetching ",e,"..."):r?a.default.createElement("div",null,"ERROR: ",u.message):a.default.createElement(l.Builder,null)}},33922:function(e,t,n){var r=this&&this.__importDefault||function(e){return e&&e.__esModule?e:{default:e}};Object.defineProperty(t,"__esModule",{value:!0}),t.EditProvider=void 0;var a=r(n(67294)),o=n(47785),i=n(66444),l=n(70774),u=n(85767);t.EditProvider=function(){return a.default.createElement(o.Provider,{store:u.store},a.default.createElement(l.EditorGlobalStyles,null),a.default.createElement(i.Edit,null))}},80926:function(e,t,n){Object.defineProperty(t,"__esModule",{value:!0}),t.save=void 0;var r=n(20991);t.save=(0,r.createAction)("form/save")},85767:function(e,t,n){var r=this&&this.__importDefault||function(e){return e&&e.__esModule?e:{default:e}};Object.defineProperty(t,"__esModule",{value:!0}),t.useAppSelector=t.useAppDispatch=t.store=void 0;var a=n(47785),o=n(20991),i=r(n(53894)),l=n(90531),u=r(n(89372)),d=r(n(45323)),c=r(n(99614)),s=r(n(56859)),f=r(n(61552)),p=r(n(68887)),v=r(n(1965)),h=r(n(68974)),m=r(n(33314)),g=r(n(65673));t.store=(0,o.configureStore)({middleware:[i.default,l.statePersistMiddleware],reducer:{form:f.default,fields:s.default,context:d.default,integrations:p.default,layouts:v.default,pages:h.default,rows:m.default,cells:u.default,search:g.default,drag:c.default}});t.useAppDispatch=function(){return(0,a.useDispatch)()},t.useAppSelector=a.useSelector},90531:function(e,t,n){var r=this&&this.__importDefault||function(e){return e&&e.__esModule?e:{default:e}};Object.defineProperty(t,"__esModule",{value:!0}),t.statePersistMiddleware=t.TOPIC_UPDATED=t.TOPIC_CREATED=t.TOPIC_ERRORS=t.TOPIC_SAVE=void 0;var a=r(n(9669)),o=r(n(1798)),i=n(80926);t.TOPIC_SAVE=Symbol("form.save"),t.TOPIC_ERRORS=Symbol("form.save.errors"),t.TOPIC_CREATED=Symbol("form.save.crated"),t.TOPIC_UPDATED=Symbol("form.save.updated"),o.default.clearAllSubscriptions();var l=function(e,n){o.default.publish(t.TOPIC_ERRORS,{dispatch:e,response:n})};t.statePersistMiddleware=function(e){return function(n){return function(r){if(r&&(n(r),r.type===String(i.save))){var u=e.dispatch,d={dispatch:u,state:e.getState(),persist:{}};o.default.publishSync(t.TOPIC_SAVE,d);var c=d.state.form.id;c?a.default.put("/client/api/forms/".concat(c),d.persist).then((function(e){return function(e,n){o.default.publish(t.TOPIC_UPDATED,{dispatch:e,response:n})}(u,e)})).catch((function(e){return l(u,e)})):a.default.post("/client/api/forms",d.persist).then((function(e){return function(e,n){o.default.publish(t.TOPIC_CREATED,{dispatch:e,response:n})}(u,e)})).catch((function(e){return l(u,e)}))}}}}},89372:function(e,t,n){var r,a=this&&this.__assign||function(){return a=Object.assign||function(e){for(var t,n=1,r=arguments.length;n<r;n++)for(var a in t=arguments[n])Object.prototype.hasOwnProperty.call(t,a)&&(e[a]=t[a]);return e},a.apply(this,arguments)},o=this&&this.__spreadArray||function(e,t,n){if(n||2===arguments.length)for(var r,a=0,o=t.length;a<o;a++)!r&&a in t||(r||(r=Array.prototype.slice.call(t,0,a)),r[a]=t[a]);return e.concat(r||Array.prototype.slice.call(t))};Object.defineProperty(t,"__esModule",{value:!0}),t.selectCellsInRow=t.remove=t.add=t.moveTo=t.set=t.cellsSlice=void 0;var i=n(20991);t.cellsSlice=(0,i.createSlice)({name:"cells",initialState:[],reducers:{set:function(e,t){e.splice.apply(e,o([0,e.length],t.payload,!1))},add:function(e,t){var n=Math.max.apply(Math,o([-1],e.filter((function(e){return e.rowUid===t.payload.rowUid})).map((function(e){return e.order})),!1));e.push(a(a({},t.payload),{order:n+1}))},remove:function(e,t){e=e.filter((function(e){return e.uid!==t.payload}))},moveTo:function(e,t){console.log(t)}}}),t.set=(r=t.cellsSlice.actions).set,t.moveTo=r.moveTo,t.add=r.add,t.remove=r.remove;t.selectCellsInRow=function(e){return function(t){return t.cells.filter((function(t){return t.rowUid===e.uid})).sort((function(e,t){return e.order-t.order}))}},t.default=t.cellsSlice.reducer},45323:function(e,t,n){var r,a=this&&this.__assign||function(){return a=Object.assign||function(e){for(var t,n=1,r=arguments.length;n<r;n++)for(var a in t=arguments[n])Object.prototype.hasOwnProperty.call(t,a)&&(e[a]=t[a]);return e},a.apply(this,arguments)};Object.defineProperty(t,"__esModule",{value:!0}),t.selectFocus=t.selectCurrentPage=t.unfocus=t.setFocusedItem=t.setPage=t.FocusType=void 0;var o=n(20991);!function(e){e.Page="page",e.Field="field",e.Row="row",e.Cell="cell"}(t.FocusType||(t.FocusType={}));var i=(0,o.createSlice)({name:"context",initialState:{page:null,focus:{active:!1,type:null,uid:null}},reducers:{setPage:function(e,t){var n=t.payload;e.page=n},setFocusedItem:function(e,t){var n=t.payload;e.focus=a({active:!0},n)},focus:function(e){e.focus.active=!0},unfocus:function(e){e.focus.active=!1}}});t.setPage=(r=i.actions).setPage,t.setFocusedItem=r.setFocusedItem,t.unfocus=r.unfocus;t.selectCurrentPage=function(e){var t=e.context.page;return t?e.pages.find((function(e){return e.uid===t})):e.pages.find(Boolean)};t.selectFocus=function(e){return e.context.focus},t.default=i.reducer},99614:function(e,t,n){var r;Object.defineProperty(t,"__esModule",{value:!0}),t.selectCurrentPosition=t.selectCurrentCell=t.selectCurrentRow=t.setPosition=t.setCell=t.setRow=t.dragSlice=void 0;var a=n(20991),o={cellUid:void 0,rowUid:void 0,position:void 0};t.dragSlice=(0,a.createSlice)({name:"drag",initialState:o,reducers:{setRow:function(e,t){e.rowUid=t.payload},setCell:function(e,t){e.cellUid=t.payload},setPosition:function(e,t){e.position=t.payload}}}),t.setRow=(r=t.dragSlice.actions).setRow,t.setCell=r.setCell,t.setPosition=r.setPosition;t.selectCurrentRow=function(e){return e.drag.rowUid};t.selectCurrentCell=function(e){return e.drag.cellUid};t.selectCurrentPosition=function(e){return e.drag.position},t.default=t.dragSlice.reducer},56859:function(e,t,n){var r,a=this&&this.__spreadArray||function(e,t,n){if(n||2===arguments.length)for(var r,a=0,o=t.length;a<o;a++)!r&&a in t||(r||(r=Array.prototype.slice.call(t,0,a)),r[a]=t[a]);return e.concat(r||Array.prototype.slice.call(t))},o=this&&this.__importDefault||function(e){return e&&e.__esModule?e:{default:e}};Object.defineProperty(t,"__esModule",{value:!0}),t.selectField=t.edit=t.remove=t.add=t.set=t.fieldsSlice=void 0;var i=n(90531),l=n(20991),u=o(n(1798));t.fieldsSlice=(0,l.createSlice)({name:"fields",initialState:[],reducers:{set:function(e,t){e.splice.apply(e,a([0,e.length],t.payload,!1))},add:function(e,t){var n=t.payload,r=n.uid,a=n.fieldType,o={};a.properties.forEach((function(e){return o[e.handle]=e.value})),e.push({uid:r,typeClass:a.typeClass,properties:o})},remove:function(e,t){var n=t.payload;e.splice(e.findIndex((function(e){return e.uid===n})),1)},edit:function(e,t){var n=t.payload,r=n.uid,a=n.property.handle,o=n.value;e.find((function(e){return e.uid===r})).properties[a]=o}}}),t.set=(r=t.fieldsSlice.actions).set,t.add=r.add,t.remove=r.remove,t.edit=r.edit;t.selectField=function(e){return function(t){return t.fields.find((function(t){return t.uid===e}))}},t.default=t.fieldsSlice.reducer;u.default.subscribe(i.TOPIC_SAVE,(function(e,t){var n=t.state;t.persist.fields=n.fields}))},61552:function(e,t,n){var r,a=this&&this.__importDefault||function(e){return e&&e.__esModule?e:{default:e}};Object.defineProperty(t,"__esModule",{value:!0}),t.selectFormSetting=t.selectFormSettings=t.selectForm=t.clearErrors=t.setErrors=t.removeError=t.addError=t.modifySettings=t.update=t.formSlice=void 0;var o=n(90531),i=n(90531),l=n(90531),u=n(20991),d=a(n(1798)),c={id:null,uid:(0,n(21614).v4)(),type:"Solspace\\Freeform\\Form\\Types\\Regular",name:"New Form",handle:"newForm",settings:{}};t.formSlice=(0,u.createSlice)({name:"form",initialState:c,reducers:{update:function(e,t){var n=t.payload;Object.assign(e,n)},modifySettings:function(e,t){var n=t.payload,r=n.namespace,a=n.key,o=n.value;e.settings[r]||(e.settings[r]={}),e.settings[r][a]=o},removeError:function(e,t){var n=t.payload;delete e.errors[n]},addError:function(e,t){var n=t.payload;e.errors[n.key]=n.errors},setErrors:function(e,t){var n=t.payload;e.errors=n},clearErrors:function(e){e.errors=void 0}}}),t.update=(r=t.formSlice.actions).update,t.modifySettings=r.modifySettings,t.addError=r.addError,t.removeError=r.removeError,t.setErrors=r.setErrors,t.clearErrors=r.clearErrors;t.selectForm=function(e){return e.form};t.selectFormSettings=function(e){return function(t){var n;return(null===(n=t.form.settings)||void 0===n?void 0:n[e])||{}}};t.selectFormSetting=function(e,t){return function(n){var r,a;return null===(a=null===(r=n.form.settings)||void 0===r?void 0:r[e])||void 0===a?void 0:a[t]}},t.default=t.formSlice.reducer;d.default.subscribe(l.TOPIC_SAVE,(function(e,t){var n=t.state,r=t.persist,a=n.form,o=a.id,i=a.uid,l=a.type,u=a.settings;r.form={id:o,uid:i,type:l,settings:u}})),d.default.subscribe(i.TOPIC_ERRORS,(function(e,n){var r,a=n.dispatch,o=n.response;a((0,t.clearErrors)()),a((0,t.setErrors)(null===(r=o.errors)||void 0===r?void 0:r.form))})),d.default.subscribe(o.TOPIC_CREATED,(function(e,n){var r=n.dispatch,a=n.response;r((0,t.update)({id:a.data.form.id}))}))},68887:function(e,t,n){var r,a=this&&this.__assign||function(){return a=Object.assign||function(e){for(var t,n=1,r=arguments.length;n<r;n++)for(var a in t=arguments[n])Object.prototype.hasOwnProperty.call(t,a)&&(e[a]=t[a]);return e},a.apply(this,arguments)};Object.defineProperty(t,"__esModule",{value:!0}),t.selectIntegration=t.modifyIntegrationSetting=t.toggleIntegration=t.addIntegrations=t.integrationsSlice=void 0;var o=n(20991),i=function(e,t){return e.find((function(e){return e.id===t}))};t.integrationsSlice=(0,o.createSlice)({name:"integrations",initialState:[],reducers:{addIntegrations:function(e,t){t.payload.forEach((function(t){e.push(a({updated:1,dirty:!1},t))}))},toggleIntegration:function(e,t){var n=i(e,t.payload);n.enabled=!n.enabled},modifyIntegrationSetting:function(e,t){var n=t.payload,r=n.id,a=n.key,o=n.value,l=function(e,t){return e.settings.find((function(e){return e.handle===t}))}(i(e,r),a);l.value=o}}}),t.addIntegrations=(r=t.integrationsSlice.actions).addIntegrations,t.toggleIntegration=r.toggleIntegration,t.modifyIntegrationSetting=r.modifyIntegrationSetting;t.selectIntegration=function(e){return function(t){return t.integrations.find((function(t){return t.id===e}))}},t.default=t.integrationsSlice.reducer},1965:function(e,t,n){var r,a=this&&this.__spreadArray||function(e,t,n){if(n||2===arguments.length)for(var r,a=0,o=t.length;a<o;a++)!r&&a in t||(r||(r=Array.prototype.slice.call(t,0,a)),r[a]=t[a]);return e.concat(r||Array.prototype.slice.call(t))};Object.defineProperty(t,"__esModule",{value:!0}),t.selectPageLayout=t.selectLayout=t.remove=t.add=t.set=t.layoutsSlice=void 0;var o=n(90531),i=n(20991);t.layoutsSlice=(0,i.createSlice)({name:"layouts",initialState:[{uid:"layout-uid-1"},{uid:"layout-uid-2"}],reducers:{set:function(e,t){e.splice.apply(e,a([0,e.length],t.payload,!1))},add:function(e,t){e.push(t.payload)},remove:function(e,t){e=e.filter((function(e){return e.uid!==t.payload}))}}}),t.set=(r=t.layoutsSlice.actions).set,t.add=r.add,t.remove=r.remove;t.selectLayout=function(e){return function(t){return t.layouts.find((function(t){return t.uid===e}))}};t.selectPageLayout=function(e){return function(t){return t.layouts.find((function(t){return t.uid===(null==e?void 0:e.layoutUid)}))}},t.default=t.layoutsSlice.reducer;PubSub.subscribe(o.TOPIC_SAVE,(function(e,t){var n=t.state,r=t.persist,a=n.layouts,o=n.cells,i=n.rows,l=n.pages;r.layout={pages:l,layouts:a,rows:i,cells:o}}))},68974:function(e,t,n){var r,a=this&&this.__assign||function(){return a=Object.assign||function(e){for(var t,n=1,r=arguments.length;n<r;n++)for(var a in t=arguments[n])Object.prototype.hasOwnProperty.call(t,a)&&(e[a]=t[a]);return e},a.apply(this,arguments)},o=this&&this.__spreadArray||function(e,t,n){if(n||2===arguments.length)for(var r,a=0,o=t.length;a<o;a++)!r&&a in t||(r||(r=Array.prototype.slice.call(t,0,a)),r[a]=t[a]);return e.concat(r||Array.prototype.slice.call(t))};Object.defineProperty(t,"__esModule",{value:!0}),t.selectPage=t.selectPages=t.remove=t.swap=t.add=t.set=t.pagesSlice=void 0;var i=n(20991);t.pagesSlice=(0,i.createSlice)({name:"pages",initialState:[],reducers:{set:function(e,t){e.splice.apply(e,o([0,e.length],t.payload,!1))},add:function(e,t){var n,r=null!==(n=Math.max.apply(Math,e.map((function(e){return e.order}))))&&void 0!==n?n:-1;e.push(a(a({},t.payload),{order:r+1}))},remove:function(e,t){var n=0;e=e.filter((function(e){return e.uid!==t.payload})).map((function(e){return a(a({},e),{order:n++})}))},swap:function(e,t){var n=e.find((function(e){return e.uid===t.payload.currentUid})),r=e.find((function(e){return e.uid===t.payload.targetUid})),a=n.order;n.order=r.order,r.order=a}}}),t.set=(r=t.pagesSlice.actions).set,t.add=r.add,t.swap=r.swap,t.remove=r.remove;t.selectPages=function(e){return e.pages};t.selectPage=function(e){return function(t){return t.pages.find((function(t){return t.uid===e}))}},t.default=t.pagesSlice.reducer},33314:function(e,t,n){var r,a=this&&this.__spreadArray||function(e,t,n){if(n||2===arguments.length)for(var r,a=0,o=t.length;a<o;a++)!r&&a in t||(r||(r=Array.prototype.slice.call(t,0,a)),r[a]=t[a]);return e.concat(r||Array.prototype.slice.call(t))};Object.defineProperty(t,"__esModule",{value:!0}),t.selectRowsInLayout=t.remove=t.swap=t.add=t.set=t.rowsSlice=void 0;var o=n(20991);t.rowsSlice=(0,o.createSlice)({name:"rows",initialState:[],reducers:{set:function(e,t){e.splice.apply(e,a([0,e.length],t.payload,!1))},add:function(e,t){var n=t.payload,r=n.layoutUid,o=n.uid,i=Math.max.apply(Math,a([-1],e.filter((function(e){return e.layoutUid===r})).map((function(e){return e.order})),!1));e.push({uid:o,order:i+1,layoutUid:r})},remove:function(e,t){e=e.filter((function(e){return e.uid!==t.payload}))},swap:function(e,t){var n=e.find((function(e){return e.uid===t.payload.currentUid})),r=e.find((function(e){return e.uid===t.payload.targetUid})),a=n.order;n.order=r.order,r.order=a}}}),t.set=(r=t.rowsSlice.actions).set,t.add=r.add,t.swap=r.swap,t.remove=r.remove;t.selectRowsInLayout=function(e){return function(t){return e?t.rows.filter((function(t){return t.layoutUid===e.uid})).sort((function(e,t){return e.order-t.order})):[]}},t.default=t.rowsSlice.reducer},65673:function(e,t,n){var r;Object.defineProperty(t,"__esModule",{value:!0}),t.selectQuery=t.clearQuery=t.updateQuery=t.searchSlice=t.Search=void 0;var a=n(20991);!function(e){e.Fields="fields"}(t.Search||(t.Search={}));t.searchSlice=(0,a.createSlice)({name:"search",initialState:{fields:""},reducers:{updateQuery:function(e,t){e[t.payload.type]=t.payload.query},clearQuery:function(e,t){e[t.payload]=""}}}),t.updateQuery=(r=t.searchSlice.actions).updateQuery,t.clearQuery=r.clearQuery;t.selectQuery=function(e){return function(t){return t.search[e]}},t.default=t.searchSlice.reducer},73396:function(e,t,n){Object.defineProperty(t,"__esModule",{value:!0}),t.addNewField=void 0;var r=n(11037),a=n(89372),o=n(45323),i=n(56859),l=n(33314),u=n(21614);t.addNewField=function(e){return function(t,n){var d=(0,u.v4)(),c=(0,u.v4)(),s=(0,u.v4)(),f=n(),p=(0,o.selectCurrentPage)(f);if(!p)throw new Error("No pages present");t((0,i.add)({fieldType:e,uid:d})),t((0,l.add)({layoutUid:p.layoutUid,uid:s})),t((0,a.add)({type:r.CellType.Field,rowUid:s,targetUid:d,uid:c}))}}},6175:function(e,t,n){Object.defineProperty(t,"__esModule",{value:!0}),t.addNewPage=void 0;var r=n(1965),a=n(68974),o=n(21614);t.addNewPage=function(){return function(e,t){var n=(0,o.v4)(),i=(0,o.v4)(),l=t().pages.length+1;e((0,r.add)({uid:i})),e((0,a.add)({uid:n,label:"Page ".concat(l),layoutUid:i,order:l}))}}},89141:function(e,t,n){Object.defineProperty(t,"__esModule",{value:!0}),t.Forms=t.Form=void 0;var r=n(33922);Object.defineProperty(t,"Form",{enumerable:!0,get:function(){return r.EditProvider}});var a=n(85581);Object.defineProperty(t,"Forms",{enumerable:!0,get:function(){return a.List}})},62819:function(e,t,n){var r=this&&this.__makeTemplateObject||function(e,t){return Object.defineProperty?Object.defineProperty(e,"raw",{value:t}):e.raw=t,e},a=this&&this.__importDefault||function(e){return e&&e.__esModule?e:{default:e}};Object.defineProperty(t,"__esModule",{value:!0}),t.Subtitle=t.Title=t.Card=t.Wrapper=void 0;var o,i,l,u,d=a(n(12788));t.Wrapper=d.default.ul(o||(o=r(["\n  display: grid;\n  grid-template-columns: repeat(4, 1fr);\n  gap: var(--l);\n\n  margin-top: var(--l);\n"],["\n  display: grid;\n  grid-template-columns: repeat(4, 1fr);\n  gap: var(--l);\n\n  margin-top: var(--l);\n"]))),t.Card=d.default.li(i||(i=r(["\n  padding: var(--xl);\n\n  background: #fff;\n  box-shadow: 0 0 0 1px #cdd8e4, 0 2px 12px rgb(205 216 228 / 50%);\n  border-radius: var(--large-border-radius);\n"],["\n  padding: var(--xl);\n\n  background: #fff;\n  box-shadow: 0 0 0 1px #cdd8e4, 0 2px 12px rgb(205 216 228 / 50%);\n  border-radius: var(--large-border-radius);\n"]))),t.Title=d.default.h3(l||(l=r(["\n  margin: 0;\n\n  font-size: 22px;\n  font-weight: normal;\n  text-align: center;\n"],["\n  margin: 0;\n\n  font-size: 22px;\n  font-weight: normal;\n  text-align: center;\n"]))),t.Subtitle=d.default.small(u||(u=r(["\n  display: block;\n\n  color: grey;\n  font-size: 12px;\n  font-family: monospace;\n  text-align: center;\n"],["\n  display: block;\n\n  color: grey;\n  font-size: 12px;\n  font-family: monospace;\n  text-align: center;\n"])))},85581:function(e,t,n){var r=this&&this.__importDefault||function(e){return e&&e.__esModule?e:{default:e}};Object.defineProperty(t,"__esModule",{value:!0}),t.List=void 0;var a=r(n(67294)),o=n(49818),i=n(23493),l=n(62819);t.List=function(){var e=(0,i.useQueryForms)(),t=e.data,n=e.isFetching,r=e.isError,u=e.error;return!t&&n?a.default.createElement("div",null,"fetching forms..."):r?a.default.createElement("div",null,"ERROR ",u.message):a.default.createElement("div",null,a.default.createElement("h1",null,"Forms",n&&a.default.createElement("span",null,"is fetching")),a.default.createElement(l.Wrapper,null,a.default.createElement(l.Card,null,a.default.createElement(l.Title,null,a.default.createElement(o.Link,{to:"new"},"Create new Form")),a.default.createElement(l.Subtitle,null,"click me")),t.map((function(e){return a.default.createElement(l.Card,{key:e.id},a.default.createElement(l.Title,null,a.default.createElement(o.Link,{to:"".concat(e.id)},e.name)),a.default.createElement(l.Subtitle,null,e.handle))}))))}},98320:function(e,t,n){var r=this&&this.__createBinding||(Object.create?function(e,t,n,r){void 0===r&&(r=n);var a=Object.getOwnPropertyDescriptor(t,n);a&&!("get"in a?!t.__esModule:a.writable||a.configurable)||(a={enumerable:!0,get:function(){return t[n]}}),Object.defineProperty(e,r,a)}:function(e,t,n,r){void 0===r&&(r=n),e[r]=t[n]}),a=this&&this.__setModuleDefault||(Object.create?function(e,t){Object.defineProperty(e,"default",{enumerable:!0,value:t})}:function(e,t){e.default=t}),o=this&&this.__importStar||function(e){if(e&&e.__esModule)return e;var t={};if(null!=e)for(var n in e)"default"!==n&&Object.prototype.hasOwnProperty.call(e,n)&&r(t,e,n);return a(t,e),t};Object.defineProperty(t,"__esModule",{value:!0}),t.Settings=void 0;var i=o(n(67294));t.Settings=function(){return(0,i.useEffect)((function(){return console.log("+++ settings"),function(){console.log("--- settings")}}),[]),i.default.createElement("div",null,"Settings")}},32523:function(e,t,n){Object.defineProperty(t,"__esModule",{value:!0}),t.useClickOutside=void 0;var r=n(67294);t.useClickOutside=function(e,t){var n=(0,r.useRef)();return(0,r.useEffect)((function(){var r=function(r){t&&n.current&&!n.current.contains(r.target)&&"function"==typeof e&&e()};return document.addEventListener("click",r,!0),function(){document.removeEventListener("click",r,!0)}}),[n,t]),n}},25818:function(e,t,n){Object.defineProperty(t,"__esModule",{value:!0}),t.useDebounce=void 0;var r=n(67294);t.useDebounce=function(e,t){void 0===t&&(t=500);var n=(0,r.useState)(e),a=n[0],o=n[1];return(0,r.useEffect)((function(){var n=setTimeout((function(){return o(e)}),t);return function(){return clearTimeout(n)}}),[e,t]),a}},12020:function(e,t,n){Object.defineProperty(t,"__esModule",{value:!0}),t.useOnKeypress=void 0;var r=n(67294);t.useOnKeypress=function(e){var t=e.meetsCondition,n=e.callback,a=e.type,o=void 0===a?"keyup":a;(0,r.useEffect)((function(){return(void 0===t||t)&&document.addEventListener(o,n),!1===t&&document.removeEventListener(o,n),function(){document.removeEventListener(o,n)}}),[t])}},92629:function(e,t,n){var r=this&&this.__importDefault||function(e){return e&&e.__esModule?e:{default:e}};Object.defineProperty(t,"__esModule",{value:!0});var a=r(n(67294)),o=r(n(20745)),i=n(88767),l=n(20938),u=n(49818);n(87662);var d=n(63438),c=n(16583),s=n(33658),f=n(89141),p=n(98320),v=n(48101),h=r(n(38957)),m=document.getElementById("freeform-client");o.default.createRoot(m).render(a.default.createElement(u.BrowserRouter,{basename:(0,v.generateUrl)("/client",!1)},a.default.createElement(i.QueryClientProvider,{client:d.queryClient},a.default.createElement(l.ReactQueryDevtools,null),a.default.createElement(c.CpNavigation,null),a.default.createElement(u.Routes,null,a.default.createElement(u.Route,{path:"/",element:a.default.createElement(h.default,null)},a.default.createElement(u.Route,{index:!0,element:a.default.createElement(s.Dashboard,null)}),a.default.createElement(u.Route,{path:"forms"},a.default.createElement(u.Route,{path:"new/*",element:a.default.createElement(f.Form,null)}),a.default.createElement(u.Route,{path:":formId/*",element:a.default.createElement(f.Form,null)}),a.default.createElement(u.Route,{index:!0,element:a.default.createElement(f.Forms,null)})),a.default.createElement(u.Route,{path:"settings",element:a.default.createElement(p.Settings,null)}))))))},42367:function(e,t,n){var r=this&&this.__importDefault||function(e){return e&&e.__esModule?e:{default:e}};Object.defineProperty(t,"__esModule",{value:!0}),t.useFieldType=t.useFetchFieldPropertySections=t.useFetchFieldTypes=void 0;var a=n(88767),o=r(n(9669));t.useFetchFieldTypes=function(e){var t=(void 0===e?{}:e).select;return(0,a.useQuery)("field-types",(function(){return o.default.get("/client/api/fields/types").then((function(e){return e.data}))}),{staleTime:1/0,select:t})};t.useFetchFieldPropertySections=function(){return(0,a.useQuery)(["field-types","property-sections"],(function(){return o.default.get("/client/api/fields/types/sections").then((function(e){return e.data})).then((function(e){return e.sort((function(e,t){return e.order-t.order}))}))}),{staleTime:1/0})};t.useFieldType=function(e){var n=(0,t.useFetchFieldTypes)().data;if(n)return n.find((function(t){return t.typeClass===e}))}},23493:function(e,t,n){var r=this&&this.__importDefault||function(e){return e&&e.__esModule?e:{default:e}};Object.defineProperty(t,"__esModule",{value:!0}),t.useQueryFormSettings=t.useQuerySingleForm=t.useQueryForms=void 0;var a=n(88767),o=n(85767),i=n(89372),l=n(45323),u=n(56859),d=n(61552),c=n(1965),s=n(68974),f=n(33314),p=r(n(9669));t.useQueryForms=function(){return(0,a.useQuery)("forms",(function(){return p.default.get("/client/api/forms").then((function(e){return e.data}))}))};t.useQuerySingleForm=function(e){var t=(0,o.useAppDispatch)();return(0,a.useQuery)(["forms",e],(function(){return p.default.get("/client/api/forms/".concat(e)).then((function(e){return e.data}))}),{staleTime:1/0,enabled:!!e,onSuccess:function(e){var n,r=e.layout,a=r.fields,o=r.pages,p=r.layouts,v=r.rows,h=r.cells;t((0,d.update)(e)),t((0,u.set)(a)),t((0,s.set)(o)),t((0,c.set)(p)),t((0,f.set)(v)),t((0,i.set)(h)),t((0,l.setPage)(null===(n=o.find(Boolean))||void 0===n?void 0:n.uid))}})};t.useQueryFormSettings=function(){return(0,a.useQuery)(["forms","settings"],(function(){return p.default.get("/client/api/forms/settings").then((function(e){return e.data}))}))}},41648:function(e,t,n){var r=this&&this.__importDefault||function(e){return e&&e.__esModule?e:{default:e}};Object.defineProperty(t,"__esModule",{value:!0}),t.useQuerySingleFormIntegration=t.useQueryFormIntegrations=t.useQueryIntegrations=void 0;var a=n(88767),o=n(47785),i=n(68887),l=r(n(9669));t.useQueryIntegrations=function(){return(0,a.useQuery)(["integrations"],(function(){return l.default.get("/client/api/integrations").then((function(e){return e.data}))}),{staleTime:1/0})};t.useQueryFormIntegrations=function(e){var t=(0,o.useDispatch)();return(0,a.useQuery)(["form-integrations"],(function(){return l.default.get("/client/api/forms/".concat(e,"/integrations")).then((function(e){return e.data}))}),{staleTime:1/0,cacheTime:1/0,onSuccess:function(e){t((0,i.addIntegrations)(e))}})};t.useQuerySingleFormIntegration=function(e,t){return(0,a.useQuery)(["form-integrations",t],(function(){return l.default.get("/client/api/forms/".concat(e,"/integrations/").concat(t)).then((function(e){return e.data}))}),{staleTime:1/0})}},240:function(e,t,n){var r=this&&this.__makeTemplateObject||function(e,t){return Object.defineProperty?Object.defineProperty(e,"raw",{value:t}):e.raw=t,e},a=this&&this.__importDefault||function(e){return e&&e.__esModule?e:{default:e}};Object.defineProperty(t,"__esModule",{value:!0}),t.colors=t.shadows=t.borderRadius=t.spacings=t.BoxShadow=void 0;var o,i=a(n(12788));t.BoxShadow=i.default.div(o||(o=r(["\n  box-shadow: 0 0 0 1px #cdd8e4, 0 2px 12px rgb(205 216 228 / 50%); ;\n"],["\n  box-shadow: 0 0 0 1px #cdd8e4, 0 2px 12px rgb(205 216 228 / 50%); ;\n"]))),t.spacings={xs:"var(--xs)",sm:"var(--s)",md:"var(--m)",lg:"var(--l)",xl:"var(--xl)"},t.borderRadius={sm:"var(--small-border-radius)",md:"var(--medium-border-radius)",lg:"var(--large-border-radius)"},t.shadows={box:"0 0 0 1px #cdd8e4, 0 2px 12px rgb(205 216 228 / 50%)",bottom:"inset 0 -1px 0 0 rgb(154 165 177 / 25%)",top:"inset 0 1px 0 0 rgb(154 165 177 / 25%)",left:"inset 1px 0 0 0 rgb(154 165 177 / 25%)",right:"inset -1px 0 0 0 rgb(154 165 177 / 25%)"},t.colors={error:"var(--error-color)",warning:"var(--warning-color)",success:"var(--success-color)",notice:"var(--notice-color)",enabled:"var(--enabled-color)",pending:"var(--pending-color)",disabled:"var(--disabled-color)",white:"var(--white)",black:"var(--black)",gray050:"var(--gray-050)",gray100:"var(--gray-100)",gray200:"var(--gray-200)",gray300:"var(--gray-300)",gray350:"var(--gray-350)",gray400:"var(--gray-400)",gray500:"var(--gray-500)",gray550:"var(--gray-550)",gray600:"var(--gray-600)",gray700:"var(--gray-700)",gray800:"var(--gray-800)",gray900:"var(--gray-900)",gray1000:"var(--gray-1000)",blue050:"var(--blue-050)",blue100:"var(--blue-100)",blue200:"var(--blue-200)",blue300:"var(--blue-300)",blue400:"var(--blue-400)",blue500:"var(--blue-500)",blue600:"var(--blue-600)",blue700:"var(--blue-700)",blue800:"var(--blue-800)",blue900:"var(--blue-900)",cyan050:"var(--cyan-050)",cyan100:"var(--cyan-100)",cyan200:"var(--cyan-200)",cyan300:"var(--cyan-300)",cyan400:"var(--cyan-400)",cyan500:"var(--cyan-500)",cyan600:"var(--cyan-600)",cyan700:"var(--cyan-700)",cyan800:"var(--cyan-800)",cyan900:"var(--cyan-900)",pink050:"var(--pink-050)",pink100:"var(--pink-100)",pink200:"var(--pink-200)",pink300:"var(--pink-300)",pink400:"var(--pink-400)",pink500:"var(--pink-500)",pink600:"var(--pink-600)",pink700:"var(--pink-700)",pink800:"var(--pink-800)",pink900:"var(--pink-900)",red050:"var(--red-050)",red100:"var(--red-100)",red200:"var(--red-200)",red300:"var(--red-300)",red400:"var(--red-400)",red500:"var(--red-500)",red600:"var(--red-600)",red700:"var(--red-700)",red800:"var(--red-800)",red900:"var(--red-900)",yellow050:"var(--yellow-050)",yellow100:"var(--yellow-100)",yellow200:"var(--yellow-200)",yellow300:"var(--yellow-300)",yellow400:"var(--yellow-400)",yellow500:"var(--yellow-500)",yellow600:"var(--yellow-600)",yellow700:"var(--yellow-700)",yellow800:"var(--yellow-800)",yellow900:"var(--yellow-900)",teal050:"var(--teal-050)",teal100:"var(--teal-100)",teal200:"var(--teal-200)",teal300:"var(--teal-300)",teal400:"var(--teal-400)",teal500:"var(--teal-500)",teal550:"var(--teal-550)",teal600:"var(--teal-600)",teal700:"var(--teal-700)",teal800:"var(--teal-800)",teal900:"var(--teal-900)"}},88454:function(e,t){var n,r=this&&this.__extends||(n=function(e,t){return n=Object.setPrototypeOf||{__proto__:[]}instanceof Array&&function(e,t){e.__proto__=t}||function(e,t){for(var n in t)Object.prototype.hasOwnProperty.call(t,n)&&(e[n]=t[n])},n(e,t)},function(e,t){if("function"!=typeof t&&null!==t)throw new TypeError("Class extends value "+String(t)+" is not a constructor or null");function r(){this.constructor=e}n(e,t),e.prototype=null===t?Object.create(t):(r.prototype=t.prototype,new r)});Object.defineProperty(t,"__esModule",{value:!0}),t.APIError=t.API_ERROR=void 0,t.API_ERROR="api_error";var a=function(e){function n(n,r){var a=e.call(this,n)||this;return a.errors={},a.name=t.API_ERROR,a.errors=r,a}return r(n,e),n}(Error);t.APIError=a},82344:function(e,t){Object.defineProperty(t,"__esModule",{value:!0}),t.range=void 0;t.range=function(e,t){void 0===t&&(e>1?(t=e,e=1):(t=e,e=0));for(var n=[],r=e;r<=t;r++)n.push(r);return n}},89361:function(e,t){Object.defineProperty(t,"__esModule",{value:!0}),t.classes=void 0;t.classes=function(){for(var e=[],t=0;t<arguments.length;t++)e[t]=arguments[t];return e.map((function(e){return"string"==typeof e&&(e=e.trim()),e})).filter((function(e){return!!e})).join(" ")},t.default=t.classes},87351:function(e,t){Object.defineProperty(t,"__esModule",{value:!0}),t.translate=t.replace=void 0;t.replace=function(e,t){void 0===t&&(t={});for(var n=0,r=Object.entries(t);n<r.length;n++){var a=r[n],o=a[0],i=a[1],l=new RegExp("\\{"+o+"\\}","g");e=e.replace(l,i.toString())}return e};t.translate=function(e,n){return void 0===n&&(n={}),"undefined"!=typeof Craft?Craft.t("freeform",e,n):(0,t.replace)(e,n)},t.default=t.translate},48101:function(e,t,n){var r=this&&this.__importDefault||function(e){return e&&e.__esModule?e:{default:e}};Object.defineProperty(t,"__esModule",{value:!0}),t.generateUrl=t.baseUrl=void 0;var a=r(n(84564));t.baseUrl=window.location.href.replace(/(.*\/freeform).*/i,"$1");t.generateUrl=function(e,n){void 0===n&&(n=!0),e=(e=(null!=e?e:"").replace(/\/+/g,"/").replace(/^\/(.*)/,"$1").replace(/\/$/,"")).length?"/".concat(e):"";var r=(0,a.default)("".concat(t.baseUrl).concat(e));return n?r.href:r.pathname}}},a={};function o(e){var t=a[e];if(void 0!==t)return t.exports;var n=a[e]={id:e,loaded:!1,exports:{}};return r[e].call(n.exports,n,n.exports,o),n.loaded=!0,n.exports}o.m=r,e=[],o.O=function(t,n,r,a){if(!n){var i=1/0;for(c=0;c<e.length;c++){n=e[c][0],r=e[c][1],a=e[c][2];for(var l=!0,u=0;u<n.length;u++)(!1&a||i>=a)&&Object.keys(o.O).every((function(e){return o.O[e](n[u])}))?n.splice(u--,1):(l=!1,a<i&&(i=a));if(l){e.splice(c--,1);var d=r();void 0!==d&&(t=d)}}return t}a=a||0;for(var c=e.length;c>0&&e[c-1][2]>a;c--)e[c]=e[c-1];e[c]=[n,r,a]},o.n=function(e){var t=e&&e.__esModule?function(){return e.default}:function(){return e};return o.d(t,{a:t}),t},n=Object.getPrototypeOf?function(e){return Object.getPrototypeOf(e)}:function(e){return e.__proto__},o.t=function(e,r){if(1&r&&(e=this(e)),8&r)return e;if("object"==typeof e&&e){if(4&r&&e.__esModule)return e;if(16&r&&"function"==typeof e.then)return e}var a=Object.create(null);o.r(a);var i={};t=t||[null,n({}),n([]),n(n)];for(var l=2&r&&e;"object"==typeof l&&!~t.indexOf(l);l=n(l))Object.getOwnPropertyNames(l).forEach((function(t){i[t]=function(){return e[t]}}));return i.default=function(){return e},o.d(a,i),a},o.d=function(e,t){for(var n in t)o.o(t,n)&&!o.o(e,n)&&Object.defineProperty(e,n,{enumerable:!0,get:t[n]})},o.g=function(){if("object"==typeof globalThis)return globalThis;try{return this||new Function("return this")()}catch(e){if("object"==typeof window)return window}}(),o.o=function(e,t){return Object.prototype.hasOwnProperty.call(e,t)},o.r=function(e){"undefined"!=typeof Symbol&&Symbol.toStringTag&&Object.defineProperty(e,Symbol.toStringTag,{value:"Module"}),Object.defineProperty(e,"__esModule",{value:!0})},o.nmd=function(e){return e.paths=[],e.children||(e.children=[]),e},function(){var e={47:0};o.O.j=function(t){return 0===e[t]};var t=function(t,n){var r,a,i=n[0],l=n[1],u=n[2],d=0;if(i.some((function(t){return 0!==e[t]}))){for(r in l)o.o(l,r)&&(o.m[r]=l[r]);if(u)var c=u(o)}for(t&&t(n);d<i.length;d++)a=i[d],o.o(e,a)&&e[a]&&e[a][0](),e[a]=0;return o.O(c)},n=self.webpackChunk_ff_client=self.webpackChunk_ff_client||[];n.forEach(t.bind(null,0)),n.push=t.bind(null,n.push.bind(n))}(),o.nc=void 0;var i=o.O(void 0,[736],(function(){return o(92629)}));i=o.O(i)}();
